+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
 affiliations:
     - name: Department of Mathematics, Imperial College London, UK
       index: 1
-date: 26 June 2024
+date: 27 June 2024
 bibliography: paper.bib
 ---
 
