+++ paper/paper.bib
@@ -7,13 +7,13 @@
     year = {2013},
 }
 
-
 @book{aurenhammer2013voronoi,
     author = {Franz Aurenhammer and Rolf Klein and Der-Tsai Lee},
     title = {Voronoi Diagrams and Delaunay Triangulations},
     publisher = {World Scientific},
     year = {2013},
-    url = {https://www.worldscientific.com/doi/abs/10.1142/8685},
+    url = {https://doi.org/10.1142/8685},
+    doi = {10.1142/8685},
 }
 
 @article{mucke1999fast,
@@ -40,6 +40,7 @@ title = {Delaunay refinement mesh generation of curve-bounded domains},
 author = {Serge Gosselin},
 year = {2009},
 school = {The University of British Columbia},
+doi={10.14288/1.0067778},
 }
 
 @article{bezanson2017julia,
@@ -50,6 +51,7 @@ school = {The University of British Columbia},
   number={1},
   pages={65--98},
   year={2017},
+  doi = {10.1137/141000671},
 }
 
 @article{golias1997delaunay,
@@ -59,6 +61,7 @@ journal = {Finite Elements in Analysis and Design},
 volume = {25},
 year = {1997},
 pages = {331--341},
+doi = {10.1016/S0168-874X(96)00054-6},
 }
 
 @article{ju2006adaptive,
@@ -68,6 +71,7 @@ journal = {SIAM Journal on Scientific Computing},
 volume = {28},
 pages = {2023-2053},
 year = {2006},
+doi = {10.1137/050643568},
 }
 
 @misc{lairez2024exact,
@@ -87,6 +91,7 @@ volume = {81},
 pages = {373--399},
 journal = {Current Topics in Developmental Biology},
 author = {Michael {Meyer-Hermann}},
+doi = {10.1016/S0070-2153(07)81013-1},
 }
 
 @article{du1999centroidal,
@@ -96,6 +101,7 @@ journal = {SIAM Review},
 volume = {41},
 pages = {637--676},
 year = {1999},
+doi = {10.1137/S0036144599352836},
 }
 
 @article{wang2024calibration,
@@ -104,6 +110,7 @@ year = {1999},
     year = {2024},
     journal = {Journal of Mathematical Biology},
     volume = {88},
+    doi = {10.1007/s00285-024-02045-4},
 }
 
 @incollection{shewchuk1996triangle,
@@ -118,7 +125,8 @@ publisher = {Springer-Verlag},
 pages = {203--222},
 month = may,
 year = 1996,
-note = {From the First ACM Workshop on Applied Computational Geometry}
+note = {From the First ACM Workshop on Applied Computational Geometry},
+doi = {10.1007/BFb0014497},
 }
 
 @software{MATLAB,
@@ -136,6 +144,7 @@ url = {https://www.mathworks.com}
   title={Path planning based on constrained {D}elaunay triangulation}, 
   year={2008},
   pages={5168--5173},
+  doi = {10.1109/WCICA.2008.4593771},
 }
 
 @article{SciPy,
@@ -158,6 +167,7 @@ url = {https://www.mathworks.com}
   year    = {2020},
   volume  = {17},
   pages   = {261--272},
+  doi = {10.1038/s41592-019-0686-2},
 }
 
 @book{CGAL, 
@@ -176,6 +186,7 @@ journal = {International Journal for Numerical Methods in Engineering},
 volume = {79},
 pages = {1309--1331},
 year = {2009},
+doi = {10.1002/nme.2579},
 }
 
 @article{vandenheuvel2023computational,
@@ -185,6 +196,7 @@ year = {2023},
 journal = {Chemical Engineering Journal},
 volume = {475},
[…]
