+++ paper/paper.bib
@@ -196,7 +196,7 @@ year = {2023},
 journal = {Chemical Engineering Journal},
 volume = {475},
 pages = {145776},
-doi = {https://doi.org/10.1016/j.cej.2023.145776},
+doi = {10.1016/j.cej.2023.145776},
 }
 
 @article{danisch2021makie, 
