### 2024-06-27 #6937 @DanielVandH
```
- No DOI given, and none found for title: Delaunay Mesh Generation
- No DOI given, and none found for title: Natural neighbor interpolation - critical assessme...
- No DOI given, and none found for title: ExactPredicates.jl: Fast and exact geometrical pre...
- No DOI given, and none found for title: MATLAB version: R2024a
- No DOI given, and none found for title: CGAL User and Reference Manual
```

As far as I can tell, none of these titles have a DOI assigned to them, and https://www.crossref.org/guestquery also seems to show no DOIs for them.
