### 2024-06-27 #6937 @danielskatz
> Thanks @danielskatz. I'll check out the compiling issues - it worked when I did it in https://github.com/JuliaGeometry/DelaunayTriangulation.jl/actions/runs/9697066359, did I configure the workflow file incorrectly?
>
> I'll get all the DOIs fixed up. I'll get them manually incase some of them are incorrect as you suggest, thanks.

Sometimes there can be difference between the action and what we do. I'm unsure why, but I'm pinging @openjournals/dev in case this is something they can fix.
