### 2024-06-28 cb59357
Fix various type stability issues; refactor validate_triangulation
