### 2024-06-30 43d38a1
some more stability fixes; redesign Polygon to make is_circular a field rather than using views
