+++ test/voronoi/voronoi.jl
@@ -95,6 +95,11 @@ include("../helper_functions.jl")
     end
 end
 
+#=
+DelaunayTriangulation.Polygon{Tuple{Float64, Float64}, SubArray{Int64, 1, Vector{Int64}, Tuple{UnitRange{Int64}}, true}, Vector{Tuple{Float64, Float64}}}, 
+DelaunayTriangulation.Polygon{Tuple{Float64, Float64}, Vector{Int64}, Vector{Tuple{Float64, Float64}}}}, DelaunayTriangulation.Polygon{Tuple{Float64, Float64}, NTuple{4, Int64}, NTuple{4, Tuple{Float64, Float64}}}}
+=#
+
 @testset "Smaller example, checking ray coordinates" begin
     for _ in 1:10
         tri = example_triangulation()
