### 2024-06-30 7e9adc0
Use Float64 for area and circumcenter
