+++ src/data_structures/triangulation/methods/boundary_curves.jl
@@ -11,7 +11,7 @@ is_curve_bounded
 @inline is_curve_bounded(::Number) = false
 @inline is_curve_bounded(boundary_curves::Tuple) = any(is_curve_bounded, boundary_curves)
 @inline is_curve_bounded(tri::Triangulation) = is_curve_bounded(get_boundary_curves(tri))
-@stable @inline function is_curve_bounded(boundary_nodes::AbstractVector)::Bool
+@inline function is_curve_bounded(boundary_nodes::AbstractVector)::Bool
     if has_multiple_curves(boundary_nodes)
         nc = num_curves(boundary_nodes)
         for curve_index in 1:nc
@@ -34,7 +34,7 @@ end
 
 Returns the set of boundary curves associated with `boundary_nodes` and `points`. 
 """
-@unstable @inline function to_boundary_curves(points, boundary_nodes)
+@inline function to_boundary_curves(points, boundary_nodes)
     if has_multiple_curves(boundary_nodes)
         return _to_boundary_curves_multiple_curves(points, boundary_nodes)
     elseif has_multiple_sections(boundary_nodes)
@@ -51,24 +51,20 @@ end
         return (PiecewiseLinear(points, boundary_nodes),)
     end
 end
-@unstable @inline function _to_boundary_curves_multiple_sections(points, boundary_nodes, section=1, boundary_curves=())
+@inline function _to_boundary_curves_multiple_sections(points, boundary_nodes, section=1, boundary_curves=())
     if section > num_sections(boundary_nodes)
         return boundary_curves
     else
-        section_nodes = allow_unstable() do
-            get_boundary_nodes(boundary_nodes, section)
-        end
+        section_nodes = get_boundary_nodes(boundary_nodes, section)
         new_boundary_curves = _to_boundary_curves_single_curve(points, section_nodes)
         return _to_boundary_curves_multiple_sections(points, boundary_nodes, section + 1, (boundary_curves..., new_boundary_curves...))
     end
 end
-@unstable @inline function _to_boundary_curves_multiple_curves(points, boundary_nodes, curve=1, boundary_curves=())
+@inline function _to_boundary_curves_multiple_curves(points, boundary_nodes, curve=1, boundary_curves=())
     if curve > num_curves(boundary_nodes)
         return boundary_curves
     else
-        curve_nodes = allow_unstable() do
-            get_boundary_nodes(boundary_nodes, curve)
-        end
+        curve_nodes = get_boundary_nodes(boundary_nodes, curve)
         new_boundary_curves = _to_boundary_curves_multiple_sections(points, curve_nodes)
         return _to_boundary_curves_multiple_curves(points, boundary_nodes, curve + 1, (boundary_curves..., new_boundary_curves...))
     end
@@ -94,7 +90,7 @@ triangulation. In particular:
 - `boundary_curves`: The boundary curves associated with `boundary_nodes`.
 - `boundary_nodes`: The modified boundary nodes.
 """
-@unstable @inline function convert_boundary_curves!(points, boundary_nodes, ::Type{I}) where {I<:Integer}
+@inline function convert_boundary_curves!(points, boundary_nodes, ::Type{I}) where {I<:Integer}
     boundary_curves = to_boundary_curves(points, boundary_nodes)
     !is_curve_bounded(boundary_curves) && return boundary_curves, boundary_nodes
     new_boundary_nodes = get_skeleton(boundary_nodes, I)
@@ -107,22 +103,14 @@ triangulation. In particular:
     end
     return boundary_curves, new_boundary_nodes
 end
-@stable @inline function _convert_boundary_curves_multiple_curves!(points, boundary_nodes, boundary_curves, new_boundary_nodes)
[…]
