+++ test/Untitled-2.jl
@@ -1,879 +1,0 @@
-"""
-    abstract type AbstractTriangulationState 
-
-`AbstractTriangulationState`s are used for representing the 
-state of a certain test of a triangulation. 
-
-# Interface 
-- Subtypes should have a `flag` field that is `true` if the state is positive, and `false` otherwise. 
-    This allows `test_state` to return `true` and `false` according to the flag.
-- Subtypes should implement `Base.summary(state)` which creates a string summarising the state. This is 
-    used in `show` for printing the `state`.
-"""
-abstract type AbstractTriangulationState end
-test_state(state::AbstractTriangulationState) = state.flag
-Base.show(io::IO, state::AbstractTriangulationState) = print(io, summary(state))
-
-struct TriangleOrientationState <: AbstractTriangulationState
-    flag::Bool
-    bad_triangle::NTuple{3,Int}
-    triangle_orientation::Symbol
-end
-TriangleOrientationState(tri) = test_triangle_orientation(tri)
-function Base.summary(state::TriangleOrientationState)
-    if test_state(state)
-        return "All the triangles have positive orientation."
-    else
-        return "The triangle $(state.bad_triangle) is $(state.triangle_orientation) oriented."
-    end
-end
-function test_triangle_orientation(tri)
-    for T in each_solid_triangle(tri)
-        cert = DT.triangle_orientation(tri, T)
-        flag = DT.is_positively_oriented(cert)
-        orientation = DT.is_positively_oriented(cert) ? :positively : DT.is_negatively_oriented(cert) ? :negatively : :degenerately
-        !flag && return TriangleOrientationState(flag, Int.(triangle_vertices(T)), orientation)
-    end
-    return TriangleOrientationState(true, (DT.∅, DT.∅, DT.∅), :positive)
-end
-
-struct DelaunayCriterionState <: AbstractTriangulationState
-    flag::Bool
-    bad_triangle::NTuple{3,Int}
-    bad_vertex::Int
-end
-DelaunayCriterionState(tri) = test_delaunay_criterion(tri)
-function Base.summary(state::DelaunayCriterionState)
-    if test_state(state)
-        return "All the triangles are Delaunay."
-    else
-        return "The Delaunay criterion does not hold for the triangle-vertex pair ($(state.bad_triangle), $(state.bad_vertex))."
-    end
-end
-function test_delaunay_criterion(tri)
-    points = get_points(tri)
-    triangle_tree = DT.BoundaryRTree(points)
-    segment_tree = DT.BoundaryRTree(points)
-    failures = Tuple{DT.triangle_type(tri),DT.integer_type(tri)}[]
-    for T in each_solid_triangle(tri)
-        i, j, k = DT.triangle_vertices(T)
-        p, q, r = get_point(tri, i, j, k)
-        c = DT.triangle_circumcenter(p, q, r)
-        cr = DT.triangle_circumradius(p, q, r)
-        xmin, xmax = getx(c) - cr, getx(c) + cr
-        ymin, ymax = gety(c) - cr, gety(c) + cr
-        any(!isfinite, (xmin, xmax, ymin, ymax)) && continue
-        bbox = DT.BoundingBox(xmin, xmax, ymin, ymax)
-        dbx = DT.DiametralBoundingBox(bbox, (i, j)) # just store (i, j) and get k via get_adjacent 
-        insert!(triangle_tree.tree, dbx)
-    end
-    for e in each_segment(tri)
-        i, j = DT.edge_vertices(e)
-        p, q = get_point(tri, i, j)
[…]
