+++ test/operations/delete_ghost_triangles.jl
@@ -3,7 +3,7 @@ const DT = DelaunayTriangulation
 using StructEquality
 using ..DelaunayTriangulation: Triangulation
 
-include("../helper_functions.jl")
+Base.include(@__MODULE__, "../helper_functions.jl")
 
 @testset "Deleting ghost triangles" begin
     tri, label_map, index_map = simple_geometry()
