+++ test/point_location/select_initial_point.jl
@@ -4,7 +4,7 @@ using LinearAlgebra
 using Random
 using StableRNGs
 
-include("../helper_functions.jl")
+Base.include(@__MODULE__, "../helper_functions.jl")
 
 global tri, label_map, index_map = simple_geometry()
 add_ghost_triangles!(tri)
