+++ test/predicates/boundaries_and_ghosts.jl
@@ -5,7 +5,7 @@ using StableRNGs
 const GV = DT.𝒢
 using ..DelaunayTriangulation: Certificate
 
-Base.include(@__MODULE__, "../helper_functions.jl")
+# Base.include(@__MODULE__, "../helper_functions.jl")
 
 x, y = complicated_geometry()
 rng = StableRNG(99988)
