+++ test/data_structures/bst.jl
@@ -2,7 +2,7 @@ using ..DelaunayTriangulation
 using Test 
 using DataStructures 
 const DT = DelaunayTriangulation
-Base.include(@__MODULE__, "../helper_functions.jl")
+# Base.include(@__MODULE__, "../helper_functions.jl")
 
 tree = DT.BalancedBST{Int}()
 tree_avl = AVLTree{Int}()
