### 2024-06-30 a62c10e
Add missing test images, test tutorials and applications in tests directly, execute literate scripts, refactor runtests.jl file
