+++ src/data_structures/trees/rtree.jl
@@ -496,6 +496,10 @@ mutable struct Leaf{Branch} <: AbstractNode
     parent::Union{Branch,Nothing}
     bounding_box::BoundingBox
     children::Vector{DiametralBoundingBox} # would do const, but for compat reasons I don't
+    Leaf(parent::Branch, bounding_box, children) where {Branch} = new{Branch}(parent, bounding_box, children)
+    Leaf{Branch}(parent::Branch, bounding_box, children) where {Branch} = new{Branch}(parent, bounding_box, children)
+    Leaf{Branch}(::Nothing, bounding_box, children) where {Branch} = new{Branch}(parent, bounding_box, children)
+    # need to separate out the constructors to avoid unbound type argument issues from Aqua
 end
 function Base.:(==)(leaf1::Leaf, leaf2::Leaf)
     xor(isnothing(leaf1), isnothing(leaf2)) && return false # if we test get_parent(leaf1) ≠ get_parent(leaf2), then we get a StackOverflowError
