+++ test/readme_example.jl
@@ -1,0 +1,66 @@
+using DelaunayTriangulation, CairoMakie, ReferenceTests, StableRNGs
+
+# Unconstrained 
+rng = StableRNG(123)
+points = rand(rng, 2, 50)
+tri1 = triangulate(points; rng)
+
+# Voronoi example 
+vorn2 = voronoi(tri1; rng)
+
+# Clipped Voronoi 
+vorn3 = voronoi(tri1, clip=true; rng)
+
+# Smoothed Voronoi 
+vorn4 = centroidal_smooth(vorn3; rng)
+
+# Constrained example with refinement 
+boundary_points = [(0.0, 0.0), (1.0, 0.0), (1.0, 0.3), (0.5, 0.3),
+    (0.3, 0.7), (0.1, 1.0), (0.0, 1.0), (0.0, 0.0)]
+boundary_nodes, points = convert_boundary_points_to_indices(boundary_points)
+tri5 = triangulate(points; boundary_nodes, rng)
+refine!(tri5; max_area=1e-2get_area(tri5), rng)
+
+# Disjoint constrained example with refinement 
+boundary_points = [
+    [[(0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0)]],
+    [[(0.3, 0.3), (0.3, 0.7), (0.7, 0.7), (0.7, 0.3), (0.3, 0.3)]],
+    [[(1.2, 0.0), (1.4, 0.0), (1.4, 1.2), (0.0, 1.2), (0.0, 1.1),
+        (1.1, 1.1), (1.1, 0.0), (1.2, 0.0)]]
+]
+boundary_nodes, points = convert_boundary_points_to_indices(boundary_points)
+tri6 = triangulate(points; boundary_nodes, rng)
+refine!(tri6; max_area=1e-2get_area(tri6), rng)
+
+# Curve-bounded example
+using DelaunayTriangulation: EllipticalArc
+ellipse = EllipticalArc((1.0, 0.0), (1.0, 0.0), (0.0, 0.0), 1.0, 2.0, 0.0)
+tri7 = triangulate(NTuple{2,Float64}[]; boundary_nodes=[ellipse], rng)
+refine!(tri7; max_area=1e-2get_area(tri7), rng)
+
+# Disjoint curve-bounded example 
+ellipse = EllipticalArc((7.0, 3.5), (7.0, 3.5), (0.0, 3.5), 7.0, 10.0, 0.0)
+catmull_cp = [(0.0, 0.0), (-2.0, -1.0), (-4.0, 0.0), (-5.0, 2.0), (-1.0, 4.0), (0.0, 3.0),
+    (1.0, 4.0), (5.0, 2.0), (4.0, 0.0), (2.0, -1.0), (0.0, 0.0)]
+catmull_spl = CatmullRomSpline(catmull_cp)
+circle = CircularArc((0.5, 1.5), (0.5, 1.5), (0.0, 1.0))
+circle2 = CircularArc((0.1, 1.5), (0.1, 1.5), (0.0, 1.0), positive=false)
+points = [(-4.0, -10.0), (4.0, -10.0), (4.0, -7.0), (-4.0, -7.0)]
+square = [1, 2, 3, 4, 1]
+boundary_nodes = [[square], [[ellipse]], [[catmull_spl]], [[circle]], [[circle2]]]
+tri8 = triangulate(points; boundary_nodes, rng)
+refine!(tri8; max_area=1e-2get_area(tri8), rng) # could also use find_polygon to help define a custom refinement function for each shape
+
+# Plotting 
+fig = Figure(fontsize = 42, size = (2800, 1480))
+ax = Axis(fig[1, 1], title="Unconstrained", width=600,height=600);            triplot!(ax, tri1)
[…]
