### 2024-06-30 5b89885
Put all the tests into a main testset
