+++ test/operations/delete_triangle.jl
@@ -4,7 +4,7 @@ using CairoMakie
 using DataStructures
 using ..DelaunayTriangulation: Triangulation
 
-include("../helper_functions.jl")
+_functions.jl")
 
 global tri, label_map, index_map = simple_geometry()
 add_ghost_triangles!(tri)
