+++ test/predicates/index_and_ghost_handling.jl
@@ -6,7 +6,7 @@ using StatsBase
 using ElasticArrays
 using ..DelaunayTriangulation: Certificate
 
-include("../helper_functions.jl")
+_functions.jl")
 
 global x, y = complicated_geometry()
 boundary_nodes, points = convert_boundary_points_to_indices(x, y; existing_points=ElasticMatrix{Float64}(undef, 2, 0))
