+++ test/triangulation/rectangle.jl
@@ -2,7 +2,7 @@ using ..DelaunayTriangulation
 const DT = DelaunayTriangulation
 using CairoMakie
 
-#include("../helper_functions.jl")
+#_functions.jl")
 
 @testset "Multiple boundaries" begin
     a, b, c, d = 2.0, 10.0, -5.0, 7.5
