+++ test/data_structures/max_priority_queue.jl
@@ -1,7 +1,7 @@
 using Test, DataStructures
 using ..DelaunayTriangulation
 const DT = DelaunayTriangulation
-include("../helper_functions.jl")
+_functions.jl")
 
 @test DT.hparent(2) == 1
 @test DT.hparent(3) == 1
