### 2024-06-30 b3902c1
Fix turning helperfunctions into a module
