+++ test/data_structures/polygon_hierarchy.jl
@@ -9,7 +9,7 @@ using Test
 using StructEquality
 using DelimitedFiles
 using StableRNGs
-_functions.jl")
+
 
 # Examples:
 #   I: A piecewise linear contiguous section.
