+++ test/operations/lock_convex_hull.jl
@@ -2,8 +2,8 @@ using ..DelaunayTriangulation
 const DT = DelaunayTriangulation
 using CairoMakie
 using DataStructures
+using StaticArrays
 
-_functions.jl")
 
 @testset "lock_convex_hull" begin
     for _ in 1:150
