+++ test/predicates/boundaries_and_ghosts.jl
@@ -2,10 +2,11 @@ using ..DelaunayTriangulation
 const DT = DelaunayTriangulation
 using DataStructures
 using StableRNGs
+using StaticArrays
 const GV = DT.𝒢
 using ..DelaunayTriangulation: Certificate
 
-_functions.jl")
+
 
 x, y = complicated_geometry()
 rng = StableRNG(99988)
