+++ test/data_structures/statistics.jl
@@ -3,7 +3,7 @@ const DT = DelaunayTriangulation
 using CairoMakie
 using ReferenceTests
 
-_functions.jl")
+
 
 @testset "Computing statistics" begin
     _x, _y = complicated_geometry()
