### 2024-06-30 f4390d3
Fix up some issues with docs
