+++ test/runtests.jl
@@ -32,7 +32,7 @@ using Aqua
 using Test
 
 const ALL_TEST_SCRIPTS = Set{String}()
-const NON_TEST_SCRIPTS = Set{String}(["helper_functions.jl", "triangulation_validation.jl", "runtests.jl", "triangulation\\weighted.jl"])
+const NON_TEST_SCRIPTS = Set{String}(["helper_functions.jl", "triangulation_validation.jl", "runtests.jl", "triangulation\\weighted.jl", "triangulation/weighted.jl"])
 include("helper_functions.jl")
 using .HelperFunctions
 
