### 2024-06-30 facc898
Fix tests on 1.6
