+++ docs/src/literate_applications/cell_simulations.jl
@@ -46,7 +46,7 @@ using StableRNGs
 using LinearAlgebra
 using StatsBase
 using CairoMakie
-@kwdef mutable struct CellModel{P}
+Base.@kwdef mutable struct CellModel{P}
     tri::Triangulation{P}
     new_r_cache::Vector{NTuple{2,Float64}} # for r(t + Δt)
     α::Float64
