### 2024-07-01 efa9d64
Remove some comments and test more of replace_triangle functions
