+++ test/data_structures/convex_hull.jl
@@ -72,3 +72,8 @@ end
     empty!(ch)
     @test isempty(DT.get_vertices(ch))
 end
+
+@testset "Issue #109" begin
+    points = rand(2, 50)
+    @test convex_hull(points).vertices == convex_hull(vcat(points, points)).vertices
+end
