+++ previews/PR86/literate_tutorials/constrained_multipolygon.jl
@@ -1,0 +1,192 @@
+# # Constrained Triangulation 
+# ## Disjoint Domains
+
+# Now we consider an example where we are triangulating disjoint domains. 
+# The way to setup these domains is similar to the previous tutorials, 
+# with the outer boundaries being counter-clockwise, and all the interior 
+# boundaries clockwise (and other interiors inside interiors counter-clockwise, 
+# etc., if you please). Here is a simple example; note the `check_arguments = false`
+# as in the previous tutorial. 
+using DelaunayTriangulation
+using CairoMakie
+using ReferenceTests #src
+fig_path = joinpath(@__DIR__, "../figures") #src
+
+θ = LinRange(0, 2π, 20) |> collect
+θ[end] = 0 # need to make sure that 2π gives the exact same coordinates as 0
+xy = Vector{Vector{Vector{NTuple{2,Float64}}}}()
+cx = 0.0
+for i in 1:2
+    global cx
+    ## Make the exterior circle
+    push!(xy, [[(cx + cos(θ), sin(θ)) for θ in θ]])
+    ## Now the interior circle - clockwise
+    push!(xy, [[(cx + 0.5cos(θ), 0.5sin(θ)) for θ in reverse(θ)]])
+    cx += 3.0
+end
+boundary_nodes, points = convert_boundary_points_to_indices(xy)
+tri = triangulate(points; boundary_nodes=boundary_nodes, check_arguments=false)
+fig, ax, sc = triplot(tri)
+fig
+@test_reference joinpath(fig_path, "constrained_ex_7.png") fig #src
+
+# Here is another example. 
+C = (15.7109521325776, 33.244486807457)
+D = (14.2705719699703, 32.8530791545746)
+E = (14.3, 27.2)
+F = (14.1, 27.0)
+G = (13.7, 27.2)
+H = (13.4, 27.5)
+I = (13.1, 27.6)
+J = (12.7, 27.4)
+K = (12.5, 27.1)
+L = (12.7, 26.7)
+M = (13.1, 26.5)
+N = (13.6, 26.4)
+O = (14.0, 26.4)
+P = (14.6, 26.5)
+Q = (15.1983491346581, 26.8128534095401)
+R = (15.6, 27.6)
+S = (15.6952958264624, 28.2344688505621)
+T = (17.8088971520274, 33.1192363585346)
+U = (16.3058917649589, 33.0722674401887)
+V = (16.3215480710742, 29.7374742376305)
+W = (16.3841732955354, 29.393035503094)
+Z = (16.6190178872649, 28.9233463196351)
+A1 = (17.0417381523779, 28.5319386667527)
+B1 = (17.5114273358368, 28.3753756055997)
+C1 = (18.1376795804487, 28.3597192994844)
+D1 = (18.7169629067146, 28.5632512789833)
+E1 = (19.2805899268653, 28.8920337074045)
+F1 = (19.26493362075, 28.4536571361762)
+G1 = (20.6426885588962, 28.4223445239456)
+H1 = (20.689657477242, 33.1035800524193)
+I1 = (19.2805899268653, 33.0722674401887)
+J1 = (19.2962462329806, 29.7531305437458)
+K1 = (19.0614016412512, 29.393035503094)
+L1 = (18.7482755189452, 29.236472441941)
+M1 = (18.4508057027546, 29.1425346052493)
+N1 = (18.1689921926793, 29.3147539725175)
[…]
