+++ dev/literate_tutorials/curve_bounded.jl
@@ -1,0 +1,284 @@
+# # Triangulating Curve-Bounded Domains 
+# 
+
+# In this tutorial, we show how we can triangulate a domain that 
+# is defined by curves rather than by straight lines. This tutorial makes use 
+# of functions introduced in the [refinement tutorial](../tutorials/refinement.md).
+#
+# Let us start by loading in the packages we will need.
+using DelaunayTriangulation
+using DelaunayTriangulation: EllipticalArc # CairoMakie also exports this
+using CairoMakie
+using StableRNGs
+using LinearAlgebra
+using ReferenceTests #src
+using Test #src
+fig_path = joinpath(@__DIR__, "../figures") #src
+
+# ## Curves 
+# In this package, only a small subset of curves are provided (although 
+# arbitrary curves could be provided by the user, as discussed in 
+# [here](../api/curves.md). The curves, defined as parametric curves over 
+# a parameter $t ∈ [0, 1]$, are:
+#
+# - [`LineSegment`](@ref): Just a simple line segment between two points.
+# - [`CircularArc`](@ref): A circular arc defined between two points and a center.
+# - [`EllipticalArc`](@ref): An elliptical arc defined by two points, a center, a major radius, a minor radius, and an angle of rotation.
+# - [`BezierCurve`](@ref): A Bézier curve defined by a set of control points.
+# - [`BSpline`](@ref): A B-spline curve defined by a set of control points. Defaults to a cubic B-spline.
+# - [`CatmullRomSpline`](@ref): A Catmull-Rom spline defined by a set of control points.
+#
+# This set of curves is sufficient to represent a large number of curves.
+#
+# ## A Circular Domain
+# We start with a simple domain: a circle. In particular, we consider a circle of 
+# radius $r = 2$ centered at $(x, y) = (1/2, 2)$. In past tutorials, we would have defined a 
+# domain like this by a set of points around a boundary, e.g.:
+n = 50
+r = 2.0
+xc, yc = 1 / 2, 2.0
+θ = range(0, 2π, length=n + 1) |> collect;
+θ[end] = θ[begin];
+x = xc .+ r * cos.(θ)
+y = yc .+ r * sin.(θ);
+
+# One problem with this approach is that we had to decide what value of $n$ to use for discretising this 
+# boundary. Instead, we can use `CircularArc`. For this, we use:
+p = (xc + r, yc)
+c = (xc, yc)
+arc = CircularArc(p, p, c)
+
+# Here, the syntax is `CircularArc(first_point, last_point, centre)`. Since the circle is closed, we 
+# use `p` for both `first_point` and `last_point`. Notice that the `arc` is a function. In particular,
+typeof(arc) |> supertype
+
+# If we wanted to look at this circle, we would need to evaluate it at a set of $t \in [0, 1]$.
+t = LinRange(0, 1, 2500)
+points = arc.(t)
+
+#-
+fig, ax, sc = lines(points)
+
+# Let's now triangulate this domain. We need to put the arc into its own vector, and we still need to pass a set of 
+# points into `triangulate`:
+points = NTuple{2,Float64}[]
+rng = StableRNG(123)
+tri = triangulate(points; boundary_nodes=[arc], rng)
+
+#-
+fig, ax, sc = triplot(tri)
+fig
+@test_reference joinpath(fig_path, "triangulate_curve_bounded_ex_1.png") fig #src
+
[…]
