+++ versions.js
@@ -1,0 +1,12 @@
+var DOC_VERSIONS = [
+  "stable",
+  "v1.0",
+  "v0.8",
+  "v0.7",
+  "v0.6",
+  "v0.5",
+  "v0.4",
+  "dev",
+];
+var DOCUMENTER_NEWEST = "v1.0.3";
+var DOCUMENTER_STABLE = "stable";

[…]
