+++ dev/literate_tutorials/refinement.jl
@@ -25,7 +25,7 @@ fig_path = joinpath(@__DIR__, "../figures") #src
 # We will constrain the triangulation such that the minimum angle is 
 # 30 degrees, and the maximum area of a triangulation is 1% of the triangulation's 
 # total area. Note that below we need to make sure `points` is mutable, else 
-# it is not possible to push points into the triangulation. Here we use a tuple, but you 
+# it is not possible to push points into the triangulation. Here we use a vector, but you 
 # could also use e.g. an `ElasticMatrix` from [ElasticArrays.jl](https://github.com/JuliaArrays/ElasticArrays.jl).
 rng = StableRNG(123)
 x = rand(rng, 50)
@@ -49,16 +49,16 @@ fig, ax, sc = triplot(orig_tri, axis=(title="Pre-refinement",))
 ax = Axis(fig[1, 2], title="Post-refinement")
 triplot!(ax, tri)
 fig
-@test_reference joinpath(fig_path, "mesh_refinement_ex_1.png") fig #src
+@test_reference joinpath(fig_path, "mesh_refinement_ex_1.png") fig by=psnr_equality(10) #src
 
 # The triangulation is now much finer. There are still some parts with 
-# many more triangles than other regions, but this is most nearly a boundary 
+# many more triangles than other regions, but these are mostly near a boundary 
 # or where was a cluster of random points. If we wanted, we could refine again 
 # to try and improve this.
 refine!(tri; min_angle=30.0, max_area=0.001A, rng) # 0.1% instead of 1%
 fig, ax, sc = triplot(tri)
 fig
-@test_reference joinpath(fig_path, "mesh_refinement_ex_2.png") fig #src
+@test_reference joinpath(fig_path, "mesh_refinement_ex_2.png") fig by=psnr_equality(10) #src
 
 # The quality has now been improved. We could also try improving the minimum 
 # angle further, but even 30 is a bit closer to the limit of convergence (which is 
@@ -74,7 +74,7 @@ statistics(test_tri)
 # resulting triangulation is given below:
 fig, ax, sc = triplot(test_tri)
 fig
-@test_reference joinpath(fig_path, "mesh_refinement_ex_3.png") fig #src
+@test_reference joinpath(fig_path, "mesh_refinement_ex_3.png") fig by=psnr_equality(10) #src
 
 # This is certainly not a suitable triangulation.
 
@@ -92,7 +92,7 @@ hist!(ax, rad2deg.(angles), bins=20:2:60)
 vlines!(ax, [30.0], color=:red)
 resize_to_layout!(fig)
 fig
-@test_reference joinpath(fig_path, "mesh_refinement_ex_4.png") fig #src
+@test_reference joinpath(fig_path, "mesh_refinement_ex_4.png") fig by=psnr_equality(10) #src
 
 # We see that indeed many of the triangle areas are very small, and the angles 
 # are all greater than 30 degrees.
@@ -119,14 +119,14 @@ rng = StableRNG(456)
 tri = triangulate(points; boundary_nodes, rng)
 fig, ax, sc = triplot(tri)
 fig
-@test_reference joinpath(fig_path, "mesh_refinement_ex_5.png") fig #src
+@test_reference joinpath(fig_path, "mesh_refinement_ex_5.png") fig by=psnr_equality(10) #src
 
 # Let us now refine this triangulation. 
 A = get_area(tri)
 refine!(tri; min_angle=27.3, max_area=0.01A, rng)
 fig, ax, sc = triplot(tri)
 fig
-@test_reference joinpath(fig_path, "mesh_refinement_ex_6.png") fig #src
[…]
