+++ versions.js
@@ -8,5 +8,5 @@ var DOC_VERSIONS = [
   "v0.4",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.0.3";
+var DOCUMENTER_NEWEST = "v1.0.4";
 var DOCUMENTER_STABLE = "stable";
