### 2024-07-02 e942c18
Update NEWS.md
