### 2024-07-04 83713b5
Revert "Fix some type instabilities and other fixes"
