+++ test/utils.jl
@@ -4,6 +4,8 @@ using LinearAlgebra
 using BenchmarkTools
 using StableRNGs
 
+include("./helper_functions.jl")
+
 @testset "is_true" begin
       @test DT.is_true(true)
       @test DT.is_true(Val(true))
@@ -686,7 +688,7 @@ end
       @test !DT.is_circular(x)
       push!(x, x[begin])
       @test DT.is_circular(x)
-      @test DT.is_circular(Float64[])
+      @test DT.is_circular([])
 end
 
 @testset "circular_equality" begin
@@ -703,7 +705,7 @@ end
             @test DT.circular_equality(x, y) && DT.circular_equality(y, x)
       end
       @test DT.circular_equality([3, 2, 1, 13, 12, 11, 5, 4, 3], [1, 13, 12, 11, 5, 4, 3, 2, 1])
-      @test DT.circular_equality(Float64[], Float64[])
+      @test DT.circular_equality([], [])
 end
 
 @testset "get_surrounding_polygon" begin
@@ -1230,18 +1232,9 @@ end
             for T in each_ghost_triangle(tri)
                   T = DT.sort_triangle(T)
                   i, j, k = triangle_vertices(T)
-                  w = get_adjacent(tri, j, i)
-                  V = (j, i, w)
+                  V = (j, i, get_adjacent(tri, j, i))
                   @test DT.replace_boundary_triangle_with_ghost_triangle(tri, V) == T
                   @test DT.replace_ghost_triangle_with_boundary_triangle(tri, T) == V
-                  if (i, j) ∉ ((624, 625), (1, 26), (2, 1), (625, 600)) # corners of the lattice have two associated ghosts
-                        V = (i, w, j)
-                        @test DT.replace_boundary_triangle_with_ghost_triangle(tri, V) == T
-                        @test DT.replace_ghost_triangle_with_boundary_triangle(tri, T) == (j, i, w)
-                        V = (w, j, i)
-                        @test DT.replace_boundary_triangle_with_ghost_triangle(tri, V) == T
-                        @test DT.replace_ghost_triangle_with_boundary_triangle(tri, T) == (j, i, w)
-                  end
             end
       end
 end
@@ -1327,9 +1320,9 @@ end
       ))
 end
 
-@testset "getxy" begin
-      @test DT.getxy((0.3, 0.5)) == (0.3, 0.5)
-      @test DT.getxy((0.3f0, 0.7f0)) == (Float64(0.3f0), Float64(0.7f0))
+@testset "_getxy" begin
+      @test DT._getxy((0.3, 0.5)) == (0.3, 0.5)
+      @test DT._getxy((0.3f0, 0.7f0)) == (Float64(0.3f0), Float64(0.7f0))
 end
 
 @testset "norm" begin
@@ -1413,15 +1406,10 @@ end
       @test chain == [151, 96]
 end
 
-#=
-Tuple{Bool, Bool, Tuple{Int64, Int64, Int64}, Tuple{Float64, Float64}, Int64, Int64, Tuple{Float64, Float64}, Tuple{Float64, Float64}}, 
-Tuple{Bool, Bool, Tuple{Int64, Int64, Int64},         Vector{Float64}, Int64, Int64, Vector{Float64}, Vector{Float64}}}`, which is not a concrete type.
-=#
-
 @testset "dist" begin
-      for i in 1:10
+      for i in 1:100
             tri = triangulate(rand(StableRNG(i), 2, 500); rng=StableRNG(i + 1))
-            for j in 1:10
+            for j in 1:100
                   p = randn(StableRNG(i * j), 2)
[…]
