+++ test/point_location/jump_and_march.jl
@@ -5,7 +5,7 @@ using Random
 using StableRNGs
 using StatsBase
 
-
+include("../helper_functions.jl")
 
 tri, label_map, index_map = simple_geometry()
 add_ghost_triangles!(tri)
@@ -106,7 +106,7 @@ rep[3].y = mean([12.0, 6.0, 2.0, 4.0, 6.0, 10.0])
     ny = 10
     tri3 = DT.triangulate_rectangle(a, b, c, d, nx, ny, delete_ghosts=false)
 
-    for (tri_idx, tri) in enumerate((tri, tri2, tri3))
+    for tri in (tri, tri2, tri3)
         DT.compute_representative_points!(tri)
         rep = DT.get_representative_point_list(tri)
         if !(tri === tri2 || tri === tri3)
@@ -122,16 +122,15 @@ rep[3].y = mean([12.0, 6.0, 2.0, 4.0, 6.0, 10.0])
         end
 
         @testset "Test that we can find a point in every triangle" begin
-            for run in 1:6
+            for _ in 1:36
                 for V in each_triangle(tri.triangles)
-                    rand() < 0.5 && continue # skip 50%
                     if !DT.is_exterior_ghost_triangle(tri, triangle_vertices(V)...)
                         i, j, k = triangle_vertices(V)
                         p, q, r = get_point(tri, i, j, k)
                         local c
                         c = (p .+ q .+ r) ./ 3
                         for k in DT.each_solid_vertex(tri)
-                            _V = jump_and_march(tri, c; k, concavity_protection=true)
+                            _V = jump_and_march(tri, c; k)
                             @test DT.is_positively_oriented(DT.triangle_orientation(tri, _V))
                             if !DT.is_ghost_triangle(_V...)
                                 @test DT.compare_triangles(_V, V) &&
@@ -150,6 +149,9 @@ rep[3].y = mean([12.0, 6.0, 2.0, 4.0, 6.0, 10.0])
                                 if i1 ≠ i2
                                     i1 = i1 + 1
                                 end
+                                if i1 ≠ i2
+                                    @test false
+                                end
                                 _V = DT.construct_triangle(typeof(V), i1, DT.getj(V1), DT.getk(V1))
                                 @test DT.compare_triangles(_V, V) &&
                                       DT.is_inside(DT.point_position_relative_to_triangle(tri,
@@ -160,30 +162,28 @@ rep[3].y = mean([12.0, 6.0, 2.0, 4.0, 6.0, 10.0])
                     end
                 end
             end
-        end
 
-        @testset "Test that we don't break for points already in the triangulation" begin
-            for _ in 1:6
-                for k in DT.each_solid_vertex(tri)
-                    rand() < 1 / 2 && continue
-                    for j in DT.each_solid_vertex(tri)
-                        _V = jump_and_march(tri, get_point(tri, k); k=j)
-                        @test k ∈ triangle_vertices(_V)
-                        @test DT.is_positively_oriented(DT.triangle_orientation(tri, _V))
+            @testset "Test that we don't break for points already in the triangulation" begin
+                for _ in 1:36
+                    for k in DT.each_solid_vertex(tri)
+                        for j in DT.each_solid_vertex(tri)
+                            _V = jump_and_march(tri, get_point(tri, k); k=j)
+                            @test k ∈ triangle_vertices(_V)
+                            @test DT.is_positively_oriented(DT.triangle_orientation(tri, _V))
+                        end
                     end
                 end
             end
-        end
 
[…]
