### 2024-07-06 48c0fe2
Disable deepcopy
