### 2024-07-07 0df001b
Allow for ExactPredicates to be toggled
