+++ test/triangulation/bowyer_watson.jl
@@ -4,8 +4,7 @@ using Random
 using StableRNGs
 using CairoMakie
 using DataStructures
-
-
+using Preferences
 
 @testset "Getting the correct order" begin
     points = rand(2, 50)
@@ -98,12 +97,14 @@ end
     @test validate_triangulation(tri)
 end
 
-@testset "Lots of collinearity" begin
-    _tri = triangulate_rectangle(-3.0, 2.0, 5.0, 17.3, 23, 57; single_boundary=true)
-    @test validate_triangulation(_tri)
-    for _ in 1:100
-        tri = DT.triangulate(_tri.points)
-        @test validate_triangulation(tri)
+if load_preference(DelaunayTriangulation, "USE_EXACTPREDICATES", true)
+    @testset "Lots of collinearity" begin
+        _tri = triangulate_rectangle(-3.0, 2.0, 5.0, 17.3, 23, 57; single_boundary=true)
+        @test validate_triangulation(_tri)
+        for _ in 1:100
+            tri = DT.triangulate(_tri.points)
+            @test validate_triangulation(tri)
+        end
     end
 end
 
