+++ test/runtests.jl
@@ -40,11 +40,6 @@ using StructEquality
 using Aqua
 using Test
 
-@show DelaunayTriangulation.USE_EXACTPREDICATES
-
-throw("...")
-
-
 const ALL_TEST_SCRIPTS = Set{String}()
 const NON_TEST_SCRIPTS = Set{String}(["helper_functions.jl", "triangulation_validation.jl", "runtests.jl", "triangulation\\weighted.jl", "triangulation/weighted.jl"])
 include("helper_functions.jl")
