### 2024-07-07 37f22d9
Document the disabling of ExactPredicates; add the triangulation validator into the package
