+++ src/validation.jl
@@ -1,5 +1,3 @@
-# maybe this should belong inside src for users at some point?
-
 #=
 """
     abstract type AbstractTriangulationState 
@@ -19,6 +17,22 @@ abstract type AbstractTriangulationState end
 test_state(state::AbstractTriangulationState) = state.flag
 Base.show(io::IO, state::AbstractTriangulationState) = print(io, summary(state))
 
+function compare_edge_vectors(E1, E2)
+    E1s = sort_edge_vector(collect(E1))
+    E2s = sort_edge_vector(collect(E2))
+    return E1s == E2s
+end
+
+function sort_edge_vector(E)
+    sorted_E = similar(E)
+    for i in eachindex(E)
+        u, v = E[i]
+        e = (min(u, v), max(u, v))
+        sorted_E[i] = e
+    end
+    return sort(sorted_E)
+end
+
 struct TriangleOrientationState <: AbstractTriangulationState
     flag::Bool
     bad_triangle::NTuple{3,Int}
@@ -34,12 +48,12 @@ function Base.summary(state::TriangleOrientationState)
 end
 function test_triangle_orientation(tri)
     for T in each_solid_triangle(tri)
-        cert = DT.triangle_orientation(tri, T)
-        flag = DT.is_positively_oriented(cert)
-        orientation = DT.is_positively_oriented(cert) ? :positively : DT.is_negatively_oriented(cert) ? :negatively : :degenerately
+        cert = triangle_orientation(tri, T)
+        flag = is_positively_oriented(cert)
+        orientation = is_positively_oriented(cert) ? :positively : is_negatively_oriented(cert) ? :negatively : :degenerately
         !flag && return TriangleOrientationState(flag, Int.(triangle_vertices(T)), orientation)
     end
-    return TriangleOrientationState(true, (DT.∅, DT.∅, DT.∅), :positive)
+    return TriangleOrientationState(true, (∅, ∅, ∅), :positive)
 end
 
 struct DelaunayCriterionState <: AbstractTriangulationState
@@ -52,7 +66,7 @@ function Base.summary(state::DelaunayCriterionState)
     if test_state(state)
         return "All the triangles are Delaunay."
     else
-        if state.bad_vertex !== DT.∅
+        if state.bad_vertex !== ∅
             return "The Delaunay criterion does not hold for the triangle-vertex pair ($(state.bad_triangle), $(state.bad_vertex))."
         else
             return "The test of the Delaunay criterion failed as there was a BoundsError when testing the visibility."
@@ -62,23 +76,23 @@ end
 function test_delaunay_criterion(tri)
     try
         points = get_points(tri)
-        triangle_tree = DT.BoundaryRTree(points)
-        segment_tree = DT.BoundaryRTree(points)
-        failures = Tuple{DT.triangle_type(tri),DT.integer_type(tri)}[]
+        triangle_tree = BoundaryRTree(points)
+        segment_tree = BoundaryRTree(points)
+        failures = Tuple{triangle_type(tri),integer_type(tri)}[]
         for T in each_solid_triangle(tri)
-            i, j, k = DT.triangle_vertices(T)
+            i, j, k = triangle_vertices(T)
             p, q, r = get_point(tri, i, j, k)
-            c = DT.triangle_circumcenter(p, q, r)
-            cr = DT.triangle_circumradius(p, q, r)
+            c = triangle_circumcenter(p, q, r)
+            cr = triangle_circumradius(p, q, r)
             xmin, xmax = getx(c) - cr, getx(c) + cr
[…]
