+++ test/runtests.jl
@@ -1,11 +1,13 @@
 # setup LocalPreferences.toml 
 using Preferences 
-USE_EXACTPREDICATES = ENV["USE_EXACTPREDICATES"]
+USE_EXACTPREDICATES = get(ENV, "USE_EXACTPREDICATES", "default")
 if USE_EXACTPREDICATES == "true"
     set_preferences!("DelaunayTriangulation", "USE_EXACTPREDICATES" => true)
 elseif USE_EXACTPREDICATES == "false" 
     set_preferences!("DelaunayTriangulation", "USE_EXACTPREDICATES" => false)
 end # if USE_EXACTPREDICATES == "default", do nothing
+
+@info "Testing with USE_EXACTPREDICATES = $USE_EXACTPREDICATES"
 
 # get all the compilation out of the way
 using BenchmarkTools
