+++ test/helper_functions.jl
@@ -2235,12 +2235,12 @@ export get_child_from_tree
 export traverse_tree
 export poor_triangulation_example
 export example_triangulation
-export test_adjacent2vertex_map_matches_adjacent_map
-export test_state 
-export test_adjacent_map_matches_adjacent2vertex_map
-export test_each_edge_has_two_incident_triangles
-export test_triangle_orientation
-export test_iterators
+export DelaunayTriangulation.test_adjacent2vertex_map_matches_adjacent_map
+export DelaunayTriangulation.test_state 
+export DelaunayTriangulation.test_adjacent_map_matches_adjacent2vertex_map
+export DelaunayTriangulation.test_each_edge_has_two_incident_triangles
+export DelaunayTriangulation.test_triangle_orientation
+export DelaunayTriangulation.test_iterators
 export example_empty_triangulation
 export shewchuk_example_constrained
 export test_segment_triangle_intersections
