### 2024-07-07 8989cc9
Add NEWS, fix contributing, and fixes
