+++ docs/src/extended/utils.md
@@ -33,5 +33,5 @@ DefaultAdjacentValue
 GhostVertex
 ε
 ∅
-USE_EXACTPREDICATE
+USE_EXACTPREDICATES
 ```
