### 2024-07-08 9f14596
Change jump_and_march to find_triangle
