+++ test/point_location/ghost_search.jl
@@ -30,87 +30,87 @@ k_list = unique(boundary_nodes[1][1])
 @testset "Finding points outside" begin
     q = [-3.94232, 15.6067]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test i == index_map["h"] && j == index_map["g"]
     end
     q = [-5.0, 5.0]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test i == index_map["a"] && j == index_map["h"]
     end
     q = [-4.8934362297993, -5.6330436693732]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test i == index_map["b"] && j == index_map["a"]
     end
     q = [15.0, -5.0]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test i == index_map["c"] && j == index_map["b"]
     end
     q = [25.0, 0.0]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test i == index_map["d"] && j == index_map["c"]
     end
     q = [25.0, 15.0]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
-        @inferred DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
+        @inferred DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test i == index_map["e"] && j == index_map["d"]
     end
     q = [13.717, 22.3597]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test i == index_map["f"] && j == index_map["e"]
     end
     q = [2.3469, 22.61468]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         @test i == index_map["g"] && j == index_map["f"]
     end
     q = [-5.0, 25.0]
     for k in k_list
-        i, j = DT.exterior_jump_and_march(tri, k, q)
+        i, j = DT.exterior_find_triangle(tri, k, q)
         @test DT.is_ghost_vertex(get_adjacent(adj, i, j))
         f, g, h = index_map["f"], index_map["g"], index_map["h"]
         @test (j, i) == (f, g) || (j, i) == (g, h)
     end
[…]
