### 2024-07-08 6e940b4
Change jump_and_march to find_triangle (#133)
* Change jump_and_march to find_triangle

* news

* news

* oops

* filename
