+++ src/algorithms/point_location/jump_and_march.jl
@@ -36,7 +36,7 @@ end
 @doc """
     select_initial_point(tri::Triangulation, q; kwargs...) -> Vertex
 
-Selects the initial point for [`jump_and_march`](@ref) to start from.
+Selects the initial point for [`find_triangle`](@ref) to start from.
 
 # Arguments 
 - `tri::Triangulation`: The [`Triangulation`](@ref).
@@ -139,7 +139,7 @@ end
 """
     select_initial_triangle_interior_vertex(tri::Triangulation, k, q, store_history=Val(false), history=nothing, rng::AbstractRNG=Random.default_rng()) -> (Point, Vertex, Vertex, Point, Point)
 
-Selects the initial triangle for [`jump_and_march`](@ref) to start from, for the case where `k` is an interior vertex.
+Selects the initial triangle for [`find_triangle`](@ref) to start from, for the case where `k` is an interior vertex.
 
 # Arguments 
 - `tri::Triangulation`: The [`Triangulation`](@ref).
@@ -159,7 +159,7 @@ Selects the initial triangle for [`jump_and_march`](@ref) to start from, for the
 !!! warning "Non-convex geometries"
 
     This function assumes that the geometry is convex. To deal with this, when an infinite loop is detected 
-    we return `∅` for both `i` and `j`, and then let [`jump_and_march`](@ref) handle how to 
+    we return `∅` for both `i` and `j`, and then let [`find_triangle`](@ref) handle how to 
     correct the algorithm from there.    
 
 # Extended help
@@ -261,7 +261,7 @@ function fix_initial_collinear_edge_for_interior_vertex(tri::Triangulation, k, q
                \pᵢ
         Then we want to know if q is to the right of `ppⱼ` (here, `line_cert_j` is the collinear certificate) 
         or to the left. If it is to the left, then it is on the edge `ppⱼ`. This is not a problem, but we let 
-        [`jump_and_march`](@ref) handle this case somewhere else. Thus, we should want to find a case where `q` 
+        [`find_triangle`](@ref) handle this case somewhere else. Thus, we should want to find a case where `q` 
         is right of `ppⱼ` and not handle this interior edge case, so we check `!is_left(on_edge_cert)`. 
         =#
         if !is_left(on_edge_cert)
@@ -360,7 +360,7 @@ See also [`check_for_intersections_with_adjacent_boundary_edges`](@ref), which u
 !!! warning "Non-convex geometries"
 
     This function assumes that the geometry is convex. The function will still be able to return, but `is_outside(q_pos_cert)` may not necessarily mean `q` 
-    is outside of the triangulation. The main function [`jump_and_march`](@ref) will have to restart the algorithm if it is found that `is_outside(q_pos_cert)` 
+    is outside of the triangulation. The main function [`find_triangle`](@ref) will have to restart the algorithm if it is found that `is_outside(q_pos_cert)` 
     was incorrect.
 
 # Extended help 
@@ -588,7 +588,7 @@ function check_for_intersections_with_triangle_left_to_boundary_vertex(tri::Tria
 end
 
 """
-    exterior_jump_and_march(tri::Triangulation, k, q, ghost_vertex=𝒢) -> (Vertex, Vertex)
+    exterior_find_triangle(tri::Triangulation, k, q, ghost_vertex=𝒢) -> (Vertex, Vertex)
 
 Given a query point `q` outside of the triangulation, find the ghost triangle containing `q`.
 
@@ -610,21 +610,21 @@ Given a query point `q` outside of the triangulation, find the ghost triangle co
 # Extended help
 This function works by first finding the position of `q` relative to `pₘp`, where `pₘ` is the representative point for 
[…]
