+++ docs/src/manual/overview.md
@@ -13,3 +13,4 @@ This section gives a manual for certain parts of the package. Little code will b
 - [Geometrical Predicates](predicates.md): How geometric predicates are defined in this package.
 - [Triangulation Output](triangulation_output.md): How the output of a triangulation is defined.
 - [Voronoi Tessellation Output](voronoi_output.md): How the output of a Voronoi tessellation is defined.
+- [Disabling Exact Predicates](disabling_ea.md): How to disable ExactPredicates.jl.
