+++ src/data_structures.jl
@@ -1,0 +1,43 @@
+include("data_structures/queue/max_priority_queue.jl")
+include("data_structures/queue/queue.jl")
+include("data_structures/trees/bst.jl")
+include("data_structures/trees/rtree.jl")
+include("data_structures/trees/polygon_hierarchy.jl")
+include("data_structures/triangulation/adjacent.jl")
+include("data_structures/triangulation/adjacent2vertex.jl")
+include("data_structures/triangulation/graph.jl")
+include("data_structures/mesh_refinement/curves.jl")
+include("data_structures/representative_coordinates/representative_coordinates.jl")
+include("data_structures/representative_coordinates/cell.jl")
+include("data_structures/representative_coordinates/cell_queue.jl")
+include("data_structures/convex_hull.jl")
+include("data_structures/triangulation/triangulation.jl")
+include("data_structures/triangulation/triangulation_cache.jl")
+include("data_structures/mesh_refinement/boundary_enricher.jl")
+include("data_structures/triangulation/methods/adjacent.jl")
+include("data_structures/triangulation/methods/adjacent2vertex.jl")
+include("data_structures/triangulation/methods/boundary_curves.jl")
+include("data_structures/triangulation/methods/boundary_edge_map.jl")
+include("data_structures/triangulation/methods/boundary_nodes.jl")
+include("data_structures/triangulation/methods/convex_hull.jl")
+include("data_structures/triangulation/methods/exterior_curve_indices.jl")
+include("data_structures/triangulation/methods/ghost_vertex_map.jl")
+include("data_structures/triangulation/methods/ghost_vertex_ranges.jl")
+include("data_structures/triangulation/methods/graph.jl")
+include("data_structures/triangulation/methods/iterators.jl")
+include("data_structures/triangulation/methods/points.jl")
+include("data_structures/triangulation/methods/representative_point_list.jl")
+include("data_structures/triangulation/methods/segments.jl")
+include("data_structures/triangulation/methods/triangles.jl")
+include("data_structures/triangulation/methods/weights.jl")
+include("data_structures/triangulation/methods/checks.jl")
+include("data_structures/point_location_history.jl")
+include("data_structures/statistics/individual_triangle_statistics.jl")
+include("data_structures/statistics/triangulation_statistics.jl")
+include("data_structures/mesh_refinement/insertion_event_history.jl")
+include("data_structures/mesh_refinement/refinement_constraints.jl")
+include("data_structures/mesh_refinement/refinement_queue.jl")
+include("data_structures/mesh_refinement/refinement_arguments.jl")
+include("data_structures/voronoi.jl")
+include("data_structures/polygon.jl")
+include("data_structures/shuffled_polygon_linked_list.jl")
