+++ src/exports.jl
@@ -1,0 +1,103 @@
+const jump_and_march = find_triangle # Remove in 2.0
+
+export
+    each_triangle,
+    each_solid_triangle,
+    each_ghost_triangle,
+    each_edge,
+    each_solid_edge,
+    each_ghost_edge,
+    each_vertex,
+    each_solid_vertex,
+    each_ghost_vertex,
+    num_triangles,
+    num_solid_triangles,
+    num_ghost_triangles,
+    num_edges,
+    num_solid_edges,
+    num_ghost_edges,
+    num_vertices,
+    num_solid_vertices,
+    num_ghost_vertices,
+    get_points,
+    get_boundary_nodes,
+    get_interior_segments,
+    get_all_segments,
+    get_adjacent,
+    get_adjacent2vertex,
+    get_ghost_vertex_map,
+    get_ghost_vertex_ranges,
+    get_ghost_vertex_range,
+    get_triangles,
+    get_boundary_edge_map,
+    statistics,
+    each_segment,
+    get_area,
+    get_all_stat,
+    num_polygons,
+    get_polygon,
+    each_polygon,
+    get_polygon_point,
+    each_generator,
+    get_generator,
+    each_polygon_index,
+    each_polygon_vertex,
+    num_polygon_vertices,
+    get_neighbours,
+    get_convex_hull,
+    get_convex_hull_vertices,
+    triangle_vertices,
+    edge_vertices,
+    getx,
+    gety,
+    getxy,
+    get_point,
+    num_boundary_edges,
+    add_ghost_triangles!,
+    delete_ghost_triangles!,
+    add_point!,
+    add_triangle!,
+    delete_triangle!,
+    flip_edge!,
+    split_edge!,
+    split_triangle!,
+    legalise_edge!,
+    delete_point!,
+    add_segment!,
+    lock_convex_hull!,
+    unlock_convex_hull!,
+    triangulate,
+    triangulate_rectangle,
+    triangulate_convex,
+    brute_force_search,
+    find_triangle,
+    jump_and_march,
+    convert_boundary_points_to_indices,
+    refine!,
+    centroidal_smooth,
+    get_polygon_coordinates,
+    get_nearest_neighbour,
+    convex_hull,
+    get_convex_hull,
+    get_graph,
+    voronoi,
+    retriangulate,
+    Triangulation,
+    VoronoiTessellation,
+    ConvexHull,
+    TriangulationStatistics,
+    convex_hull!,
+    add_weight!,
+    get_weight,
+    get_weights,
+    CircularArc,
+    LineSegment,
+    EllipticalArc,
+    BezierCurve,
+    BSpline,
+    CatmullRomSpline,
+    find_polygon,
+    get_boundary_curves,
+    map_ghost_vertex,
+    polygon_bounds,
+    num_neighbours
