+++ src/geometric_primitives.jl
@@ -1,0 +1,4 @@
+include("geometric_primitives/boundary_nodes.jl")
+include("geometric_primitives/edges.jl")
+include("geometric_primitives/points.jl")
+include("geometric_primitives/triangles.jl")
