+++ src/algorithms.jl
@@ -1,0 +1,38 @@
+include("algorithms/intersections/rtree.jl")
+include("algorithms/convex_hull.jl")
+include("algorithms/point_location/brute_force.jl")
+include("algorithms/point_location/jump_and_march.jl")
+include("algorithms/point_location/nearest_neighbour.jl")
+include("algorithms/point_location/find_polygon.jl")
+include("algorithms/polygon_clipping/liang_barsky.jl")
+include("algorithms/polygon_clipping/sutherland_hodgman.jl")
+include("algorithms/pole_of_inaccessibility.jl")
+include("algorithms/triangulation/constrained_triangulation.jl")
+include("algorithms/triangulation/check_args.jl")
+include("algorithms/triangulation/main.jl")
+include("algorithms/triangulation/triangulate_curve_bounded.jl")
+include("algorithms/triangulation/mesh_refinement.jl")
+include("algorithms/triangulation/triangulate_convex.jl")
+include("algorithms/triangulation/triangulate_rectangle.jl")
+include("algorithms/triangulation/unconstrained_triangulation.jl")
+include("algorithms/triangulation/basic_operations/add_boundary_information.jl")
+include("algorithms/triangulation/basic_operations/add_ghost_triangles.jl")
+include("algorithms/triangulation/basic_operations/add_point.jl")
+include("algorithms/triangulation/basic_operations/add_segment.jl")
+include("algorithms/triangulation/basic_operations/add_triangle.jl")
+include("algorithms/triangulation/basic_operations/clear_empty_features.jl")
+include("algorithms/triangulation/basic_operations/delete_ghost_triangles.jl")
+include("algorithms/triangulation/basic_operations/delete_holes.jl")
+include("algorithms/triangulation/basic_operations/delete_point.jl")
+include("algorithms/triangulation/basic_operations/delete_triangle.jl")
+include("algorithms/triangulation/basic_operations/flip_edge.jl")
+include("algorithms/triangulation/basic_operations/legalise_edge.jl")
+include("algorithms/triangulation/basic_operations/lock_convex_hull.jl")
+include("algorithms/triangulation/basic_operations/split_edge.jl")
+include("algorithms/triangulation/basic_operations/split_triangle.jl")
+include("algorithms/triangulation/basic_operations/unlock_convex_hull.jl")
+include("algorithms/voronoi/centroidal.jl")
+include("algorithms/voronoi/clipped_coordinates.jl")
+include("algorithms/voronoi/clipped.jl")
+include("algorithms/voronoi/main.jl")
+include("algorithms/voronoi/unbounded.jl")
