+++ src/predicates.jl
@@ -1,0 +1,4 @@
+include("predicates/certificate.jl")
+include("predicates/exactpredicates_definitions.jl")
+include("predicates/predicates.jl")
+include("predicates/boundaries_and_ghosts.jl")
