+++ src/DelaunayTriangulation.jl
@@ -1,294 +1,20 @@
 module DelaunayTriangulation
 
-"""
-    DefaultAdjacentValue = 0 
+include("setup.jl")
 
-Default value used for representing an empty result 
-from an adjacency query.
-"""
-const DefaultAdjacentValue = 0
-
-"""
-    ∅ = DefaultAdjacentValue
-
-Alias for [`DefaultAdjacentValue`](@ref).
-"""
-const ∅ = DefaultAdjacentValue
-
-"""
-    GhostVertex = -1
-
-Number used for representing initial ghost vertices. 
-All other ghost vertices are derived from subtracting from 
-this number. See https://juliageometry.github.io/DelaunayTriangulation.jl/stable/manual/ghost_triangles/.
-"""
-const GhostVertex = -1
-
-"""
-    𝒢 = GhostVertex
-
-Alias for [`𝒢`](@ref).
-"""
-const 𝒢 = GhostVertex
-
-"""
-    ε = sqrt(eps(Float64))
-
-Number used as a tolerance in certain functions, e.g. 
-for mesh refinement when using [`check_precision`](@ref) to 
-avoid degenerate circumcenters.
-"""
-const ε = sqrt(eps(Float64))
-
-const INF_WARN = Ref(true)
-"""
-    toggle_inf_warn!()
-
-Toggle the warning for infinite circumcenters in the Voronoi tessellation.
-By default, this warning is enabled.
-"""
-toggle_inf_warn!() = (INF_WARN[] = !INF_WARN[])
-
-@static if VERSION ≥ v"1.6"
-    using Preferences
+@static if USE_EXACTPREDICATES
+    using ExactPredicates
 end
-
-@static if VERSION ≥ v"1.6"
-    const USE_EXACTPREDICATES = @load_preference("USE_EXACTPREDICATES", true)::Bool
-else 
-    const USE_EXACTPREDICATES = true 
-end
-@doc """
-    USE_EXACTPREDICATES
-
-Whether to use ExactPredicates.jl for computing predicates. By default, 
-this is true, but a user can change this by defining a preference with Preferences.jl, i.e. 
-you could do the following 
-
-```julia-repl
-julia> using Preferences: set_preferences!
-
-julia> set_preferences!("DelaunayTriangulation", "USE_EXACTPREDICATES" => false)
-
-julia> using DelaunayTriangulation # load only after setting the preference
-
-julia> DelaunayTriangulation.USE_EXACTPREDICATES
-false
-```
-
-You have set `USE_EXACTPREDICATES = $USE_EXACTPREDICATES`. 
-
-!!! note "Precision"
-
-    Even if you have disabled ExactPredicates.jl, the predicates 
-    are still computed in Float64 precision.
-"""
-USE_EXACTPREDICATES
-
-using ExactPredicates
 using EnumX
 using Random
 
-include("data_structures/queue/max_priority_queue.jl")
-include("data_structures/queue/queue.jl")
-include("data_structures/trees/bst.jl")
-include("data_structures/trees/rtree.jl")
-include("data_structures/trees/polygon_hierarchy.jl")
-include("data_structures/triangulation/adjacent.jl")
-include("data_structures/triangulation/adjacent2vertex.jl")
-include("data_structures/triangulation/graph.jl")
-include("data_structures/mesh_refinement/curves.jl")
-include("data_structures/representative_coordinates/representative_coordinates.jl")
-include("data_structures/representative_coordinates/cell.jl")
-include("data_structures/representative_coordinates/cell_queue.jl")
-include("data_structures/convex_hull.jl")
-include("data_structures/triangulation/triangulation.jl")
-include("data_structures/triangulation/triangulation_cache.jl")
[…]
