+++ test/runtests.jl
@@ -1,13 +1,15 @@
 # setup LocalPreferences.toml 
 using Preferences 
-USE_EXACTPREDICATES = get(ENV, "USE_EXACTPREDICATES", "default")
-if USE_EXACTPREDICATES == "true"
-    set_preferences!("DelaunayTriangulation", "USE_EXACTPREDICATES" => true)
-elseif USE_EXACTPREDICATES == "false" 
-    set_preferences!("DelaunayTriangulation", "USE_EXACTPREDICATES" => false)
-end # if USE_EXACTPREDICATES == "default", do nothing
+PREDICATES = get(ENV, "PREDICATES", "EXACT")
+if PREDICATES == "EXACT"
+    set_preferences!("DelaunayTriangulation", "PREDICATES" => "EXACT")
+elseif PREDICATES == "INEXACT" 
+    set_preferences!("DelaunayTriangulation", "PREDICATES" => "INEXACT")
+else
+    throw("Invalid PREDICATES setting, $PREDICATES.")
+end # if PREDICATES == "default", do nothing
 
-@info "Testing with USE_EXACTPREDICATES = $USE_EXACTPREDICATES"
+@info "Testing with PREDICATES = $PREDICATES"
 
 # get all the compilation out of the way
 using BenchmarkTools
@@ -61,7 +63,7 @@ end
 
 @testset verbose = true "DelaunayTriangulation.jl" begin
     @testset verbose = true "Aqua" begin
-        Aqua.test_all(DelaunayTriangulation; ambiguities=false, project_extras=false, stale_deps = load_preference(DelaunayTriangulation, "USE_EXACTPREDICATES", true)) # don't care about julia < 1.2
+        Aqua.test_all(DelaunayTriangulation; ambiguities=false, project_extras=false, stale_deps = !USE_INEXACTPREDICATES) # don't care about julia < 1.2
         Aqua.test_ambiguities(DelaunayTriangulation) # don't pick up Base and Core...
     end    
 
