### 2024-07-08 1dcf3bb
Change USE_EXACTPREDICATES to PREDICATES (#137)
* Change USE_EXACTPREDICATES to PREDICATES

* news

* fix

* fix
