+++ test/triangulation/bowyer_watson.jl
@@ -6,6 +6,7 @@ using CairoMakie
 using DataStructures
 using Preferences
 
+
 @testset "Getting the correct order" begin
     points = rand(2, 50)
     randomise = false
@@ -29,6 +30,7 @@ using Preferences
     @test all(∈(pt_order), setdiff(DT.each_point_index(points), skip_points))
 end
 
+
 @testset "Getting the initial triangle" begin
     tri = triangulate_rectangle(0.0, 10.0, 0.0, 20.0, 11, 21)
     insertion_order = collect(DT.each_point_index(tri))
@@ -38,15 +40,17 @@ end
     @test DT.is_positively_oriented(DT.triangle_orientation(tri, initial_triangle))
 end
 
+
 @testset "Initialising the Bowyer-Watson algorithm" begin
-    tri = triangulate_rectangle(0.0, 10.0, 0.0, 20.0, 11, 21, delete_ghosts=false)
+    tri = triangulate_rectangle(0.0, 10.0, 0.0, 20.0, 11, 21, delete_ghosts = false)
     _tri = DT.Triangulation(tri.points)
     DT.initialise_bowyer_watson!(_tri, DT.each_point_index(_tri) |> collect)
     S = filter(!DT.is_ghost_triangle, _tri.triangles)
     u, v, w = first(S)
     BI = DT.𝒢
     @test get_triangles(_tri) == Set(((u, v, w), (w, v, BI), (v, u, BI), (u, w, BI)))
-    @test get_adjacent(get_adjacent(_tri)) == Dict((u, v) => w,
+    @test get_adjacent(get_adjacent(_tri)) == Dict(
+        (u, v) => w,
         (v, w) => u,
         (w, u) => v,
         (w, v) => BI,
@@ -57,49 +61,59 @@ end
         (BI, v) => u,
         (u, w) => BI,
         (w, BI) => u,
-        (BI, u) => w)
+        (BI, u) => w,
+    )
     @test get_adjacent2vertex(get_adjacent2vertex(_tri)) ==
-          Dict(BI => Set{NTuple{2,Int}}([(w, v), (v, u), (u, w)]),
-        u => Set{NTuple{2,Int}}([(v, w), (BI, v), (w, BI)]),
-        v => Set{NTuple{2,Int}}([(w, u), (BI, w), (u, BI)]),
-        w => Set{NTuple{2,Int}}([(u, v), (v, BI), (BI, u)]))
+        Dict(
+        BI => Set{NTuple{2, Int}}([(w, v), (v, u), (u, w)]),
+        u => Set{NTuple{2, Int}}([(v, w), (BI, v), (w, BI)]),
+        v => Set{NTuple{2, Int}}([(w, u), (BI, w), (u, BI)]),
+        w => Set{NTuple{2, Int}}([(u, v), (v, BI), (BI, u)]),
+    )
     @test DT.get_graph(_tri).vertices == Set([u, v, w, BI])
 end
 
+
 @testset "A basic triangulation" begin
     a, b, c, d, e, f, g = [-1.78, 5.77], [-4.96, 0.31], [0.08, -3.73], [7.74, -3.03],
-    [8.0, 3.0], [-0.6, 1.57], [3.58, 6.15]
+        [8.0, 3.0], [-0.6, 1.57], [3.58, 6.15]
     points = [a, b, c, d, e, f, g]
     rng = StableRNG(8881)
-    tri = DT.triangulate(points; rng, delete_ghosts=false)
+    tri = DT.triangulate(points; rng, delete_ghosts = false)
     @static if VERSION ≥ v"1.9"
-        @inferred DT.triangulate(points; rng, delete_ghosts=false)
[…]
