+++ src/algorithms/voronoi/unbounded.jl
@@ -9,20 +9,20 @@ Initialise a `VoronoiTessellation` from the triangulation `tri`.
 # Output
 - `vorn`: The [`VoronoiTessellation`](@ref). This tessellation is not yet filled in, as all the polygons and other fields need to be properly defined. This simply defines all the initial objects that will be pushed into.
 """
-function initialise_voronoi_tessellation(tri::Tr) where {Tr<:Triangulation}
+function initialise_voronoi_tessellation(tri::Tr) where {Tr <: Triangulation}
     I = integer_type(tri)
     T = triangle_type(tri)
     F = number_type(tri)
-    P = NTuple{2,F}
+    P = NTuple{2, F}
     polygon_points = Vector{P}()
-    circumcenter_to_triangle = Dict{I,T}()
-    triangle_to_circumcenter = Dict{T,I}()
+    circumcenter_to_triangle = Dict{I, T}()
+    triangle_to_circumcenter = Dict{T, I}()
     sizehint!(polygon_points, num_triangles(tri))
     sizehint!(circumcenter_to_triangle, num_triangles(tri))
     sizehint!(triangle_to_circumcenter, num_triangles(tri))
     cur_ghost_idx = I(0)
-    cocircular_dict = Dict{P,I}()
-    encountered_circumcenters = Dict{P,I}()
+    cocircular_dict = Dict{P, I}()
+    encountered_circumcenters = Dict{P, I}()
     cocircular_circumcenters = Set{I}()
     for V in each_triangle(tri)
         V = sort_triangle(V)
@@ -50,25 +50,26 @@ function initialise_voronoi_tessellation(tri::Tr) where {Tr<:Triangulation}
                 cocircular_dict[(cx, cy)] = num_points(polygon_points)
             end
         else
-            circumcenter_to_triangle[I(𝒢)-cur_ghost_idx] = V
+            circumcenter_to_triangle[I(𝒢) - cur_ghost_idx] = V
             triangle_to_circumcenter[V] = I(𝒢) - cur_ghost_idx
             cur_ghost_idx += I(1)
         end
     end
-    polygons = Dict{I,Vector{I}}()
+    polygons = Dict{I, Vector{I}}()
     sizehint!(polygons, num_solid_vertices(tri))
     unbounded_polygons = Set{I}()
     sizehint!(unbounded_polygons, num_ghost_edges(tri))
-    generators = Dict{I,P}()
+    generators = Dict{I, P}()
     sizehint!(generators, num_solid_vertices(tri))
     for i in each_solid_vertex(tri)
         generators[i] = get_point(tri, i)
     end
     E = edge_type(tri)
-    adj = Adjacent{I,E}()
+    adj = Adjacent{I, E}()
     boundary_polygons = Set{I}()
-    return VoronoiTessellation{Tr,P,I,T,typeof(cocircular_circumcenters),E}(tri, generators, polygon_points, polygons, circumcenter_to_triangle, triangle_to_circumcenter, unbounded_polygons, cocircular_circumcenters, adj, boundary_polygons)
+    return VoronoiTessellation{Tr, P, I, T, typeof(cocircular_circumcenters), E}(tri, generators, polygon_points, polygons, circumcenter_to_triangle, triangle_to_circumcenter, unbounded_polygons, cocircular_circumcenters, adj, boundary_polygons)
 end
+
 
 """
     prepare_add_voronoi_polygon(vorn::VoronoiTessellation, i) -> (Vector, Vector)
[…]
