### 2024-07-14 d8cbb78
Add public keywords for v1.11; Add const fields using a macro for compatibility; Test on pre-release; Add docstring to Module;
