+++ test/runtests.jl
@@ -1,9 +1,9 @@
 # setup LocalPreferences.toml 
-using Preferences 
+using Preferences
 PREDICATES = get(ENV, "PREDICATES", "EXACT")
 if PREDICATES == "EXACT"
     set_preferences!("DelaunayTriangulation", "PREDICATES" => "EXACT")
-elseif PREDICATES == "INEXACT" 
+elseif PREDICATES == "INEXACT"
     set_preferences!("DelaunayTriangulation", "PREDICATES" => "INEXACT")
 elseif PREDICATES != "DEFAULT"
     throw("Invalid PREDICATES setting, $PREDICATES.")
@@ -44,7 +44,9 @@ using StructEquality
 using Aqua
 using Test
 
-@test isempty(Docs.undocumented_names(DelaunayTriangulation))
+if isdefined(Docs, :undocumented_names)
+    @test isempty(Docs.undocumented_names(DelaunayTriangulation))
+end
 
 const ALL_TEST_SCRIPTS = Set{String}()
 const NON_TEST_SCRIPTS = Set{String}(["helper_functions.jl", "triangulation_validation.jl", "runtests.jl", "triangulation\\weighted.jl", "triangulation/weighted.jl"])
@@ -52,7 +54,7 @@ include("helper_functions.jl")
 using .HelperFunctions
 
 ct() = Dates.format(now(), "HH:MM:SS")
-function safe_include(filename; name=filename, push=true, verbose = true) # Workaround for not being able to interpolate into SafeTestset test names
+function safe_include(filename; name=filename, push=true, verbose=true) # Workaround for not being able to interpolate into SafeTestset test names
     push && push!(ALL_TEST_SCRIPTS, normpath(filename))
     mod = @eval module $(gensym()) end
     @info "[$(ct())] Testing $name"
@@ -65,9 +67,9 @@ end
 
 @testset verbose = true "DelaunayTriangulation.jl" begin
     @testset verbose = true "Aqua" begin
-        Aqua.test_all(DelaunayTriangulation; ambiguities=false, project_extras=false, stale_deps = !USE_INEXACTPREDICATES) # don't care about julia < 1.2
+        Aqua.test_all(DelaunayTriangulation; ambiguities=false, project_extras=false, stale_deps=!USE_INEXACTPREDICATES) # don't care about julia < 1.2
         Aqua.test_ambiguities(DelaunayTriangulation) # don't pick up Base and Core...
-    end    
+    end
 
     @testset verbose = true "Triangulation" begin
         safe_include("triangulation/rectangle.jl")
@@ -104,7 +106,7 @@ end
         safe_include("data_structures/curves.jl")
         safe_include("data_structures/rtree.jl")
         safe_include("data_structures/bst.jl")
-        safe_include("data_structures/polygon_hierarchy.jl", verbose = false)
+        safe_include("data_structures/polygon_hierarchy.jl", verbose=false)
     end
 
     @testset verbose = true "Predicates" begin
@@ -169,7 +171,7 @@ end
             mp4_path = joinpath(dirname(dirname(pathof(DelaunayTriangulation))), "cell_simulation.mp4")
             isfile(mp4_path) && rm(mp4_path)
         end
-    
+
         @testset verbose = true "Test the tutorials" begin
             tut_dir = joinpath(dirname(dirname(pathof(DelaunayTriangulation))), "docs", "src", "literate_tutorials")
             tut_files = readdir(tut_dir)
@@ -177,11 +179,11 @@ end
                 safe_include(joinpath(tut_dir, file); push=false)
             end
         end
-    
+
         @testset verbose = true "Test the readme example" begin
             safe_include("readme_example.jl")
         end
-    
+
[…]
