+++ docs/src/tutorials/clipped_rectangle.md
@@ -15,7 +15,7 @@ but we provide the function [`get_polygon_coordinates`](@ref) for this (this is 
 Let us now demonstrate. First, we construct a tessellation of
 some example point set.
 
-````julia
+````@example clipped_rectangle
 using DelaunayTriangulation
 using CairoMakie
 A = (-3.0, 7.0)
@@ -31,43 +31,28 @@ tri = triangulate(points)
 vorn = voronoi(tri)
 ````
 
-````
-Voronoi Tessellation.
-    Number of generators: 8
-    Number of polygon vertices: 9
-    Number of polygons: 8
-````
-
 Let us show the tessellation, and the rectangle we want to clip
 the tessellation to.
 
-````julia
+````@example clipped_rectangle
 fig, ax, sc = voronoiplot(vorn)
 a, b, c, d = -2.0, 3.0, 0.0, 7.0
 lines!(ax, [(a,c),(b,c),(b,d),(a,d),(a,c)], color = :black, linewidth = 4)
 fig
 ````
 
-```@raw html
[…]
