+++ docs/src/tutorials/constrained_interior_within_interiors.md
@@ -9,7 +9,7 @@ Now we consider triangulating a domain which has not only an
 interior boundary, but also an interior boundary inside that
 interior boundary. To start, let us load the packages.
 
-````julia
+````@example constrained_interior_within_interiors
 using DelaunayTriangulation
 using CairoMakie
 using StableRNGs
@@ -24,7 +24,7 @@ the orientation of an interior curve is the opposite
 orientation to the curve it's inside of. Here are the
 points we will be triangulating.
 
-````julia
+````@example constrained_interior_within_interiors
 curve_1 = [
     [(0.0, 0.0), (5.0, 0.0), (10.0, 0.0), (15.0, 0.0), (20.0, 0.0), (25.0, 0.0)],
     [(25.0, 0.0), (25.0, 5.0), (25.0, 10.0), (25.0, 15.0), (25.0, 20.0), (25.0, 25.0)],
@@ -53,22 +53,14 @@ points = [
 boundary_nodes, points = convert_boundary_points_to_indices(curves; existing_points=points)
 ````
 
-````
[…]
