+++ docs/src/tutorials/convex_hull.md
@@ -13,7 +13,7 @@ from the point set directly, using the [monotone chain algorithm](https://en.wik
 Let us first demonstrate how we compute convex hulls from a triangulation.
 This is only possible from triangulations without a constrained boundary
 
-````julia
+````@example convex_hull
 using DelaunayTriangulation
 using CairoMakie
 using StableRNGs
@@ -21,82 +21,36 @@ using StableRNGs
 
 We construct a triangulation of a point set.
 
-````julia
+````@example convex_hull
 rng = StableRNG(123)
 points = randn(rng, 2, 250)
 tri = triangulate(points; rng)
 ````
 
-````
-Delaunay Triangulation.
-   Number of vertices: 250
-   Number of triangles: 489
-   Number of edges: 738
-   Has boundary nodes: false
-   Has ghost triangles: true
-   Curve-bounded: false
-   Weighted: false
-   Constrained: false
-````
-
 To get the convex hull, just use [`get_convex_hull`](@ref):
 
-````julia
+````@example convex_hull
 get_convex_hull(tri)
-````
-
-````
-Convex hull.
-   Vertices:
-10-element Vector{Int64}:
- 151
-  96
-  43
- 159
-  25
- 147
-  65
- 206
- 199
- 151
 ````
 
 You can also obtain the vertices directly from
 [`get_convex_hull_vertices`](@ref):
 
-````julia
+````@example convex_hull
 get_convex_hull_vertices(tri)
-````
-
-````
-10-element Vector{Int64}:
- 151
-  96
-  43
- 159
-  25
- 147
-  65
- 206
- 199
- 151
 ````
 
 The vertices refer to indices of points in `points`, and they are
 given in counter-clockwise order. To obtain the convex hull directly
 from the points without constructing the triangulation, use [`convex_hull`](@ref):
 
-````julia
+````@example convex_hull
 ch = convex_hull(points)
[…]
