+++ docs/src/tutorials/constrained_multiply_connected.md
@@ -9,7 +9,7 @@ We now consider triangulating a domain which has not only a
 boundary, but also an interior boundary. To start, let us
 load the packages.
 
-````julia
+````@example constrained_multiply_connected
 using DelaunayTriangulation
 using CairoMakie
 using StableRNGs
@@ -26,7 +26,7 @@ the interior boundaries must be clockwise so that the orientations
 of the interiors relative to the boundaries are consistent. Note again
 that neighbouring segments must connect.
 
-````julia
+````@example constrained_multiply_connected
 curve_1 = [
     [ # first segment
         (0.0, 0.0), (4.0, 0.0), (8.0, 0.0), (12.0, 0.0), (12.0, 4.0),
@@ -65,6 +65,7 @@ points = [
     (5.6, 7.8), (5.6, 7.6), (5.6, 7.4), (6.2, 7.4), (6.0, 7.6),
     (7.0, 7.8), (7.0, 7.4)]
 boundary_nodes, points = convert_boundary_points_to_indices(curves; existing_points=points);
+nothing #hide
 ````
 
 Notice that `curve_1` and `curve_2` are split up into multiple segments. For `curve_3`
@@ -73,16 +74,12 @@ treating them as a single segment.
 
 Now let us triangulate.
 
-````julia
+````@example constrained_multiply_connected
 rng = StableRNG(123) # the triangulation is not unique due to cocircular points
 tri = triangulate(points; boundary_nodes, rng)
 fig, ax, sc = triplot(tri, show_constrained_edges=true, show_convex_hull=true)
 fig
 ````
-
-```@raw html
[…]
