+++ docs/src/tutorials/nearest.md
@@ -12,7 +12,7 @@ given a point `p` find the Voronoi tile `P` containing it. Here we give an examp
 of how we can use triangulations or tessellations to find the nearest neighbour
 in the point set to a given point. First, we load in the packages we need.
 
-````julia
+````@example nearest
 using DelaunayTriangulation
 using CairoMakie
 ````
@@ -20,7 +20,7 @@ using CairoMakie
 Now we define the tessellation we will use for this example. The white points
 shown are the points that we will query.
 
-````julia
+````@example nearest
 points = [
     (-3.0, 7.0), (2.0, 6.0), (0.0, 3.0),
     (0.0, 0.0), (-5.0, 5.0), (-3.0, 1.0),
@@ -34,46 +34,26 @@ scatter!(ax,[p,q],color=:white,strokecolor=:black,strokewidth=2,markersize=14)
 fig
 ````
 
-```@raw html
-<img width=600 height=450 style='object-fit: contain; height: auto;' src="data:image/png;base64, iVBORw0KGgoAAAANSUhEUgAABLAAAAOECAIAAAA+D1+tAAAABmJLR0QA/wD/AP+gvaeTAAAgAElEQVR4nOzdeXyb1Zk2/uvRLsuWvO9OnM2J7Tg7CUsgELZAEjuFpFAGKFs6naHLTIe+7fw689LSzmc6w0tbSikFWph22jIDbWcKbSBk3yAJCdntxEsWr5Jsy7ZsybK28/tDNISQxXYknUfS9f1LlvXIF2Bk3Trn3LcihAARERERERGlHo3sAERERERERCQHC0IiIiIiIqIUxYKQiIiIiIgoRbEgJCIiIiIiSlEsCImIiIiIiFIUC0IiIiIiIqIUxYKQiIiIiIgoRbEgJCIiIiIiSlEsCImIiIiIiFIUC0IiIiIiIqIUxYKQiIiIiIgoRbEgJCIiIiIiSlEsCImIiIiIiFIUC0IiIiIiIqIUxYKQiIiIiIgoRbEgJCIiIiIiSlEsCImIiIiIiFIUC0IiIiIiIqIUxYKQiIiIiIgoRbEgJCIiIiIiSlEsCImIiIiIiFKUTnaAi+ru7t65c2d9fb3b7Z4xY8bcuXPnzJkjOxQREREREVHyUGNBGAwGf/jDH373u98dHBw89/5Vq1b94Ac/mDRpkqxgREREREREyUQRQsjOcL4nnnjimWeeAZCdnX3NNddYrdZDhw7V19cDmDhx4sGDBzMzM2VnJCIiIiIiSniqKwh37tx5ww03CCFWrlz5q1/96mzt98tf
[…]
