+++ src/public.jl
@@ -1,264 +1,269 @@
-public Adjacent,
-Adjacent2Vertex,
-ConvexHull,
-Graph,
-InsertionEventHistory,
-Triangulation,
-VoronoiTessellation,
-ZeroWeight,
-add_weight!,
-all_ghost_vertices,
-check_args,
-compute_representative_points!,
-contains_boundary_edge,
-convert_boundary_points_to_indices,
-delete_ghost_vertices_from_graph!,
-dist,
-get_adjacent,
-get_adjacent2vertex,
-get_all_boundary_nodes,
-get_all_segments,
-get_area,
-get_boundary_curves,
-get_boundary_edge_map,
-get_boundary_nodes,
-get_convex_hull,
-get_convex_hull_vertices,
-get_curve_index,
-get_edges,
-get_ghost_vertex_map,
-get_ghost_vertex_range,
-get_ghost_vertex_ranges,
-get_graph,
-get_insertion_order,
-get_interior_segments,
-get_left_boundary_node,
-get_neighbours,
-get_point,
-get_points,
-get_representative_point_coordinates,
-get_representative_point_list,
-get_right_boundary_node,
-get_section_index,
-get_triangles,
-get_vertices,
-get_weight,
-get_weights,
-has_ghost_vertices,
-has_vertex,
-is_exterior_ghost_vertex,
-is_interior_ghost_vertex,
-iterated_neighbourhood,
-map_ghost_vertex,
-num_curves,
-num_neighbours,
-num_points,
-num_sections,
-refine!,
-retriangulate,
-triangulate,
-triangulate_convex,
-triangulate_rectangle,
-validate_triangulation,
-add_boundary_information!,
-add_ghost_triangles!,
-add_point!,
-add_segment!,
-add_triangle!,
-clear_empty_features!,
-complete_split_edge_and_legalise!,
-complete_split_triangle_and_legalise!,
-delete_ghost_triangles!,
-delete_holes!,
-delete_point!,
-delete_triangle!,
-flip_edge!,
-get_surrounding_polygon,
-legalise_edge!,
-lock_convex_hull!,
-split_edge!,
-split_triangle!,
-unlock_convex_hull!,
-centroidal_smooth,
-get_boundary_polygons,
-get_centroid,
-get_circumcenter_to_triangle,
-get_cocircular_circumcenters,
-get_generator,
-get_generators,
-get_polygon,
-get_polygon_coordinates,
-get_polygon_point,
-get_polygon_points,
-get_polygons,
-get_triangle_to_circumcenter,
-get_triangulation,
-get_unbounded_polygons,
-num_generators,
-num_polygon_vertices,
-num_polygons,
-polygon_bounds,
-polygon_features,
-toggle_inf_warn!,
-voronoi,
-convex_hull,
-convex_hull!,
-AbstractParametricCurve,
-BSpline,
-BezierCurve,
-CatmullRomSpline,
-CircularArc,
-EllipticalArc,
-LineSegment,
-angle_between,
-arc_length,
-curvature,
-differentiate,
-get_circle_intersection,
-get_closest_point,
-get_equidistant_split,
-get_equivariation_split,
-get_inverse,
-marked_total_variation,
-orientation_markers,
-point_position_relative_to_curve,
-thrice_differentiate,
-total_variation,
-twice_differentiate,
-each_boundary_edge,
-each_edge,
-each_generator,
-each_ghost_edge,
-each_ghost_triangle,
-each_ghost_vertex,
-each_point,
-each_point_index,
-each_polygon,
-each_polygon_index,
-each_polygon_vertex,
-each_segment,
-each_solid_edge,
-each_solid_triangle,
-each_solid_vertex,
[…]
