### 2024-07-16 #6937 @danielskatz
👋 @vissarion & @hugoledoux - I know you are both at your limits currently, but there is at least an appearance of a COI with @Kevin-Mattheus-Moerman and this submission, so I wonder if either of you would be able to take it, even if you can't start on it immediately?
