+++ src/validation.jl
@@ -35,7 +35,7 @@ end
 
 struct TriangleOrientationState <: AbstractTriangulationState
     flag::Bool
-    bad_triangle::NTuple{3,Int}
+    bad_triangle::NTuple{3, Int}
     triangle_orientation::Symbol
 end
 TriangleOrientationState(tri) = test_triangle_orientation(tri)
@@ -58,7 +58,7 @@ end
 
 struct DelaunayCriterionState <: AbstractTriangulationState
     flag::Bool
-    bad_triangle::NTuple{3,Int}
+    bad_triangle::NTuple{3, Int}
     bad_vertex::Int
 end
 DelaunayCriterionState(tri) = test_delaunay_criterion(tri)
@@ -78,7 +78,7 @@ function test_delaunay_criterion(tri)
         points = get_points(tri)
         triangle_tree = BoundaryRTree(points)
         segment_tree = BoundaryRTree(points)
-        failures = Tuple{triangle_type(tri),integer_type(tri)}[]
+        failures = Tuple{triangle_type(tri), integer_type(tri)}[]
         for T in each_solid_triangle(tri)
             i, j, k = triangle_vertices(T)
             p, q, r = get_point(tri, i, j, k)
@@ -106,7 +106,7 @@ function test_delaunay_criterion(tri)
         end
         for r in shuffle(collect(each_solid_vertex(tri)))
             !isempty(failures) && break
-            intersects = get_intersections(triangle_tree, r, cache_id=1) # can't use multithreading here
+            intersects = get_intersections(triangle_tree, r, cache_id = 1) # can't use multithreading here
             for box in intersects
                 i, j = get_edge(box)
                 k = get_adjacent(tri, i, j)
@@ -121,7 +121,7 @@ function test_delaunay_criterion(tri)
                     for (i, j) in triangle_edges(i, j, k)
                         contains_segment(tri, i, j) && continue # visibility is defined according to the relative interior of the simplex, which means that it's fine if a segment can see the vertex
                         if rand() < 1 / 2 # just testing both
-                            cert = test_visibility(tri, i, j, r, shift=0.01, attractor=c)
+                            cert = test_visibility(tri, i, j, r, shift = 0.01, attractor = c)
                         else
                             cert = test_visibility(tri, segment_tree, i, j, r, c)
                         end
@@ -142,7 +142,7 @@ function test_delaunay_criterion(tri)
         rethrow(e)
     end
 end
-function test_visibility(tri::Triangulation, segment_tree, i, j, k, centroid=nothing)
+function test_visibility(tri::Triangulation, segment_tree, i, j, k, centroid = nothing)
     if is_boundary_edge(tri, j, i)
         i, j = j, i
     end
@@ -178,7 +178,7 @@ function test_visibility(tri::Triangulation, segment_tree, i, j, k, centroid=not
         ax, ay = getxy(a)
         xmin, ymin, xmax, ymax = min(mx, ax), min(my, ay), max(mx, ax), max(my, ay)
         bbox = BoundingBox(xmin, xmax, ymin, ymax)
-        intersections = get_intersections(segment_tree, bbox; cache_id=2)
+        intersections = get_intersections(segment_tree, bbox; cache_id = 2)
         for box in intersections
[…]
