+++ docs/make.jl
@@ -4,8 +4,10 @@ using Literate
 using Test
 using Dates
 
-DocMeta.setdocmeta!(DelaunayTriangulation, :DocTestSetup, :(using DelaunayTriangulation, Test);
-    recursive=true)
+DocMeta.setdocmeta!(
+    DelaunayTriangulation, :DocTestSetup, :(using DelaunayTriangulation, Test);
+    recursive = true,
+)
 
 const IS_LIVESERVER = false && get(ENV, "LIVESERVER_ACTIVE", "false") == "true"
 if IS_LIVESERVER
@@ -41,7 +43,7 @@ function add_just_the_code_section(dir, file)
     file_name, file_ext = splitext(file)
     file_path = joinpath(dir, file)
     new_file_path = joinpath(session_tmp, file_name * "_just_the_code" * file_ext)
-    cp(file_path, new_file_path, force=true)
+    cp(file_path, new_file_path, force = true)
     folder = splitpath(dir)[end] # literate_tutorials or literate_applications
     open(new_file_path, "a") do io
         write(io, "\n")
@@ -74,7 +76,7 @@ for folder in ("tutorials", "applications")
         #    end
         #end
         new_file_path = add_just_the_code_section(dir, file)
-        script = Literate.script(file_path, session_tmp, name=splitext(file)[1] * "_just_the_code_cleaned")
+        script = Literate.script(file_path, session_tmp, name = splitext(file)[1] * "_just_the_code_cleaned")
         code = strip(read(script, String))
         @info "[$(ct())] Processing $file: Converting markdown script"
         line_ending_symbol = occursin(code, "\r\n") ? "\r\n" : "\n"
@@ -86,12 +88,12 @@ for folder in ("tutorials", "applications")
         Literate.markdown(
             new_file_path,
             outputdir;
-            documenter=true,
-            postprocess=editurl_update ∘ post_strip,
-            credit=true,
-            name=splitext(file)[1],
-            execute=!IS_LIVESERVER,
-            flavor=Literate.DocumenterFlavor(),
+            documenter = true,
+            postprocess = editurl_update ∘ post_strip,
+            credit = true,
+            name = splitext(file)[1],
+            execute = !IS_LIVESERVER,
+            flavor = Literate.DocumenterFlavor(),
         )
     end
 end
@@ -127,7 +129,7 @@ const _PAGES = [
         "Voronoi Tessellations" => "tutorials/voronoi.md",
         "Clipped Voronoi Tessellations" => [
             "Clipping to the Convex Hull" => "tutorials/clipped.md",
-            "Clipping to a Rectangle" => "tutorials/clipped_rectangle.md"
+            "Clipping to a Rectangle" => "tutorials/clipped_rectangle.md",
         ],
         "Centroidal Voronoi Tessellations" => "tutorials/centroidal.md",
         "Point Location" => "tutorials/point_location.md",
@@ -135,7 +137,7 @@ const _PAGES = [
         "Convex Hulls" => "tutorials/convex_hull.md",
         "Pole of Inaccessibility" => "tutorials/pole_of_inaccessibility.md",
         "Point-in-Polygon Testing" => "tutorials/point_in_polygon.md",
-        "Using Custom Structs for Primitives and Boundaries" => "tutorials/custom_primitive.md"
+        "Using Custom Structs for Primitives and Boundaries" => "tutorials/custom_primitive.md",
     ],
     "Manual" => [
         "Overview" => "manual/overview.md",
@@ -168,7 +170,7 @@ const _PAGES = [
         "Data Structures" => "extended/data_structures.md",
[…]
