+++ docs/src/literate_tutorials/pole_of_inaccessibility.jl
@@ -49,7 +49,7 @@ points = [
     5.33697 1.88019
 ]'
 outer_boundary = [
-    # split into segments for demonstration purposes
+    ## split into segments for demonstration purposes
     [1, 4, 3, 2],
     [2, 9, 10, 11, 8, 7, 12],
     [12, 6, 13, 5, 14, 15, 16, 17, 16],
