+++ docs/src/tutorials/operations_ghost_triangles.md
@@ -11,22 +11,38 @@ As we discussed in the [vertex insertion/deletion example](operations_vertex_ins
 ghost triangles are needed when we are making updates outside of the boundary
 of the current triangulation.
 
-````@example operations_ghost_triangles
+````julia
 using DelaunayTriangulation
 using CairoMakie
 ````
 
 Let us take an example triangulation.
 
-````@example operations_ghost_triangles
+````julia
 points = [(-1.0, -1.0), (1.0, -1.0), (0.0, 1.0)]
 tri = triangulate(points)
 ````
 
-````@example operations_ghost_triangles
+````
+Delaunay Triangulation.
+   Number of vertices: 3
+   Number of triangles: 1
+   Number of edges: 3
+   Has boundary nodes: false
+   Has ghost triangles: true
+   Curve-bounded: false
+   Weighted: false
+   Constrained: false
+````
+
+````julia
 fig, ax, sc = triplot(tri, show_ghost_edges = true)
 fig
 ````
+
+```@raw html
[…]
