+++ docs/src/tutorials/operations_split_triangle.md
@@ -12,7 +12,7 @@ that triangle into three new triangles. This is called *triangle splitting*.
 
 Let us give an example.
 
-````@example operations_split_triangle
+````julia
 using DelaunayTriangulation
 using CairoMakie
 
@@ -20,9 +20,13 @@ points = [(0.0, 0.0), (1.0, 0.0), (0.0, 1.0)]
 p = (0.2, 0.5)
 tri = triangulate(points)
 fig, ax, sc = triplot(tri)
-scatter!(ax, [p], markersize=14)
+scatter!(ax, [p], markersize = 14)
 fig
 ````
+
+```@raw html
+<img width=600 height=450 style='object-fit: contain; height: auto;' src="data:image/png;base64, iVBORw0KGgoAAAANSUhEUgAABLAAAAOECAIAAAA+D1+tAAAABmJLR0QA/wD/AP+gvaeTAAAgAElEQVR4nOzdeVRUd4L//VtSCCiiGBW3oInGJW5V7AKiqKCgMW64e39PP8+M6Z6Zp25l+szT09O/c+rWbD0958xM3ZrTPW3PpLurS03c447iiqAiSxUhmhiXGBQVd0QFIVU8f9C/JJ0QFwS+tbxff9F1q8rPyUkH39xLXV1LS4sEAAAAAAg+3UQPAAAAAACIQRACAAAAQJAiCAEAAAAgSBGEAAAAABCkCEIAAAAACFIEIQAAAAAEKYIQAAAAAIIUQQgAAAAAQYogBAAAAIAgRRACAAAAQJAiCAEAAAAgSBGEAAAAABCkCEIAAAAACFIEIQAAAAAEKYIQAAAAAIIUQQgAAAAAQYogBAAAAIAgRRACAAAAQJAiCAEAAAAgSBGEAAAAABCkCEIAAAAACFIEIQAAAAAEKYIQAAAAAIIUQQgAAAAAQYogBAAAAIAgRRACAAAAQJAiCAEAAAAgSBGEAAAAABCkCEIAAAAACFIEIQAAAAAEKb3oAT5Bp9OJngAAAAAAbWhpaem8N+cMIQAAAAAEKc4Qfq1Ty/tF1dTUeDyeIUOGhISEiN4CAHgB1dXVkiTFxsaKHgIAeAEej6empiYkJGTIkCGit3ytC65k5AwhAAAAAAQpghAAAAAAghRBCAAAAABBiiAEAAAAgCBFEAIAAABAkCIIAQAAACBIEYQAAAAAEKQIQgAAAAAIUgQhAAAAAAQpghAAAAAAghRBCAAAAABBiiAEAAAAgCBFEAIAAABAkCIIAQAAACBIEYQAAAAAEKQIQgAAAAAIUgQhAAAAAAQpghAAAAAAghRBCAAAAABBiiAEAAAAgCBFEAIAAABAkCIIAQAAACBIEYQAAAAAEKQIQgAAAAAIUl0ahPfu3Rs3btyPfvSj9r385MmT77zzTnp6elxc3Pz58997770nT5507EIAAAAACB76rvzDtm7devbs2ddee60dr
[…]
