+++ docs/src/tutorials/operations_convex_hull_locking.md
@@ -15,7 +15,7 @@ inside `refine!` when providing an unconstrained triangulation
 for mesh refinement. Let us give an example of how this can be done,
 in case you want to do this for your own application.
 
-````@example operations_convex_hull_locking
+````julia
 using DelaunayTriangulation
 using CairoMakie
 
@@ -24,12 +24,31 @@ tri = triangulate(points)
 get_boundary_nodes(tri)
 ````
 
+````
+Int64[]
+````
+
 As you can see, the boundary nodes field is empty.
 We can lock the convex hull using [`lock_convex_hull!`](@ref):
 
-````@example operations_convex_hull_locking
+````julia
 lock_convex_hull!(tri)
 get_boundary_nodes(tri)
+````
+
+````
+11-element Vector{Int64}:
+  4
+  1
+ 37
+  6
+ 27
+ 24
+ 31
+  5
+ 14
+  8
+  4
 ````
 
 Now the boundary nodes field is not empty. Note that if you try
@@ -37,9 +56,13 @@ and lock the convex hull again, you will get an error because
 `DelaunayTriangulation.has_boundary_nodes(tri)` is now true.
 To now unlock the convex hull, we use [`unlock_convex_hull!`](@ref):
 
-````@example operations_convex_hull_locking
+````julia
 unlock_convex_hull!(tri)
 get_boundary_nodes(tri)
+````
+
+````
+Int64[]
 ````
 
 This function will error if it detects that the existing boundary
@@ -48,6 +71,7 @@ isn't actually equal to the convex hull.
 Note that this locking/unlocking doesn't actually change anything about the triangulation,
 it just adds information into `tri` to treat it as if you had provided
 the convex hull as a constrained boundary to start with.
+
 ## Just the code
 An uncommented version of this example is given below.
 You can view the source code for this file [here](https://github.com/JuliaGeometry/DelaunayTriangulation.jl/tree/main/docs/src/literate_tutorials/operations_convex_hull_locking.jl).
