+++ docs/src/tutorials/voronoi.md
@@ -9,7 +9,7 @@ tessellations and work with them. Voronoi tessellations are built from
 a dual Delaunay triangulation using [`voronoi`](@ref). To start, let us
 load in the packages.
 
-````@example voronoi
+````julia
 using DelaunayTriangulation
 using CairoMakie
 using StableRNGs
@@ -18,25 +18,36 @@ using StableRNGs
 We build the tessellation by constructing the triangulation,
 and then passing that triangulation into `voronoi`.
 
-````@example voronoi
+````julia
 points = [
     (-3.0, 7.0), (1.0, 6.0), (-1.0, 3.0),
     (-2.0, 4.0), (3.0, -2.0), (5.0, 5.0),
-    (-4.0, -3.0), (3.0, 8.0)
+    (-4.0, -3.0), (3.0, 8.0),
 ]
 rng = StableRNG(123)
 tri = triangulate(points; rng)
 vorn = voronoi(tri)
 ````
 
+````
+Voronoi Tessellation.
+    Number of generators: 8
+    Number of polygon vertices: 9
+    Number of polygons: 8
+````
+
 To visualise the tessellation, you can use `voronoiplot`. Here,
 we also compare the tessellation with its dual triangulation.
 
-````@example voronoi
-fig, ax, sc = voronoiplot(vorn, markersize=13, colormap=:matter, strokecolor=:white, strokewidth=5)
+````julia
+fig, ax, sc = voronoiplot(vorn, markersize = 13, colormap = :matter, strokecolor = :white, strokewidth = 5)
 triplot!(ax, tri)
 fig
 ````
+
+```@raw html
[…]
