### 2024-07-21 df20ca8
Update docs/src/literate_tutorials/operations_vertex_insertion_deletion.jl
