### 2024-07-21 cb9de09
Update docs/src/literate_tutorials/operations_vertex_insertion_deletion.jl
