### 2024-07-21 3dfd046
Fix tests on prerelease
