+++ docs/Project.toml
@@ -2,7 +2,6 @@
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-DelaunayTriangulation = "927a84f5-c5f4-47a5-9785-b46e178433df"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 ElasticArrays = "fdbdab4c-e67f-52f5-8c3f-e7b388dad3d4"
@@ -17,3 +16,22 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[compat]
+BenchmarkTools = "1.5"
+CairoMakie = "0.12"
+Documenter = "1.5.0"
+DocumenterTools = "0.1"
+ElasticArrays = "1.2"
+Literate = "2.19"
+LiveServer = "1.3"
+Preferences = "1.4"
+ReferenceTests = "0.10"
+SimpleGraphs = "0.8"
+StableRNGs = "1.0"
+StatsBase = "0.34"
+Dates = "1.11"
+InteractiveUtils = "1.11"
+LinearAlgebra = "1.11"
+SparseArrays = "1.11"
+Test = "1.11"
