+++ test/Project.toml
@@ -35,35 +35,3 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 Aqua = "0.8.7"
-BenchmarkTools = "1.5"
-CairoMakie = "0.12"
-ColorSchemes = "3.26"
-DataStructures = "0.18"
-DelimitedFiles = "1.9"
-Documenter = "1.5"
-DocumenterTools = "0.1"
-ElasticArrays = "1.2"
-ExactPredicates = "2.2"
-FastGaussQuadrature = "1.0"
-ForwardDiff = "0.10"
-GeometryBasics = "0.4"
-OrderedCollections = "1.6"
-Preferences = "1.4"
-ReferenceTests = "0.10"
-SafeTestsets = "0.1"
-Setfield = "1.1"
-SimpleGraphs = "0.8"
-SpatialIndexing = "0.1"
-StableRNGs = "1.0"
-StaticArrays = "1.9"
-StaticArraysCore = "1.4"
-StatsBase = "0.34"
-StructEquality = "2.1"
-Dates = "1.11"
-Downloads = "1.6"
-InteractiveUtils = "1.11"
-LinearAlgebra = "1.11"
-Pkg = "1.11"
-Random = "1.11"
-SparseArrays = "1.11"
-Test = "1.11"
