### 2024-07-21 cb79081
Update src/algorithms/triangulation/check_args.jl
Co-authored-by: Daniel VandenHeuvel <95613936+DanielVandH@users.noreply.github.com>
