### 2024-07-21 2bcadd8
Update NEWS.md
