### 2024-07-21 bef0192
Add `public`, `const`, and test on pre-release (#140)
