+++ NEWS.md
@@ -14,6 +14,7 @@
 - The `.md` files for tutorials and applications in the docs have been properly updated to match their literate counterparts. See [#140](https://github.com/JuliaGeometry/DelaunayTriangulation.jl/pull/140).
 - We now use a workflow to enforce changes to `NEWS.md` for any PRs. See [#140](https://github.com/JuliaGeometry/DelaunayTriangulation.jl/pull/140).
 - Improved the error message for an incorrect orientation. See [#144](https://github.com/JuliaGeometry/DelaunayTriangulation.jl/pull/144).
+- Added a CONTRIBUTING.md file and issue/PR templates. See [#160](https://github.com/JuliaGeometry/DelaunayTriangulation.jl/pull/160).
 
 ## v1.0.5 
 
