+++ .github/ISSUE_TEMPLATE/documentation-issues.md
@@ -1,0 +1,16 @@
+---
+name: Documentation issues
+about: Report a problem with DelaunayTriangulation's documentation
+title: "[Documentation]: "
+labels: ''
+assignees: ''
+
+---
+
+**Describe the issue**
+
+Describe the problem with the documentation. Please include a link to the relevant part of the documentation, including any code snippets if applicable.
+
+**Suggestion**
+
+Suggest how the issue could be improved.
