+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,40 @@
+---
+name: Bug report
+about: Report a bug with DelaunayTriangulation.jl
+title: "[BUG]: "
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+
+Describe the bug here.
+
+**Reproducer**
+
+Give a minimal working example of your bug here, e.g. something like
+
+```julia
+setup()
+function()
+< error message, or incorrect result >
+```
+
+Describe what you expected to obtain.
+
+**Environment information**
+
+Post the result of doing
+
+```julia
+using InteractiveUtils
+versioninfo()
+```
+
+and
+
+```julia
+using Pkg
+Pkg.status()
+```
