### 2024-07-22 cbf26a5
Fix convert_bouindary_points_to_indices for vector of vector coordinates
