### 2024-07-22 4ad705e
Fix convert_boundary_points_to_indices for vector of vector coordinates (#161)
