+++ docs/src/api/primitives.md
@@ -10,7 +10,9 @@ CurrentModule = DelaunayTriangulation
 getx
 gety
 getxy
+is_point2 
 getpoint
+get_point
 push_point!
 pop_point!
 set_point!
