### 2024-07-24 ba71e01
Remove allocation in add_vertex! (why is it even there..)
