+++ src/data_structures/triangulation/adjacent2vertex.jl
@@ -13,14 +13,14 @@ The map taking `w` to the set of all `(u, v)` such that `(u, v, w)` is a positiv
     Adjacent2Vertex(adj2v::Dict{IntegerType, EdgesType})
 """
 struct Adjacent2Vertex{I,Es}
-  adjacent2vertex::Dict{I,Es}
-  Adjacent2Vertex(adj2v::Dict{I,Es}) where {I,Es} = new{I,Es}(adj2v)
+    adjacent2vertex::Dict{I,Es}
+    Adjacent2Vertex(adj2v::Dict{I,Es}) where {I,Es} = new{I,Es}(adj2v)
 end
 Adjacent2Vertex{I,Es}() where {I,Es} = Adjacent2Vertex(Dict{I,Es}())
 Base.:(==)(adj2v::Adjacent2Vertex, adj2v2::Adjacent2Vertex) = get_adjacent2vertex(adj2v) == get_adjacent2vertex(adj2v2)
 function Base.show(io::IO, m::MIME"text/plain", adj2v::Adjacent2Vertex{I,Es}) where {I,Es}
-  println(io, "Adjacent2Vertex{", I, ", ", Es, "} with map:")
-  show(io, m, get_adjacent2vertex(adj2v))
+    println(io, "Adjacent2Vertex{", I, ", ", Es, "} with map:")
+    show(io, m, get_adjacent2vertex(adj2v))
 end
 Base.sizehint!(adj2v::Adjacent2Vertex, n) = Base.sizehint!(get_adjacent2vertex(adj2v), n)
 
@@ -86,8 +86,8 @@ Set{Tuple{Int64, Int64}} with 3 elements:
 ```
 """
 function get_adjacent2vertex(adj2v::Adjacent2Vertex, w)
-  dict = get_adjacent2vertex(adj2v)
-  return dict[w]
+    dict = get_adjacent2vertex(adj2v)
+    return dict[w]
 end
 
 """
@@ -123,15 +123,15 @@ Dict{Int64, Set{Tuple{Int64, Int64}}} with 2 entries:
 ```
 """
 function add_adjacent2vertex!(adj2v::Adjacent2Vertex{I,Es}, w, uv) where {I,Es}
-  dict = get_adjacent2vertex(adj2v)
-  existing_edges = get!(Es, dict, w)
-  add_edge!(existing_edges, uv)
-  return adj2v
+    dict = get_adjacent2vertex(adj2v)
+    existing_edges = get!(Es, dict, w)
+    add_edge!(existing_edges, uv)
+    return adj2v
 end
 function add_adjacent2vertex!(adj2v::Adjacent2Vertex{I,Es}, w, u, v) where {I,Es}
-  E = edge_type(Es)
-  uv = construct_edge(E, u, v)
-  return add_adjacent2vertex!(adj2v, w, uv)
+    E = edge_type(Es)
+    uv = construct_edge(E, u, v)
+    return add_adjacent2vertex!(adj2v, w, uv)
 end
 
 """
@@ -165,14 +165,14 @@ Dict{Int64, Set{Tuple{Int64, Int64}}} with 2 entries:
 ```
 """
 function delete_adjacent2vertex!(adj2v::Adjacent2Vertex, w, uv)
-  existing_edges = get_adjacent2vertex(adj2v, w)
-  delete_edge!(existing_edges, uv)
-  return adj2v
+    existing_edges = get_adjacent2vertex(adj2v, w)
+    delete_edge!(existing_edges, uv)
+    return adj2v
 end
 function delete_adjacent2vertex!(adj2v::Adjacent2Vertex{I,Es}, w, u, v) where {I,Es}
-  E = edge_type(Es)
[…]
