### 2024-07-24 db8acc9
Move out the config checks at the start of triangulate into their own function (#162)
