+++ src/algorithms/triangulation/main.jl
@@ -128,10 +128,7 @@ function triangulate(points::P;
     polygonise_n=4096,
     coarse_n=0,
     enrich=false) where {P,I,E,V,Es,Ts,M,H}
-    number_type(points) == Float64 || @warn "Using non-Float64 coordinates may cause issues. If you run into problems, consider using Float64 coordinates." maxlog=1
-    is_weighted(weights) && throw(ArgumentError("Weighted triangulations are not yet fully implemented."))
-    is_constrained = !(isnothing(segments) || isempty(segments)) || !(isnothing(boundary_nodes) || !has_boundary_nodes(boundary_nodes))
-    is_weighted(weights) && is_constrained && throw(ArgumentError("You cannot compute a constrained triangulation with weighted points."))
+    check_config(points, weights, segments, boundary_nodes)
     if enrich || (isempty(boundary_curves) && is_curve_bounded(boundary_nodes)) # If boundary_curves is not empty, then we are coming from triangulate_curve_bounded
         return triangulate_curve_bounded(points; segments, boundary_nodes, weights, IntegerType, EdgeType, TriangleType, EdgesType, TrianglesType, randomise, delete_ghosts, delete_empty_features, try_last_inserted_point, skip_points, num_sample_rule, rng, insertion_order, recompute_representative_points, delete_holes, check_arguments, polygonise_n, coarse_n)
     end
@@ -142,6 +139,14 @@ function triangulate(points::P;
     tri = Triangulation(points; IntegerType, EdgeType, TriangleType, EdgesType, TrianglesType, weights, boundary_curves, boundary_enricher, build_cache = Val(true))
     return _triangulate!(tri, segments, boundary_nodes, randomise, try_last_inserted_point, skip_points, num_sample_rule, rng, insertion_order,
         recompute_representative_points, delete_holes, full_polygon_hierarchy, delete_ghosts, delete_empty_features)
+end
+
+function check_config(points, weights, segments, boundary_nodes)
+    number_type(points) == Float64 || @warn "Using non-Float64 coordinates may cause issues. If you run into problems, consider using Float64 coordinates." maxlog=1
+    is_weighted(weights) && throw(ArgumentError("Weighted triangulations are not yet fully implemented."))
+    is_constrained = !(isnothing(segments) || isempty(segments)) || !(isnothing(boundary_nodes) || !has_boundary_nodes(boundary_nodes))
+    is_weighted(weights) && is_constrained && throw(ArgumentError("You cannot compute a constrained triangulation with weighted points."))
+    return nothing
 end
 
 function _triangulate!(tri::Triangulation, segments, boundary_nodes, randomise, try_last_inserted_point, skip_points, num_sample_rule, rng, insertion_order,
