### 2024-07-25 ee4fa67
Fix some weighted triangulation tests
