### 2024-07-26 5c7ea82
Use adaptive predicates
