+++ src/algorithms/convex_hull.jl
@@ -1,28 +1,36 @@
 """
-    convex_hull(points; IntegerType::Type{I}=Int) where {I} -> ConvexHull 
+    convex_hull(points; predicates::AbstractPredicateType=def_alg222(), IntegerType::Type{I}=Int) where {I} -> ConvexHull 
 
 Computes the convex hull of `points`. The monotone chain algorithm is used.
 
 # Arguments 
 - `points`: The set of points.
 
+# Keyword Arguments 
+- `IntegerType=Int`: The integer type to use for the vertices.
+- `predicates::AbstractPredicateType=def_alg222()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
+
 # Output
 - `ch`: The [`ConvexHull`](@ref). 
 """
-function convex_hull(points; IntegerType::Type{I}=Int) where {I}
+function convex_hull(points; predicates::AbstractPredicateType=def_alg222(), IntegerType::Type{I}=Int) where {I}
     ch = ConvexHull(points, I[])
     sizehint!(ch, num_points(points))
-    return convex_hull!(ch)
+    return convex_hull!(ch; predicates)
 end
 
 """
-    convex_hull!(ch::ConvexHull{P,I}) where {P,I}
+    convex_hull!(ch::ConvexHull{P,I}; predicates::AbstractPredicateType=def_alg222()) where {P,I}
 
 Using the points in `ch`, computes the convex hull in-place. 
 
+The `predicates` keywrod argument determines how predicates are computed, and should be 
+one of [`Exact`](@ref), [`Fast`](@ref), and [`Adaptive`](@ref) (the default).
+See the documentation for more information about these choices.
+
 See also [`convex_hull`](@ref).
 """
-function convex_hull!(ch::ConvexHull{P,I}) where {P,I}
+function convex_hull!(ch::ConvexHull{P,I}; predicates::AbstractPredicateType=def_alg222()) where {P,I}
     indices = get_vertices(ch)
     points = get_points(ch)
     empty!(indices)
@@ -37,7 +45,7 @@ function convex_hull!(ch::ConvexHull{P,I}) where {P,I}
         push!(indices, insertion_order[begin], insertion_order[begin+1], insertion_order[begin])
         return ch
     elseif n == 3
-        i, j, k = construct_positively_oriented_triangle(NTuple{3,I}, insertion_order[begin], insertion_order[begin+1], insertion_order[begin+2], points)
+        i, j, k = construct_positively_oriented_triangle(predicates,NTuple{3,I}, insertion_order[begin], insertion_order[begin+1], insertion_order[begin+2], points)
         push!(indices, i, j, k, i)
         return ch
     end
@@ -46,13 +54,13 @@ function convex_hull!(ch::ConvexHull{P,I}) where {P,I}
     sizehint!(lower, max(4, floor(I, cbrt(n))))
     sizehint!(upper, max(4, floor(I, cbrt(n))))
     for i in eachindex(insertion_order)
-        while length(upper) ≥ 2 && is_left(point_position_relative_to_line(get_point(points, upper[end-1]), get_point(points, upper[end]), get_point(points, insertion_order[i])))
+        while length(upper) ≥ 2 && is_left(point_position_relative_to_line(predicates,get_point(points, upper[end-1]), get_point(points, upper[end]), get_point(points, insertion_order[i])))
             pop!(upper)
         end
         push!(upper, insertion_order[i])
     end
     for i in lastindex(insertion_order):-1:firstindex(insertion_order)
[…]
