+++ src/algorithms/triangulation/basic_operations/lock_convex_hull.jl
@@ -1,8 +1,11 @@
 """
-    lock_convex_hull!(tri::Triangulation)
+    lock_convex_hull!(tri::Triangulation; rng=Random.default_rng(), predicates::AbstractPredicateType=def_alg222())
 
 Locks the convex hull of the unconstrained triangulation `tri` so that it is now treated as a constrained triangulation 
 with boundary given by its convex hull.
+
+The random number generator (used inside [`add_segment!`](@ref) can be provided with the `rng` keyword argument,
+and similarly for `predicates`.
 
 !!! warning 
 
@@ -10,7 +13,7 @@ with boundary given by its convex hull.
     then this edge will be deleted from `tri.interior_segments`; this will be undone from `unlock_convex_hull!`, 
     possibly splitting the segments in case they were split before unlocking.
 """
-function lock_convex_hull!(tri::Triangulation)
+function lock_convex_hull!(tri::Triangulation; rng, predicates::AbstractPredicateType=def_alg222())
     if has_boundary_nodes(tri)
         throw(ArgumentError("Cannot lock the convex hull of a triangulation with boundary nodes."))
     end
@@ -38,7 +41,7 @@ function lock_convex_hull!(tri::Triangulation)
         end
     end
     for e in keys(bnn_map)
-        add_segment!(tri, e)
+        add_segment!(tri, e; rng, predicates)
     end
     return tri
 end
