+++ src/predicates/predicate_definitions.jl
@@ -29,7 +29,7 @@ SOFTWARE.
 Computes `ExactPredicates.ext((ux, uy), (vx, vy))`, i.e. 
 returns `ux * vy - uy * vx`.
 """
-ext(ux, uy, vx, vy) = ux * vy - uy * vx
+@inline ext(ux, uy, vx, vy) = ux * vy - uy * vx
 
 """
     inp(ux, uy, vx, vy)
@@ -37,7 +37,7 @@ ext(ux, uy, vx, vy) = ux * vy - uy * vx
 Computes `ExactPredicates.inp((ux, uy), (vx, vy))`, i.e.
 returns `ux * vx + uy * vy`.
 """
-inp(ux, uy, vx, vy) = ux * vx + uy * vy
+@inline inp(ux, uy, vx, vy) = ux * vx + uy * vy
 
 """
     det(a, b, c, d)
@@ -45,7 +45,7 @@ inp(ux, uy, vx, vy) = ux * vx + uy * vy
 Computes `ExactPredicates.det(a, b, c, d)`,
 i.e. returns `a*d - b*c`.
 """
-det(a, b, c, d) = a * d - b * c
+@inline det(a, b, c, d) = a * d - b * c
 
 """
     det(a, b, c, d, e, f, g, h, i)
@@ -56,7 +56,7 @@ i.e. returns the determinant of
 \\det \\begin{bmatrix} a & b & c \\\\ d & e & f \\\\ g & h & i \\end{bmatrix}
 ```
 """
-det(a, b, c, d, e, f, g, h, i) = a * det(e, f, h, i) - d * det(b, c, h, i) + g * det(b, c, e, f)
+@inline det(a, b, c, d, e, f, g, h, i) = a * det(e, f, h, i) - d * det(b, c, h, i) + g * det(b, c, e, f)
 
 """
     opposite_signs(x, y) -> Bool
@@ -64,14 +64,14 @@ det(a, b, c, d, e, f, g, h, i) = a * det(e, f, h, i) - d * det(b, c, h, i) + g *
 From ExactPredicates.jl, returns `true` if `x` and `y` have opposite signs, and `false` otherwise.
 Assumes that `x` and `y` are in `[-1, 0, 1]`.
 """
-opposite_signs(x, y) = xor(x, y) == -2
+@inline opposite_signs(x, y) = xor(x, y) == -2
 
 """
     sgn(x) -> Int 
 
 Returns `Int(sign(x))`.
-""" 
-sgn(x) = Int(sign(x))
+"""
+@inline sgn(x) = Int(sign(x))
 
 """
     orient_predicate([method::AbstractPredicateType,] p, q, r) -> Integer 
@@ -86,14 +86,14 @@ The `method` argument determines how this result is computed, and should be
 one of [`Exact`](@ref), [`Fast`](@ref), and [`Adaptive`](@ref) (the default).
 See the documentation for more information about these choices.
 """
-function orient_predicate(method::AbstractPredicateType, p, q, r) 
+@inline function orient_predicate(method::AbstractPredicateType, p, q, r)
     return orient(method, getxy(p), getxy(q), getxy(r))
 end
-orient_predicate(p, q, r) = orient_predicate(def_alg222(), p, q, r)
+@inline orient_predicate(p, q, r) = orient_predicate(def_alg222(), p, q, r)
 
-orient(::Fast, p, q, r) = sgn(AP.orient2fast(p, q, r))
-orient(::Exact, p, q, r) = EP.orient(_getxy(p), _getxy(q), _getxy(r))
-orient(::Adaptive, p, q, r) = AP.orient2p(p, q, r)
[…]
