+++ src/algorithms/triangulation/basic_operations/legalise_edge.jl
@@ -1,5 +1,5 @@
 """
-    legalise_edge!(tri::Triangulation, i, j, r, store_event_history=Val(false), event_history=nothing; predicates::AbstractPredicateType=def_alg222())
+    legalise_edge!(tri::Triangulation, i, j, r, store_event_history=Val(false), event_history=nothing; predicates::AbstractPredicateType=Adaptive())
 
 Legalises the edge `(i, j)` and other neighbouring edges in `tri` if they are illegal, assuming the vertex `r` 
 was just added into a triangle that contains `(i, j)`. [`flip_edge!`](@ref) is used.
@@ -15,7 +15,7 @@ See also [`is_legal`](@ref).
 - `event_history=nothing`: The event history. Only updated if `store_event_history` is true, in which case it needs to be an [`InsertionEventHistory`](@ref) object.
 
 # Keyword Arguments 
-- `predicates::AbstractPredicateType=def_alg222()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
+- `predicates::AbstractPredicateType=Adaptive()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
 
 # Outputs 
 There is no output, as `tri` is updated in-place.
@@ -26,7 +26,7 @@ There is no output, as `tri` is updated in-place.
     To get around this, we only store in these fields the triangles necessary to allow [`undo_insertion!`](@ref) to work,
     so that at a triangle that might have appeared in both will only appear in one.
 """
-function legalise_edge!(tri::Triangulation, i, j, r, store_event_history=Val(false), event_history=nothing; predicates::AbstractPredicateType=def_alg222())
+function legalise_edge!(tri::Triangulation, i, j, r, store_event_history=Val(false), event_history=nothing; predicates::AbstractPredicateType=Adaptive())
     cert = is_legal(predicates, tri, i, j)
     if is_illegal(cert) 
         e = get_adjacent(tri, j, i)
