+++ src/algorithms/triangulation/constrained_triangulation.jl
@@ -400,8 +400,8 @@ function replace_ghost_vertex_information(tri::Triangulation, ghost_vertex_map, 
 end
 
 """
-    is_vertex_closer_than_neighbours([predicates::AbstractPredicateType=Adaptive(),] tri::Triangulation, u, v, jᵢ, jᵢ₋₁, jᵢ₊₁) -> Bool
-    is_vertex_closer_than_neighbours([predicates::AbstractPredicateType=Adaptive(),] tri::Triangulation, list::ShuffledPolygonLinkedList, u, v, j) -> Bool
+    is_vertex_closer_than_neighbours([predicates::AbstractPredicateKernel=def_alg222(),] tri::Triangulation, u, v, jᵢ, jᵢ₋₁, jᵢ₊₁) -> Bool
+    is_vertex_closer_than_neighbours([predicates::AbstractPredicateKernel=def_alg222(),] tri::Triangulation, list::ShuffledPolygonLinkedList, u, v, j) -> Bool
 
 Tests if the vertex `jᵢ` is closer to the line `(u, v)` than its neighbours `jᵢ₋₁` and `jᵢ₊₁`, assuming all these 
 vertices are to the left of the line.
@@ -409,7 +409,7 @@ vertices are to the left of the line.
 See also [`point_closest_to_line`](@ref).
 
 # Arguments 
-- `predicates::AbstractPredicateType=Adaptive()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
+- `predicates::AbstractPredicateKernel=def_alg222()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
 - `tri::Triangulation`: The [`Triangulation`](@ref).
 - `u`, `v`: The vertices of the line. 
 - `jᵢ`, `jᵢ₋₁`, `jᵢ₊₁`: The vertices to compare. 
@@ -419,19 +419,19 @@ The second method extracts these latter two vertices using the doubly-linked `li
 # Outputs 
 - `flag`: Whether `jᵢ` is closer to the line than `jᵢ₋₁` and `jᵢ₊₁`.
 """
-function is_vertex_closer_than_neighbours(predicates::AbstractPredicateType, tri::Triangulation, u, v, jᵢ, jᵢ₋₁, jᵢ₊₁)
+function is_vertex_closer_than_neighbours(predicates::AbstractPredicateKernel, tri::Triangulation, u, v, jᵢ, jᵢ₋₁, jᵢ₊₁)
     prev_comp = point_closest_to_line(predicates, tri, u, v, jᵢ, jᵢ₋₁)
     is_further(prev_comp) && return false # If we just do is_closer, then we can run into an infinite loop when all neighbours are equidistant 
     next_comp = point_closest_to_line(predicates, tri, u, v, jᵢ, jᵢ₊₁)
     is_further(next_comp) && return false
     return true
 end
-function is_vertex_closer_than_neighbours(predicates::AbstractPredicateType, tri::Triangulation, list::ShuffledPolygonLinkedList, u, v, j)
+function is_vertex_closer_than_neighbours(predicates::AbstractPredicateKernel, tri::Triangulation, list::ShuffledPolygonLinkedList, u, v, j)
     jᵢ, jᵢ₋₁, jᵢ₊₁ = get_triplet(list, j)
[…]
