+++ src/data_structures/mesh_refinement/boundary_enricher.jl
@@ -775,21 +775,27 @@ end
 end
 
 """
-    point_position_relative_to_curve(enricher::BoundaryEnricher, curve_index, p) -> Certificate
+    point_position_relative_to_curve([kernel::AbstractPredicateKernel=def_alg222(),] enricher::BoundaryEnricher, curve_index, p) -> Certificate
 
 Returns a [`Certificate`](@ref) which is 
 
 - `Left`: If `p` is to the left of the `curve_index`th curve.
 - `Right`: If `p` is to the right of the `curve_index`th curve.
 - `On`: If `p` is on the `curve_index`th curve.
+
+The `kernel` argument determines how this result is computed, and should be 
+one of [`Exact`](@ref), [`Fast`](@ref), and [`Adaptive`](@ref) (the default).
+See the documentation for more information about these choices.
 """
-@inline function point_position_relative_to_curve(enricher::BoundaryEnricher, curve_index, p)
+@inline function point_position_relative_to_curve(kernel::AbstractPredicateKernel, enricher::BoundaryEnricher, curve_index, p)
     boundary_curves = get_boundary_curves(enricher)
-    return point_position_relative_to_curve(boundary_curves, curve_index, p)
+    return point_position_relative_to_curve(kernel, boundary_curves, curve_index, p)
 end
-@inline function point_position_relative_to_curve(boundary_curves::C, curve_index, p) where {C<:Tuple}
-    return eval_fnc_at_het_tuple_element_with_arg(point_position_relative_to_curve, boundary_curves, (p,), curve_index)
+@inline point_position_relative_to_curve(enricher::BoundaryEnricher, curve_index, p) = point_position_relative_to_curve(def_alg222(), enricher, curve_index, p)
+@inline function point_position_relative_to_curve(kernel::AbstractPredicateKernel, boundary_curves::C, curve_index, p) where {C<:Tuple}
+    return eval_fnc_at_het_tuple_element_with_arg_and_prearg(point_position_relative_to_curve, boundary_curves, kernel, (p,), curve_index)
 end
+@inline point_position_relative_to_curve(boundary_curves::C, curve_index, p) where {C<:Tuple} = point_position_relative_to_curve(def_alg222(),boundary_curves,curve_index,p)
 
 """
     angle_between(enricher::BoundaryEnricher, curve_index1, curve_index2) -> Float64 

[…]
