+++ src/algorithms/triangulation/triangulate_curve_bounded.jl
@@ -2,7 +2,7 @@
     triangulate_curve_bounded(points::P;
     segments=nothing,
     boundary_nodes=nothing,
-    predicates::AbstractPredicateKernel=def_alg222(),
+    predicates::AbstractPredicateKernel=Adaptive(),
     IntegerType::Type{I}=Int,
     polygonise_n=4096,
     coarse_n=0,
@@ -36,7 +36,7 @@ See also [`BoundaryEnricher`](@ref) and [`enrich_boundary!`](@ref).
 function triangulate_curve_bounded(points::P;
     segments=nothing,
     boundary_nodes=nothing,
-    predicates::AbstractPredicateKernel=def_alg222(),
+    predicates::AbstractPredicateKernel=Adaptive(),
     IntegerType::Type{I}=Int,
     polygonise_n=4096,
     coarse_n=0,
@@ -60,7 +60,7 @@ function triangulate_curve_bounded(points::P;
         kwargs...)
 end
 function _triangulate_curve_bounded(points::P, enricher;
-    predicates::AbstractPredicateKernel=def_alg222(),
+    predicates::AbstractPredicateKernel=Adaptive(),
     IntegerType::Type{I}=Int,
     check_arguments=true,
     delete_ghosts=false,
@@ -171,7 +171,7 @@ function _coarse_discretisation_contiguous!(points, boundary_nodes, boundary_cur
 end
 
 """
-    enrich_boundary!(enricher::BoundaryEnricher; predicates::AbstractPredicateKernel=def_alg222())
+    enrich_boundary!(enricher::BoundaryEnricher; predicates::AbstractPredicateKernel=Adaptive())
 
 Enriches the initial boundary defined inside `enricher`, implementing the algorithm of Gosselin and Ollivier-Gooch (2007).
 At the termination of the algorithm, all edges will contain no other points inside their 
@@ -181,7 +181,7 @@ The `predicates` argument determines how predicates are computed, and should be
 one of [`Exact`](@ref), [`Fast`](@ref), and [`Adaptive`](@ref) (the default).
 See the documentation for more information about these choices.
 """
-function enrich_boundary!(enricher::BoundaryEnricher; predicates::AbstractPredicateKernel=def_alg222())
+function enrich_boundary!(enricher::BoundaryEnricher; predicates::AbstractPredicateKernel=Adaptive())
     queue = get_queue(enricher)
     points = get_points(enricher)
     enqueue_all!(queue, each_point_index(points))
@@ -190,7 +190,7 @@ function enrich_boundary!(enricher::BoundaryEnricher; predicates::AbstractPredic
     end
     return enricher
 end
-function _enrich_boundary_itr!(enricher::BoundaryEnricher, predicates::AbstractPredicateKernel=def_alg222())
+function _enrich_boundary_itr!(enricher::BoundaryEnricher, predicates::AbstractPredicateKernel=Adaptive())
     queue = get_queue(enricher)
     points = get_points(enricher)
     spatial_tree = get_spatial_tree(enricher)
@@ -217,7 +217,7 @@ function _enrich_boundary_itr!(enricher::BoundaryEnricher, predicates::AbstractP
 end
 
 """
-    split_subcurve!(enricher::BoundaryEnricher, i, j, predicates::AbstractPredicateKernel=def_alg222()) -> Bool
+    split_subcurve!(enricher::BoundaryEnricher, i, j, predicates::AbstractPredicateKernel=Adaptive()) -> Bool
 
 Splits the curve associated with the edge `(i, j)` into two subcurves by inserting a point `r` between `(i, j)` such that the 
 total variation of the subcurve is equal on `(i, r)` and `(r, j)`. The returned value is a `flag` that is `true` 
[…]
