+++ src/algorithms/triangulation/unconstrained_triangulation.jl
@@ -37,13 +37,13 @@ Gets the initial triangle for the Bowyer-Watson algorithm.
 # Arguments 
 - `tri`: The triangulation.
 - `insertion_order`: The insertion order of the points. See [`get_insertion_order`](@ref).
-- `predicates::AbstractPredicateKernel=def_alg222()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
+- `predicates::AbstractPredicateKernel=Adaptive()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
 - `itr=0`: To avoid issues with degenerate triangles and infinite loops, this counts the number of times `insertion_order` had to be shifted using `circshift!` to find an initial non-degenerate triangle.
 
 # Output
 - `initial_triangle`: The initial triangle.
 """
-function get_initial_triangle(tri::Triangulation, insertion_order, predicates::AbstractPredicateKernel=def_alg222(), itr=0)
+function get_initial_triangle(tri::Triangulation, insertion_order, predicates::AbstractPredicateKernel=Adaptive(), itr=0)
     i, j, k = @view insertion_order[1:3] # insertion_order got converted into a Vector, so indexing is safe 
     initial_triangle = construct_positively_oriented_triangle(tri, i, j, k, predicates)
     i, j, k = triangle_vertices(initial_triangle)
@@ -62,19 +62,19 @@ function get_initial_triangle(tri::Triangulation, insertion_order, predicates::A
 end
 
 """
-    initialise_bowyer_watson!(tri::Triangulation, insertion_order, predicates::AbstractPredicateKernel=def_alg222()) -> Triangulation
+    initialise_bowyer_watson!(tri::Triangulation, insertion_order, predicates::AbstractPredicateKernel=Adaptive()) -> Triangulation
 
 Initialises the Bowyer-Watson algorithm.
 
 # Arguments
 - `tri`: The triangulation.
 - `insertion_order`: The insertion order of the points. See [`get_insertion_order`](@ref).
-- `predicates::AbstractPredicateKernel=def_alg222()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
+- `predicates::AbstractPredicateKernel=Adaptive()`: Method to use for computing predicates. Can be one of [`Fast`](@ref), [`Exact`](@ref), and [`Adaptive`](@ref). See the documentation for a further discussion of these methods.
 
 # Output
 `tri` is updated in place to contain the initial triangle from which the Bowyer-Watson algorithm starts.
 """
-function initialise_bowyer_watson!(tri::Triangulation, insertion_order, predicates::AbstractPredicateKernel=def_alg222())
+function initialise_bowyer_watson!(tri::Triangulation, insertion_order, predicates::AbstractPredicateKernel=Adaptive())
     I = integer_type(tri)
     initial_triangle = get_initial_triangle(tri, insertion_order, predicates)
     add_triangle!(tri, initial_triangle; update_ghost_edges=true)
@@ -114,7 +114,7 @@ function get_initial_search_point(tri::Triangulation, num_points, new_point, ins
 end
 
 """
-    add_point_bowyer_watson!(tri::Triangulation, new_point, initial_search_point::I, rng::Random.AbstractRNG=Random.default_rng(), update_representative_point=true, store_event_history=Val(false), event_history=nothing, peek::P=Val(false), predicates::AbstractPredicateKernel=def_alg222()) -> Triangle 
[…]
