### 2024-07-27 b7a8a9a
See also in docstrings; Show off kernel in readme
