+++ test/predicates/index_and_ghost_handling.jl
@@ -199,7 +199,7 @@ for PT in subtypes(DT.AbstractPredicateKernel)
             @test DT.is_inside(DT.point_position_relative_to_circumcircle(PT(), tri, DT.𝒢, 2, 1, 27))
             p28 = @SVector[-5.247746, 0.905588]
             push!(pts, p28)
-            @test DT.is_outside(DT.point_position_relative_to_circumcircle(PT(), PT(), tri, 2, 1, DT.𝒢, 28))
+            @test DT.is_outside(DT.point_position_relative_to_circumcircle(PT(),tri, 2, 1, DT.𝒢, 28))
             @test DT.is_outside(DT.point_position_relative_to_circumcircle(PT(), tri, 1, DT.𝒢, 2, 28))
             @test DT.is_outside(DT.point_position_relative_to_circumcircle(PT(), tri, DT.𝒢, 2, 1, 28))
             @test DT.is_outside(DT.point_position_relative_to_circumcircle(PT(), tri, 8, 7, 11, 28))
@@ -421,18 +421,18 @@ for PT in subtypes(DT.AbstractPredicateKernel)
             u1_i = length(pts)
             i, j = index_map["c"], index_map["e"]
             k = index_map["d"]
-            cert1 = DT.point_position_on_line_segment(PT(), tri, i, j, k)
-            cert2 = DT.point_position_on_line_segment(PT(), tri, i, j, t1_i)
-            cert3 = DT.point_position_on_line_segment(PT(), tri, i, j, u1_i)
-            cert4 = DT.point_position_on_line_segment(PT(), tri, j, i, k)
-            cert5 = DT.point_position_on_line_segment(PT(), tri, j, i, t1_i)
-            cert6 = DT.point_position_on_line_segment(PT(), tri, j, i, u1_i)
-            cert7 = DT.point_position_on_line_segment(PT(), tri, j, i, i)
-            cert8 = DT.point_position_on_line_segment(PT(), tri, j, i, j)
-            cert9 = DT.point_position_on_line_segment(PT(), tri, i, j, i)
-            cert10 = DT.point_position_on_line_segment(PT(), tri, i, j, j)
-            cert11 = DT.point_position_on_line_segment(PT(), tri, i, j, pts[j])
-            cert12 = DT.point_position_on_line_segment(PT(), tri, pts[i], j, i)
+            cert1 = DT.point_position_on_line_segment(tri, i, j, k)
+            cert2 = DT.point_position_on_line_segment(tri, i, j, t1_i)
+            cert3 = DT.point_position_on_line_segment(tri, i, j, u1_i)
+            cert4 = DT.point_position_on_line_segment(tri, j, i, k)
+            cert5 = DT.point_position_on_line_segment(tri, j, i, t1_i)
+            cert6 = DT.point_position_on_line_segment(tri, j, i, u1_i)
+            cert7 = DT.point_position_on_line_segment(tri, j, i, i)
+            cert8 = DT.point_position_on_line_segment(tri, j, i, j)
+            cert9 = DT.point_position_on_line_segment(tri, i, j, i)
+            cert10 = DT.point_position_on_line_segment( tri, i, j, j)
+            cert11 = DT.point_position_on_line_segment( tri, i, j, pts[j])
+            cert12 = DT.point_position_on_line_segment( tri, pts[i], j, i)
             @test all(DT.is_on, (cert1, cert4))
             @test all(DT.is_left, (cert3, cert5))
             @test all(DT.is_right, (cert2, cert6))
@@ -444,20 +444,20 @@ for PT in subtypes(DT.AbstractPredicateKernel)
             w1_i = length(pts)
             i, j = index_map["a"], index_map["c"]
             k = index_map["b"]
[…]
