+++ test/refinement/refine.jl
@@ -1215,9 +1215,8 @@ end
             in_dt_encroached_edges_lens_45, not_in_dt_encroached_edges_lens_45 = slow_encroachment_test_diametral_lens(tri, 45.0)
             in_dt_encroached_edges_lens_30, not_in_dt_encroached_edges_lens_30 = slow_encroachment_test_diametral_lens(tri, 30.0)
             in_dt_encroached_edges_lens_20, not_in_dt_encroached_edges_lens_20 = slow_encroachment_test_diametral_lens(tri, 20.0)
-            in_dt_encroached_edges_lens_10, not_in_dt_encroached_edges_lens_10 = slow_encroachment_test_diametral_lens(tri, 15.0)
             all_bn = DT.get_all_boundary_nodes(tri)
-            for (lens_angle, not_in_dt_encroached_edges) in zip((45.0, 30.0, 20.0, 15.0), (not_in_dt_encroached_edges_lens_45, not_in_dt_encroached_edges_lens_30, not_in_dt_encroached_edges_lens_20, not_in_dt_encroached_edges_lens_10))
+            for (lens_angle, not_in_dt_encroached_edges) in zip((45.0, 30.0, 20.0), (not_in_dt_encroached_edges_lens_45, not_in_dt_encroached_edges_lens_30, not_in_dt_encroached_edges_lens_20))
                 @info "Testing encroached edge detection. lens angle: $lens_angle"
                 args = DT.RefinementArguments(tri; use_lens=true, min_angle=lens_angle)
                 for (e, (b, k)) in not_in_dt_encroached_edges
@@ -1231,7 +1230,7 @@ end
                     end
                 end
             end
-            for (lens_angle, in_dt_encroached_edges) in zip((45.0, 30.0, 20.0, 15.0), (in_dt_encroached_edges_lens_45, in_dt_encroached_edges_lens_30, in_dt_encroached_edges_lens_20, in_dt_encroached_edges_lens_10))
+            for (lens_angle, in_dt_encroached_edges) in zip((45.0, 30.0, 20.0), (in_dt_encroached_edges_lens_45, in_dt_encroached_edges_lens_30, in_dt_encroached_edges_lens_20))
                 args = DT.RefinementArguments(tri; use_lens=true, min_angle=lens_angle)
                 for (e, (b, k)) in in_dt_encroached_edges
                     if DT.initial(e) ∈ all_bn && DT.terminal(e) ∈ all_bn && !DT.contains_segment(tri, e)
@@ -1712,9 +1711,10 @@ end
 
 @testset "enqueueing and splitting all encroached segments" begin
     for PT in (DT.Exact, DT.Adaptive)
-        for iii in 1:25
+        for iii in 1:5
             for use_lens in (false, true)
                 for pass in 1:2
+                    @info "Testing enqueuing and splitting all encroached segments. Predicates: $PT; Run: $iii; lens: $use_lens; Block: $pass"
                     if pass == 1
                         p1 = (0.0, 0.0)
                         p2 = (1.0, 0.0)
@@ -2151,25 +2151,27 @@ end
     end
 
     @testset "check_for_steiner_point_on_segment" begin
-        points = [(0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.0, 1.0)]
-        tri = triangulate(points; randomise=false)
-        push!(points, (1 / 2, 1 / 2))
-        V = (1, 2, 3)
-        V′ = (1, 2, 3)
-        new_point = 5
-        flag = DT.Cert.On
-        ison_flag = DT.check_for_steiner_point_on_segment(tri, V, V′, new_point, flag)
-        @test !ison_flag
-        add_segment!(tri, 1, 3)
[…]
