+++ test/refinement/refine.jl
@@ -2392,9 +2392,9 @@ end
                                 (7.0, 7.8), (7.0, 7.4)]
                             boundary_nodes, points = convert_boundary_points_to_indices(curves; existing_points=points)
                             tri = triangulate(points; boundary_nodes, rng)
-                            max_area = max_area * get_area(tri)
-                            refine!(tri; min_angle, min_area, max_area, use_circumcenter=true, seditious_angle, use_lens, rng)
-                            @test validate_refinement(tri; min_angle, min_area, max_area, use_circumcenter=true, seditious_angle, use_lens, check_conformal=false)
+                            _max_area = max_area * get_area(tri)
+                            refine!(tri; min_angle, min_area, max_area=_max_area, use_circumcenter=true, seditious_angle, use_lens, rng)
+                            @test validate_refinement(tri; min_angle, min_area, max_area=_max_area, use_circumcenter=true, seditious_angle, use_lens, check_conformal=false)
                             if _rng_num(idx1, idx2, idx3, idx4, idx5) == _rng_num(1, 3, 1, 3, 1)
                                 fig, ax, sc = triplot(tri)
                                 @test_reference "a_constrained_triangulation_with_multiple_holes_circle.png" fig by = psnr_equality(15)
@@ -2557,7 +2557,7 @@ end
     end
 
     @testset "Refining disjoint sets" begin
-        @info "Testing the refinement of disjoint sets" 
+        @info "Testing the refinement of disjoint sets"
         θ = LinRange(0, 2π, 30) |> collect
         θ[end] = 0
         xy = Vector{Vector{Vector{NTuple{2,Float64}}}}()
@@ -2812,58 +2812,59 @@ end
 
     @testset "Test that nothing is breaking for Float32 inputs" begin
         for PT in (DT.Exact, DT.Adaptive)
-        for i in 1:25
-            p1 = (0.0f0, 0.0f0)
-            p2 = (1.0f0, 0.0f0)
-            p3 = (0.0f0, 1.0f0)
-            p4 = (1.0f0, 1.0f0)
-            pts = [p1, p2, p3, p4]
-            tri = triangulate(pts; delete_ghosts=false, predicates=PT())
-            validate_statistics(tri)
-            refine!(tri; max_area=0.001, max_points=25_000, use_circumcenter=true, predicates=PT())
-            stats = statistics(tri)
-            @inferred statistics(tri)
-            @test DT.get_smallest_angle(stats) ≥ deg2rad(30.0)
-            @test DT.get_largest_area(stats) ≤ 0.001f0
-            @test !DT.is_constrained(tri)
-            @test DT.convex_hull(tri).vertices == DT.convex_hull(tri.points).vertices
-            @test validate_triangulation(tri)
-            @test validate_refinement(tri; max_area=0.001, max_points=25_000, use_circumcenter=true)
-        end
+            for i in 1:25
+                p1 = (0.0f0, 0.0f0)
+                p2 = (1.0f0, 0.0f0)
+                p3 = (0.0f0, 1.0f0)
+                p4 = (1.0f0, 1.0f0)
+                pts = [p1, p2, p3, p4]
+                tri = triangulate(pts; delete_ghosts=false, predicates=PT())
+                validate_statistics(tri)
+                refine!(tri; max_area=0.001, max_points=25_000, use_circumcenter=true, predicates=PT())
+                stats = statistics(tri)
+                @inferred statistics(tri)
[…]
