### 2024-07-27 d0fd937
triangle area
