+++ src/data_structures/statistics/individual_triangle_statistics.jl
@@ -1,5 +1,4 @@
 # Some formulas in this file come from https://perso.uclouvain.be/jean-francois.remacle/LMECA2170/robnotes.pdf.
-# Unfortunately, I can't use the robust forms of the formulas because we don't have implementations for them (only the versions that take the sign of the orient predicate).
 
 """
     IndividualTriangleStatistics{T}
@@ -67,7 +66,7 @@ function IndividualTriangleStatistics(p, q, r, sink=(NaN, NaN))
     F = number_type(p)
     ℓmin², ℓmed², ℓmax² = squared_triangle_lengths(p, q, r)
     ℓmin, ℓmed, ℓmax = sqrt(ℓmin²), sqrt(ℓmed²), sqrt(ℓmax²)
-    A = triangle_area(ℓmin², ℓmed², ℓmax²)
+    A = triangle_area(p, q, r)
     circumcenter = triangle_circumcenter(p, q, r, A)
     circumradius = triangle_circumradius(A, ℓmin², ℓmed², ℓmax²)
     radius_edge_ratio = triangle_radius_edge_ratio(circumradius, ℓmin)
@@ -96,49 +95,32 @@ function IndividualTriangleStatistics(p, q, r, sink=(NaN, NaN))
     )
 end
 
-## We could simplify some of this even further, e.g. keeping some more things as their square temporarily.
-## Is it really worth the effort, though?
-
 """
-    triangle_area(ℓ₁², ℓ₂², ℓ₃²) -> Number
+    triangle_area(p, q, r) -> Number 
 
-Compute the area of a triangle given the squares of its edge lengths. If there are precision issues that cause the area to be negative, then the area is set to zero.
-
-See also [`squared_triangle_area`](@ref).
+Returns the signed area of a triangle `(p, q, r)`. The area is positive if `(p, q, r)`
+is positively oriented.
 """
-function triangle_area(ℓ₁²::Number, ℓ₂²::Number, ℓ₃²::Number)
-    A² = squared_triangle_area(ℓ₁², ℓ₂², ℓ₃²)
-    A = A² < 0.0 ? zero(A²) : sqrt(A²) # needed e.g. if A² is like -1.084020e-19
-    return A
+function triangle_area(p, q, r)
+    return AdaptivePredicates.orient2(getxy(p), getxy(q), getxy(r)) / 2
 end
 
-@doc raw"""
-    squared_triangle_area(ℓ₁², ℓ₂², ℓ₃²) -> Number
-
-Compute the squared area of a triangle given the squares of its edge lengths. Heron's formula is used, so that the squared area is 
-
-```math 
-A^2 = \dfrac{1}{16}\left[4\ell_1^2\ell_2^2 - \left(\ell_1^2 + \ell_2^2 - \ell_3^2\right)^2\right]..
-```
-
-See also [`squared_triangle_area_v2`](@ref).
 """
-squared_triangle_area(ℓ₁²::Number, ℓ₂²::Number, ℓ₃²::Number) = (4ℓ₁² * ℓ₂² - (ℓ₁² + ℓ₂² - ℓ₃²)^2) / 16 # Heron's formula
+    triangle_area(ℓ₁²::Number, ℓ₂²::Number, ℓ₃²::Number) -> Number 
 
-@doc raw"""
-    squared_triangle_area_v2(ℓ₁², ℓ₂², ℓ₃²) -> Number
-
-Compute the squared area of a triangle given the squares of its edge lengths, given in sorted order so that `ℓ₁² ≤ ℓ₂² ≤ ℓ₃²`. This is a more numerically stable version of [`squared_triangle_area`](@ref) using 
-the formula from [Kahan (2014)](https://people.eecs.berkeley.edu/~wkahan/Triangle.pdf):
-
-```math 
[…]
