### 2024-07-27 fc9a4fa
triangle area
