+++ test/helper_functions.jl
@@ -499,22 +499,22 @@ function validate_statistics(tri::DT.Triangulation, stats=statistics(tri))
     sort!(all_vertices)
     sort!(solid_vertices)
     sort!(ghost_vertices)
-
+    NT = DT.number_type(tri)
     ## Build up the individual statistics 
-    areas = Dict{T,Float64}()
-    lengths = Dict{T,NTuple{3,Float64}}()
-    circumcenters = Dict{T,NTuple{2,Float64}}()
-    circumradii = Dict{T,Float64}()
-    angles = Dict{T,NTuple{3,Float64}}()
-    radius_edge_ratio = Dict{T,Float64}()
-    edge_midpoints = Dict{T,NTuple{3,NTuple{2,Float64}}}()
-    aspect_ratio = Dict{T,Float64}()
-    inradius = Dict{T,Float64}()
-    perimeter = Dict{T,Float64}()
-    centroid = Dict{T,NTuple{2,Float64}}()
-    offcenters = Dict{T,NTuple{2,Float64}}()
-    sinks = Dict{T,NTuple{2,Float64}}()
-    total_A = 0.0
+    areas = Dict{T,NT}()
+    lengths = Dict{T,NTuple{3,NT}}()
+    circumcenters = Dict{T,NTuple{2,NT}}()
+    circumradii = Dict{T,NT}()
+    angles = Dict{T,NTuple{3,NT}}()
+    radius_edge_ratio = Dict{T,NT}()
+    edge_midpoints = Dict{T,NTuple{3,NTuple{2,NT}}}()
+    aspect_ratio = Dict{T,NT}()
+    inradius = Dict{T,NT}()
+    perimeter = Dict{T,NT}()
+    centroid = Dict{T,NTuple{2,NT}}()
+    offcenters = Dict{T,NTuple{2,NT}}()
+    sinks = Dict{T,NTuple{2,NT}}()
+    total_A = zero(NT)
     for T in each_solid_triangle(tri)
         u, v, w = DT.triangle_vertices(T)
         p, q, r = get_point(tri, u, v, w)
@@ -588,6 +588,26 @@ function validate_statistics(tri::DT.Triangulation, stats=statistics(tri))
     @test stats.individual_statistics == DT.get_individual_statistics(stats)
     @test stats.area ≈ DT.get_area(stats) rtol = 1e-4 atol = 1e-4
     @test stats.area ≈ total_A rtol = 1e-4 atol = 1e-4
+
+    ## Inference 
+    for _ in 1:5
+        T = rand(DT.each_solid_triangle(tri))
+        p, q, r = @inferred DT.get_point(tri, DT.triangle_vertices(T)...)
+        ℓmin², ℓmed², ℓmax² = @inferred DT.squared_triangle_lengths(p, q, r)
+        ℓmin, ℓmed, ℓmax = sqrt(ℓmin²), sqrt(ℓmed²), sqrt(ℓmax²)
+        A = @inferred DT.triangle_area(p, q, r)
+        circumcenter = @inferred DT.triangle_circumcenter(p, q, r, A)
+        circumradius = @inferred DT.triangle_circumradius(A, ℓmin², ℓmed², ℓmax²)
+        radius_edge_ratio = @inferred DT.triangle_radius_edge_ratio(circumradius, ℓmin)
+        edge_midpoints = @inferred DT.triangle_edge_midpoints(p, q, r)
+        perimeter = @inferred DT.triangle_perimeter(ℓmin, ℓmed, ℓmax)
+        inradius = @inferred DT.triangle_inradius(A, perimeter)
+        aspect_ratio = @inferred DT.triangle_aspect_ratio(inradius, circumradius)
[…]
