+++ docs/src/manual/primitives.md
@@ -10,7 +10,7 @@ For defining new methods, as discussed in [this tutorial](../tutorials/custom_pr
 
 - Points are constantly being accessed via index, so a vector-like structure is recommended, and fixed-size tuples (like `Tuple` or a `SVector` from StaticArrays.jl) are recommended for individual points. For mutability, e.g. in the case of refinement or translating points with centroidal Voronoi tessellations, you of course want a mutable structure.
 - Edges and triangles are never accessed via index, so a vector-like structure is not needed. Instead, these collections are always being used for finding specific objects and for deleting/pushing objects, so a set-like structure is highly recommended. For individual edges and triangles, fixed-size tuples of length 2 and 3, respectively, are recommended.
-- `Int` is recommended to be used for vertices, like in an edge or a triangle, and for coordinates `Float64` is highly recommended. As discussed in the [Geometrical Predicates](predicates.md) section, the use of `Float64` is important for the robustness of the algorithms in this package - using `Float32` is highly likely to lead to errors or infinite loops in the algorithms.
+- `Int` is recommended to be used for vertices, like in an edge or a triangle, and for coordinates `Float64` is highly recommended. As discussed in the [Geometrical Predicates](predicates.md) section, the use of `Float64` is important for the robustness of the algorithms in this package - using `Float32` is highly likely to lead to errors or infinite loops in the algorithms; as of v1.1.0, the use of `Adaptive()` predicate kernel makes issues from `Float32` less likely.
 
 Of course, how you implement the methods for your custom types comes down to what you need. 
 
