### 2024-07-27 5c11fcc
precision should depend on number precision; improve centroid formula
