+++ test/helper_functions.jl
@@ -591,22 +591,22 @@ function validate_statistics(tri::DT.Triangulation, stats=statistics(tri))
 
     ## Inference 
     for _ in 1:5
-        T = rand(DT.each_solid_triangle(tri))
-        p, q, r = @inferred DT.get_point(tri, DT.triangle_vertices(T)...)
-        ℓmin², ℓmed², ℓmax² = @inferred DT.squared_triangle_lengths(p, q, r)
-        ℓmin, ℓmed, ℓmax = sqrt(ℓmin²), sqrt(ℓmed²), sqrt(ℓmax²)
-        A = @inferred DT.triangle_area(p, q, r)
-        circumcenter = @inferred DT.triangle_circumcenter(p, q, r, A)
-        circumradius = @inferred DT.triangle_circumradius(A, ℓmin², ℓmed², ℓmax²)
-        radius_edge_ratio = @inferred DT.triangle_radius_edge_ratio(circumradius, ℓmin)
-        edge_midpoints = @inferred DT.triangle_edge_midpoints(p, q, r)
-        perimeter = @inferred DT.triangle_perimeter(ℓmin, ℓmed, ℓmax)
-        inradius = @inferred DT.triangle_inradius(A, perimeter)
-        aspect_ratio = @inferred DT.triangle_aspect_ratio(inradius, circumradius)
-        centroid = @inferred DT.triangle_centroid(p, q, r)
-        angles = @inferred DT.triangle_angles(p, q, r)
-        offcenter = @inferred DT.triangle_offcenter(p, q, r, circumcenter)
-        sink = @inferred DT.triangle_sink(tri, T)
+        _T = rand(DT.each_solid_triangle(tri))
+        _p, _q, _r = @inferred DT.get_point(tri, DT.triangle_vertices(_T)...)
+        _ℓmin², _ℓmed², _ℓmax² = @inferred DT.squared_triangle_lengths(_p, _q, _r)
+        _ℓmin, _ℓmed, _ℓmax = sqrt(_ℓmin²), sqrt(_ℓmed²), sqrt(_ℓmax²)
+        _A = @inferred DT.triangle_area(_p, _q, _r)
+        _circumcenter = @inferred DT.triangle_circumcenter(_p, _q, _r, _A)
+        _circumradius = @inferred DT.triangle_circumradius(_A, _ℓmin², _ℓmed², _ℓmax²)
+        _radius_edge_ratio = @inferred DT.triangle_radius_edge_ratio(_circumradius, _ℓmin)
+        _edge_midpoints = @inferred DT.triangle_edge_midpoints(_p, _q, _r)
+        _perimeter = @inferred DT.triangle_perimeter(_ℓmin, _ℓmed, _ℓmax)
+        _inradius = @inferred DT.triangle_inradius(_A, _perimeter)
+        _aspect_ratio = @inferred DT.triangle_aspect_ratio(_inradius, _circumradius)
+        _centroid = @inferred DT.triangle_centroid(_p, _q, _r)
+        _angles = @inferred DT.triangle_angles(_p, _q, _r)
+        _offcenter = @inferred DT.triangle_offcenter(_p, _q, _r, _circumcenter)
+        _sink = @inferred DT.triangle_sink(tri, _T)
     end
 
     ## Test the number statistics 
