+++ src/predicates/predicate_definitions.jl
@@ -83,17 +83,17 @@ Returns
 - `-1`: `(p, q, r)` is negatively oriented.
 
 The `kernel` argument determines how this result is computed, and should be 
-one of [`Exact`](@ref), [`Fast`](@ref), and [`Adaptive`](@ref) (the default).
+one of [`ExactKernel`](@ref), [`FastKernel`](@ref), and [`AdaptiveKernel`](@ref) (the default).
 See the documentation for more information about these choices.
 """
 @inline function orient_predicate(kernel::AbstractPredicateKernel, p, q, r)
     return orient(kernel, getxy(p), getxy(q), getxy(r))
 end
-@inline orient_predicate(p, q, r) = orient_predicate(Adaptive(), p, q, r)
+@inline orient_predicate(p, q, r) = orient_predicate(AdaptiveKernel(), p, q, r)
 
-@inline orient(::Fast, p, q, r) = sgn(AP.orient2fast(p, q, r))
-@inline orient(::Exact, p, q, r) = EP.orient(_getxy(p), _getxy(q), _getxy(r))
-@inline orient(::Adaptive, p, q, r) = AP.orient2p(p, q, r)
+@inline orient(::FastKernel, p, q, r) = sgn(AP.orient2fast(p, q, r))
+@inline orient(::ExactKernel, p, q, r) = EP.orient(_getxy(p), _getxy(q), _getxy(r))
+@inline orient(::AdaptiveKernel, p, q, r) = AP.orient2p(p, q, r)
 
 """
     orient_predicate([kernel::AbstractPredicateKernel,] p, q, r) -> Integer
@@ -124,17 +124,17 @@ Here, a positively oriented tetrahedron `(p, q, r, s)` takes the form
                                  '\\.    
 
 The `kernel` argument determines how this result is computed, and should be 
-one of [`Exact`](@ref), [`Fast`](@ref), and [`Adaptive`](@ref) (the default).
+one of [`ExactKernel`](@ref), [`FastKernel`](@ref), and [`AdaptiveKernel`](@ref) (the default).
 See the documentation for more information about these choices.
 """
 @inline function orient_predicate(kernel::AbstractPredicateKernel, p, q, r, s)
     return orient(kernel, getxyz(p), getxyz(q), getxyz(r), getxyz(s))
 end
-@inline orient_predicate(p, q, r, s) = orient_predicate(Adaptive(), p, q, r, s)
+@inline orient_predicate(p, q, r, s) = orient_predicate(AdaptiveKernel(), p, q, r, s)
 
-@inline orient(::Fast, p, q, r, s) = sgn(AP.orient3fast(p, q, r, s))
-@inline orient(::Exact, p, q, r, s) = EP.orient(_getxyz(p), _getxyz(q), _getxyz(r), _getxyz(s))
-@inline orient(::Adaptive, p, q, r, s) = AP.orient3p(p, q, r, s)
+@inline orient(::FastKernel, p, q, r, s) = sgn(AP.orient3fast(p, q, r, s))
+@inline orient(::ExactKernel, p, q, r, s) = EP.orient(_getxyz(p), _getxyz(q), _getxyz(r), _getxyz(s))
+@inline orient(::AdaptiveKernel, p, q, r, s) = AP.orient3p(p, q, r, s)
 
 """
     incircle_predicate([kernel::AbstractPredicateKernel,] a, b, c, p) -> Integer
@@ -146,17 +146,17 @@ Assuming that `(a, b, c)` is a positively oriented triangle, returns
 - `-1`: If `p` is outside the circle defined by `(a, b, c)`.
 
 The `kernel` argument determines how this result is computed, and should be 
[…]
