+++ src/public.jl
@@ -267,9 +267,9 @@
             pole_of_inaccessibility,
             is_point2,
             AbstractPredicateKernel,
-            Fast,
-            Exact,
-            Adaptive,
+            FastKernel,
+            ExactKernel,
+            AdaptiveKernel,
             orient_predicate,
             incircle_predicate,
             parallelorder_predicate,

[…]
