+++ docs/src/literate_tutorials/pole_of_inaccessibility.jl
@@ -47,7 +47,8 @@ points = [
     7.27436 2.7979
     3.0 4.0
     5.33697 1.88019]'
-outer_boundary = [ # split into segments for demonstration purposes
+outer_boundary = [ 
+    ## split into segments for demonstration purposes
     [1, 4, 3, 2],
     [2, 9, 10, 11, 8, 7, 12],
     [12, 6, 13, 5, 14, 15, 16, 17, 16],
