+++ test/predicates/general.jl
@@ -61,24 +61,28 @@ end
             @test DT.orient_predicate(DT.FastKernel(), p, q, r) === Int(sign(AdaptivePredicates.orient2fast(p, q, r))) === ExactPredicates.orient(p, q, r)
             @test DT.orient_predicate(DT.AdaptiveKernel(), p, q, r) === AdaptivePredicates.orient2p(p, q, r) === ExactPredicates.orient(p, q, r)
             @inferred DT.orient_predicate(rt(), p, q, r)
+            @test DT.orient_predicate(DT.AdaptiveKernel(), p, q, r) === DT.orient_predicate(p, q, r) 
 
             p, q, r, s = eachcol(rand(3, 4))
             @test DT.orient_predicate(DT.ExactKernel(), p, q, r, s) === ExactPredicates.orient(p, q, r, s)
             @test DT.orient_predicate(DT.FastKernel(), p, q, r, s) === Int(sign(AdaptivePredicates.orient3fast(p, q, r, s))) === ExactPredicates.orient(p, q, r, s)
             @test DT.orient_predicate(DT.AdaptiveKernel(), p, q, r, s) === AdaptivePredicates.orient3p(p, q, r, s) === ExactPredicates.orient(p, q, r, s)
             @inferred DT.orient_predicate(rt(), p, q, r, s)
+            @test DT.orient_predicate(DT.AdaptiveKernel(), p, q, r) === DT.orient_predicate(p, q, r) 
 
             a, b, c, p = eachcol(rand(2, 4))
             @test DT.incircle_predicate(DT.ExactKernel(), a, b, c, p) === ExactPredicates.incircle(a, b, c, p)
             @test DT.incircle_predicate(DT.FastKernel(), a, b, c, p) === Int(sign(AdaptivePredicates.incirclefast(a, b, c, p))) === ExactPredicates.incircle(a, b, c, p)
             @test DT.incircle_predicate(DT.AdaptiveKernel(), a, b, c, p) === AdaptivePredicates.incirclep(a, b, c, p) === ExactPredicates.incircle(a, b, c, p)
             @inferred DT.incircle_predicate(rt(), a, b, c, p)
+            @test DT.incircle_predicate(DT.AdaptiveKernel(), a, b, c, p) === DT.incircle_predicate(a, b, c, p)
 
             a, b, p, q = eachcol(rand(2, 4))
             @test DT.parallelorder_predicate(DT.ExactKernel(), a, b, p, q) === ExactPredicates.parallelorder(a, b, p, q)
             @test DT.parallelorder_predicate(DT.FastKernel(), a, b, p, q) === DT.parallelorder(DT.FastKernel(), a, b, p, q) === ExactPredicates.parallelorder(a, b, p, q)
             @test DT.parallelorder_predicate(DT.AdaptiveKernel(), a, b, p, q) === ExactPredicates.parallelorder(a, b, p, q) === ExactPredicates.parallelorder(a, b, p, q)
             @inferred DT.parallelorder_predicate(rt(), a, b, p, q)
+            @test DT.parallelorder_predicate(DT.AdaptiveKernel(), a, b, p, q) ===DT.parallelorder_predicate(a, b, p, q)
 
             p, a, b = eachcol(rand(2, 3))
             @test DT.sameside_predicate(a, b, p) === ExactPredicates.sameside(p, a, b)
@@ -89,6 +93,7 @@ end
             @test DT.meet_predicate(DT.FastKernel(), p, q, a, b) === ExactPredicates.meet(p, q, a, b)
             @test DT.meet_predicate(DT.AdaptiveKernel(), p, q, a, b) === ExactPredicates.meet(p, q, a, b)
             @inferred DT.meet_predicate(rt(), p, q, a, b)
[…]
