### 2024-07-28 50de5e3
Provide a choice of predicate kernel. Make AdaptivePredicates.jl the new default (#165)
