+++ test/refinement/curve_bounded.jl
@@ -10,7 +10,7 @@ using Test
 using StructEquality
 using DelimitedFiles
 using StableRNGs
-using Preferences
+
 const SACM = DT.SmallAngleComplexMember
 const SAC = DT.SmallAngleComplex
 
@@ -412,7 +412,7 @@ end
         @test DT.get_boundary_curves(enricher) == _boundary_curves
         hierarchy = DT.get_polygon_hierarchy(enricher)
         _hierarchy = DT.construct_polygon_hierarchy(points, _boundary_nodes, _boundary_curves)
-        DT.expand_bounds!(_hierarchy, DT.ε)
+        DT.expand_bounds!(_hierarchy, DT.ε(Float64))
         @test compare_trees(hierarchy, _hierarchy)
         for i in eachindex(_boundary_curves)
             @test DT.get_boundary_curve(enricher, i) == _boundary_curves[i]
@@ -863,20 +863,22 @@ end
 
     curve = enricher_IV.boundary_curves[1]
     p = (10randn(), 10randn())
-    @test DT.point_position_relative_to_curve(enricher_IV, 1, p) == DT.point_position_relative_to_curve(curve, p)
-    @inferred DT.point_position_relative_to_curve(enricher_IV, 1, p)
+    @test DT.point_position_relative_to_curve(rt(), enricher_IV, 1, p) == DT.point_position_relative_to_curve(rt(), curve, p)
+    @test DT.point_position_relative_to_curve(enricher_IV, 1, p) == DT.point_position_relative_to_curve(AdaptiveKernel(), curve, p)
+    @inferred DT.point_position_relative_to_curve(rt(), enricher_IV, 1, p)
     curve = enricher_XI.boundary_curves[2]
     p = (10randn(), 10randn())
-    @test DT.point_position_relative_to_curve(enricher_XI, 2, p) == DT.point_position_relative_to_curve(curve, p)
-    @inferred DT.point_position_relative_to_curve(enricher_XI, 2, p)
+    @test DT.point_position_relative_to_curve(rt(), enricher_XI, 2, p) == DT.point_position_relative_to_curve(rt(), curve, p)
+    @inferred DT.point_position_relative_to_curve(rt(), enricher_XI, 2, p)
     curve = enricher_XI.boundary_curves[8]
     p = (10randn(), 10randn())
-    @test DT.point_position_relative_to_curve(enricher_XI, 8, p) == DT.point_position_relative_to_curve(curve, p)
-    @inferred DT.point_position_relative_to_curve(enricher_XI, 8, p)
+    @test DT.point_position_relative_to_curve(rt(), enricher_XI, 8, p) == DT.point_position_relative_to_curve(rt(), curve, p)
+    @inferred DT.point_position_relative_to_curve(rt(), enricher_XI, 8, p)
     curve = enricher_X.boundary_curves[3]
     p = (10randn(), 10randn())
-    @test DT.point_position_relative_to_curve(enricher_X, 3, p) == DT.point_position_relative_to_curve(curve, p)
-    @inferred DT.point_position_relative_to_curve(enricher_X, 3, p)
+    @test DT.point_position_relative_to_curve(rt(), enricher_X, 3, p) == DT.point_position_relative_to_curve(rt(), curve, p)
+    @inferred DT.point_position_relative_to_curve(rt(), enricher_X, 3, p)
[…]
