+++ test/voronoi/voronoi.jl
@@ -23,13 +23,13 @@ using GeometryBasics
             F = (1.0, 4.0)
             G = (-3.0, 5.0)
             pts = [A, B, C, D, E, F, G]
-            tri = triangulate(pts; delete_ghosts=false, randomise=false)
+            tri = triangulate(pts; delete_ghosts = false, randomise = false)
 
-            vorn = voronoi(tri; predicates=PT())
+            vorn = voronoi(tri; predicates = PT())
             for (i, p) in DT.get_generators(vorn)
                 @test get_point(tri, i) == get_generator(vorn, i) == p
             end
-            @test validate_tessellation(vorn; predicates=PT())
+            @test validate_tessellation(vorn; predicates = PT())
             @test DT.get_triangulation(vorn) == tri
             circumcenter_to_triangle = DT.get_circumcenter_to_triangle(vorn)
             triangle_to_circumcenter = DT.get_triangle_to_circumcenter(vorn)
@@ -53,45 +53,59 @@ using GeometryBasics
             @test isempty(DT.get_boundary_polygons(vorn))
             @test DT.circular_equality(
                 get_polygon(vorn, 1),
-                DT.get_triangle_to_circumcenter.(Ref(vorn), [
-                    (7, 4, 1), (4, 6, 1), (6, 2, 1), (1, 2, -1), (7, 1, -1), (7, 4, 1)
-                ])
+                DT.get_triangle_to_circumcenter.(
+                    Ref(vorn), [
+                        (7, 4, 1), (4, 6, 1), (6, 2, 1), (1, 2, -1), (7, 1, -1), (7, 4, 1),
+                    ],
+                ),
             )
             @test DT.circular_equality(
                 get_polygon(vorn, 2),
-                DT.get_triangle_to_circumcenter.(Ref(vorn), [
-                    (2, 5, -1), (1, 2, -1), (6, 2, 1), (6, 5, 2), (2, 5, -1)
-                ])
+                DT.get_triangle_to_circumcenter.(
+                    Ref(vorn), [
+                        (2, 5, -1), (1, 2, -1), (6, 2, 1), (6, 5, 2), (2, 5, -1),
+                    ],
+                ),
             )
             @test DT.circular_equality(
                 get_polygon(vorn, 3),
-                DT.get_triangle_to_circumcenter.(Ref(vorn), [
-                    (5, 3, -1), (5, 4, 3), (4, 7, 3), (3, 7, -1), (5, 3, -1)
-                ])
+                DT.get_triangle_to_circumcenter.(
+                    Ref(vorn), [
+                        (5, 3, -1), (5, 4, 3), (4, 7, 3), (3, 7, -1), (5, 3, -1),
+                    ],
+                ),
             )
             @test DT.circular_equality(
                 get_polygon(vorn, 4),
-                DT.get_triangle_to_circumcenter.(Ref(vorn), [
-                    (5, 6, 4), (4, 6, 1), (7, 4, 1), (4, 7, 3), (5, 4, 3), (5, 6, 4)
-                ])
+                DT.get_triangle_to_circumcenter.(
+                    Ref(vorn), [
+                        (5, 6, 4), (4, 6, 1), (7, 4, 1), (4, 7, 3), (5, 4, 3), (5, 6, 4),
+                    ],
+                ),
             )
             @test DT.circular_equality(
                 get_polygon(vorn, 5),
-                DT.get_triangle_to_circumcenter.(Ref(vorn), [
-                    (6, 5, 2), (5, 6, 4), (5, 4, 3), (5, 3, -1), (2, 5, -1), (6, 5, 2)
-                ])
+                DT.get_triangle_to_circumcenter.(
+                    Ref(vorn), [
[…]
