### 2024-07-28 2b3314e
Update NEWS.md
