### 2024-07-28 8202822
Wrap matlab code in code blocks so it's easier to interpret in `test/helper_functions.jl` (#167)
