+++ paper/paper.bib
@@ -84,6 +84,16 @@ doi = {10.1137/050643568},
     commit = {a9dce0334ee62104a14930fe206764bf802c23db},
 }
 
+@misc{churavy2024adaptive,
+    title = {{AdaptivePredicates.jl: Port of Shewchuk's robust predicates into Julia}},
+    author = {Valentin Churavy and Daniel VandenHeuvel},
+    year = {2024},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    howpublished = {\url{https://github.com/JuliaGeometry/AdaptivePredicates.jl}},
+    commit = {4e30b4b3f1a59d346369aa0970eecc9d82e6b3aa},
+}
+
 @article{hermann2008delaunay,
 title = {{D}elaunay-object-dynamics: {C}ell mechanics with a {3D} kinetic and dynamic weighted {D}elaunay triangulation},
 year = {2008},
@@ -250,3 +260,12 @@ journal = {Journal of Physics A: Mathematical and Theoretical},
 doi = {10.1088/1751-8121/ac4a1d},
 }
 
+@article{shewchuk1997adaptive,
+title = {Adaptive precision floating-point arithmetic and fast robust geometric predicates},
+journal = {Discrete Computational Geometry},
+volume = {18},
+pages = {305--363},
+year = {1997},
+author = {Jonathan Shewchuk},
+doi = {10.1007/PL00009321},
+}
