### 2024-07-28 #6937 @DanielVandH
Just had to make a quick update to the summary (specifically "making use of AdaptivePredicates.jl and ExactPredicates.jl for computing predicates with robust arithmetic", AdaptivePredicates.jl is the new default rather than only ExactPredicates.jl) after I upgraded DelaunayTriangulation.jl to v1.1.0
