### 2024-08-03 ba611fa
some notes in docstrings
