+++ src/data_structures/voronoi.jl
@@ -10,7 +10,7 @@ See also [`voronoi`](@ref).
     Accessing the fields themselves using e.g. `vorn.field` is not recommended and is not intended 
     to be in the public API. You should be using the 
     accessor functions, e.g. instead of `vorn.adjacent` do `get_adjacent(vorn)`. Similarly, for the iterators,
-    e.g. `vorn.generators`, `each_generators(tri)` is recommended instead.
+    e.g. `vorn.generators`, `each_generators(vorn)` is recommended instead.
 
 # Fields 
 - `triangulation::Tr`: The underlying triangulation. The tessellation is dual to this triangulation, 
