### 2024-08-03 fb56c79
some notes in docstrings (#171)
