### 2024-08-03 e93bb68
Update NEWS.md
