### 2024-08-04 99cb5e8
make sort_triangle public
