### 2024-08-04 907368d
make sort_triangle public (#174)
