+++ versions.js
@@ -1,0 +1,13 @@
+var DOC_VERSIONS = [
+  "stable",
+  "v1.1",
+  "v1.0",
+  "v0.8",
+  "v0.7",
+  "v0.6",
+  "v0.5",
+  "v0.4",
+  "dev",
+];
+var DOCUMENTER_NEWEST = "v1.1.2";
+var DOCUMENTER_STABLE = "stable";
