+++ versions.js
@@ -9,5 +9,5 @@ var DOC_VERSIONS = [
   "v0.4",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.1.2";
+var DOCUMENTER_NEWEST = "v1.1.3";
 var DOCUMENTER_STABLE = "stable";
