### 2024-08-09 78c793a
Update curve_bounded.jl
