### 2024-08-12 68b9eb6
Update centroidal.md
