+++ docs/src/tutorials/centroidal.md
@@ -23,7 +23,7 @@ vorn = voronoi(tri, clip = true)
 ````
 
 To now compute the centroidal tessellation, use [`centroidal_smooth`](@ref). (
-If you want to straight from a triangulation to a centroidal tessellation, you
+If you want to go straight from a triangulation to a centroidal tessellation, you
 can also just use `smooth_vorn = voronoi(tri, clip = true, smooth = true)`.)
 
 ````@example centroidal
