### 2024-08-31 6a857dc
Fix curve-bounded refinement with custom structs
