+++ src/data_structures/mesh_refinement/boundary_enricher.jl
@@ -105,8 +105,8 @@ get_members(complex::SmallAngleComplex) = complex.members
 
 Returns a map from an apex vertex to a list of all curves that define a small angle complex associated with that apex vertex.
 """
-function get_small_angle_complexes(points, boundary_nodes, boundary_curves, segments=nothing; IntegerType=Int)
-    d = Dict{IntegerType,Vector{SmallAngleComplex{IntegerType}}}()
+function get_small_angle_complexes(points, boundary_nodes, boundary_curves, segments = nothing; IntegerType = Int)
+    d = Dict{IntegerType, Vector{SmallAngleComplex{IntegerType}}}()
     if has_multiple_curves(boundary_nodes)
         _get_small_angle_complexes_multiple_curves!(d, boundary_nodes, boundary_curves, IntegerType)
     elseif has_multiple_sections(boundary_nodes)
@@ -208,7 +208,7 @@ Segments are stored twice. The vertices associated with a vertex are sorted coun
 to define the coordinates.
 """
 function construct_segment_map(segments, points, ::Type{I}) where {I}
-    segment_map = Dict{I,Vector{I}}()
+    segment_map = Dict{I, Vector{I}}()
     for e in each_edge(segments)
         i, j = edge_vertices(e)
         iset = get!(Vector{I}, segment_map, i)
@@ -225,12 +225,12 @@ function construct_segment_map(segments, points, ::Type{I}) where {I}
         qx, qy = getxy(q)
         base = (qx - px, qy - py)
         sort!(
-            vertices, by=_vertex -> begin
+            vertices, by = _vertex -> begin
                 _q = get_point(points, _vertex)
                 _qx, _qy = getxy(_q)
                 next_base = (_qx - px, _qy - py)
                 return angle_between(base, next_base)
-            end, rev=false,
+            end, rev = false,
         )
     end
     return segment_map
@@ -251,12 +251,12 @@ function sort_members!(complex::SmallAngleComplex, points)
     qx, qy = getxy(q)
     base = (qx - px, qy - py)
     sort!(
-        members, by=member -> begin
+        members, by = member -> begin
             _q = get_point(points, get_next_edge(member))
             _qx, _qy = getxy(_q)
             next_base = (_qx - px, _qy - py)
             return angle_between(base, next_base)
-        end, rev=false,
+        end, rev = false,
     )
     return complex
 end
@@ -367,37 +367,37 @@ field will no longer aliased with the input `boundary_nodes`, although `points` 
 The argument `n` is used in [`polygonise`](@ref) for filling out the boundary temporarily in order to construct the [`PolygonHierarchy`](@ref). The argument `coarse_n` defines the initial coarse discretisation 
 through [`coarse_discretisation!`](@ref); the default `n=0` means that the coarse discretisation will be performed until the maximum total variation of a subcurve is less than π/2.
 """
-struct BoundaryEnricher{P,B,C,I,BM,S,E}
+struct BoundaryEnricher{P, B, C, I, BM, S, E}
     points::P
     boundary_nodes::B
     segments::S
     boundary_curves::C
     polygon_hierarchy::PolygonHierarchy{I}
-    parent_map::Dict{E,I}
-    curve_index_map::Dict{I,I}
+    parent_map::Dict{E, I}
+    curve_index_map::Dict{I, I}
     boundary_edge_map::BM
     spatial_tree::BoundaryRTree{P}
     queue::Queue{I}
-    small_angle_complexes::Dict{I,Vector{SmallAngleComplex{I}}}
[…]
