+++ test/refinement/curve_bounded.jl
@@ -90,7 +90,7 @@ segments_III = Set([(n + 2, n + 1)])
 fpoints_III_extra_segments = flatten_boundary_nodes(points_III_extra_segments, curve_III, segments_III)
 
 curve_IV = [CircularArc((1.0, 0.0), (1.0, 0.0), (0.0, 0.0))]
-points_IV = NTuple{2,Float64}[]
+points_IV = NTuple{2, Float64}[]
 fpoints_IV = flatten_boundary_nodes(points_IV, curve_IV)
 points_IV_extra = copy(points_IV)
 push!(points_IV_extra, (0.99, 0.14), (0.0, 0.99), (-0.99, 0.0), (0.99, 0.0), (0.0, -0.99), (0.0, 0.0), (0.5, -0.5))
@@ -130,7 +130,7 @@ curve_VIII = [
     [1, 2, 3, 4, 5],
     [DT.EllipticalArc((0.0, 0.0), (2.0, -2.0), (1.0, -1.0), sqrt(2), sqrt(2), 45.0)],
     [6, 7, 8, 9, 10],
-    [CatmullRomSpline([(10.0, -3.0), (20.0, 0.0), (18.0, 0.0), (10.0, 0.0)], lookup_steps=5000)],
+    [CatmullRomSpline([(10.0, -3.0), (20.0, 0.0), (18.0, 0.0), (10.0, 0.0)], lookup_steps = 5000)],
 ]
 points_VIII = [
     (10.0, 0.0), (8.0, 0.0), (4.0, 0.0), (2.0, 2.0), (0.0, 0.0), (2.0, -2.0),
@@ -142,13 +142,13 @@ push!(points_VIII_extra, (5.0, -0.01), (10.0, -1.0), (15.0, -1.95), (0.0, -1.0),
 
 curve_IX =
     [
-        [
-            [1, 2, 3, 4, 5, 6, 7, 1],
-        ],
-        [
-            [CircularArc((0.6, 0.5), (0.6, 0.5), (0.5, 0.5), positive=false)],
-        ],
-    ]
+    [
+        [1, 2, 3, 4, 5, 6, 7, 1],
+    ],
+    [
+        [CircularArc((0.6, 0.5), (0.6, 0.5), (0.5, 0.5), positive = false)],
+    ],
+]
 points_IX = [(0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.5, 1.5), (0.0, 1.0), (0.0, 0.5), (0.0, 0.2)]
 fpoints_IX = flatten_boundary_nodes(points_IX, curve_IX)
 points_IX_extra = copy(points_IX)
@@ -186,7 +186,7 @@ curve_XI = [
         [12, 11, 10, 12],
     ],
     [
-        [CircularArc((1.1, -3.0), (1.1, -3.0), (0.0, -3.0), positive=false)],
+        [CircularArc((1.1, -3.0), (1.1, -3.0), (0.0, -3.0), positive = false)],
     ],
 ]
 points_XI = [(-2.0, 0.0), (0.0, 0.0), (2.0, 0.0), (-2.0, -5.0), (2.0, -5.0), (2.0, -1 / 10), (-2.0, -1 / 10), (-1.0, -3.0), (0.0, -4.0), (0.0, -2.3), (-0.5, -3.5), (0.9, -3.0)]
@@ -206,7 +206,7 @@ points_XII = [
     (0.0, -0.1), (-0.1, 0.0),
     (0.4, 0.2), (0.2, 0.4), (0.0, 0.6),
 ]
[…]
