+++ src/algorithms/triangulation/triangulate_curve_bounded.jl
@@ -34,29 +34,29 @@ See also [`BoundaryEnricher`](@ref) and [`enrich_boundary!`](@ref).
     please see [`refine!`](@ref).
 """
 function triangulate_curve_bounded(
-    points::P;
-    segments=nothing,
-    boundary_nodes=nothing,
-    predicates::AbstractPredicateKernel=AdaptiveKernel(),
-    IntegerType::Type{I}=Int,
-    EdgeType::Type{E}=isnothing(segments) ? NTuple{2,IntegerType} : (edge_type ∘ typeof)(segments),
-    EdgesType::Type{Es}=isnothing(segments) ? Set{EdgeType} : typeof(segments),
-    polygonise_n=4096,
-    coarse_n=0,
-    check_arguments=true,
-    delete_ghosts=false,
-    delete_empty_features=true,
-    recompute_representative_points=true,
-    rng::Random.AbstractRNG=Random.default_rng(),
-    insertion_order=nothing, # use this so that it gets ignored by the kwargs
-    kwargs...,
-) where {P,I,E,Es}
-    enricher = BoundaryEnricher(points, boundary_nodes, segments; IntegerType, EdgeType, EdgesType, n=polygonise_n, coarse_n)
+        points::P;
+        segments = nothing,
+        boundary_nodes = nothing,
+        predicates::AbstractPredicateKernel = AdaptiveKernel(),
+        IntegerType::Type{I} = Int,
+        EdgeType::Type{E} = isnothing(segments) ? NTuple{2, IntegerType} : (edge_type ∘ typeof)(segments),
+        EdgesType::Type{Es} = isnothing(segments) ? Set{EdgeType} : typeof(segments),
+        polygonise_n = 4096,
+        coarse_n = 0,
+        check_arguments = true,
+        delete_ghosts = false,
+        delete_empty_features = true,
+        recompute_representative_points = true,
+        rng::Random.AbstractRNG = Random.default_rng(),
+        insertion_order = nothing, # use this so that it gets ignored by the kwargs
+        kwargs...,
+    ) where {P, I, E, Es}
+    enricher = BoundaryEnricher(points, boundary_nodes, segments; IntegerType, EdgeType, EdgesType, n = polygonise_n, coarse_n)
     return _triangulate_curve_bounded(
         points, enricher;
         predicates,
         IntegerType,
-        EdgeType, 
+        EdgeType,
         EdgesType,
         check_arguments,
         delete_ghosts,
@@ -68,19 +68,19 @@ function triangulate_curve_bounded(
     )
 end
 function _triangulate_curve_bounded(
-    points::P, enricher;
-    predicates::AbstractPredicateKernel=AdaptiveKernel(),
-    IntegerType::Type{I}=Int,
-    EdgeType::Type{E},
-    EdgesType::Type{Es},
-    check_arguments=true,
-    delete_ghosts=false,
-    delete_empty_features=true,
-    recompute_representative_points=true,
-    rng::Random.AbstractRNG=Random.default_rng(),
-    insertion_order=nothing, # use this so that it gets ignored by the kwargs
-    kwargs...,
-) where {P,I,E,Es}
+        points::P, enricher;
+        predicates::AbstractPredicateKernel = AdaptiveKernel(),
+        IntegerType::Type{I} = Int,
+        EdgeType::Type{E},
+        EdgesType::Type{Es},
+        check_arguments = true,
+        delete_ghosts = false,
+        delete_empty_features = true,
+        recompute_representative_points = true,
+        rng::Random.AbstractRNG = Random.default_rng(),
+        insertion_order = nothing, # use this so that it gets ignored by the kwargs
+        kwargs...,
+    ) where {P, I, E, Es}
     check_arguments && check_args(enricher)
[…]
