### 2024-09-01 24041c3
Fix curve-bounded refinement with custom structs (#175)
