+++ src/geometric_primitives/points.jl
@@ -531,7 +531,7 @@ julia> points
 set_point!
 set_point!(points::AbstractVector, i, x, y) = points[i] = (x, y)
 set_point!(points::AbstractVector{<:Vector}, i, x, y) = points[i] = [x, y]
-set_point!(points::AbstractMatrix, i, x, y) = points[:, i] .= (x, y)
+set_point!(points::AbstractMatrix, i, x, y) = (points[1, i] = x; points[2, i] = y)
 set_point!(points, i, p) = set_point!(points, i, getx(p), gety(p))
 
 """
