+++ test/interfaces/points.jl
@@ -7,7 +7,7 @@ import GeometryBasics: Point2f
 
 global p1 = [1.3, 2.5]
 global p2 = (1.3, 2.5)
-global p3 = SVector{2, Float32}((1.3, 2.5))
+global p3 = SVector{2,Float32}((1.3, 2.5))
 
 @testset "Individual points" begin
     @testset "Getting coordinates" begin
@@ -52,9 +52,9 @@ global pts4 = ((2.0, 3.5), (1.7, 23.3), (-1.0, 0.0))
             @test DT.get_point(pts, (2.0, 5.3), (17.0, 5.3)) == ((2.0, 5.3), (17.0, 5.3))
             @inferred DT.get_point(pts, (2.0, 5.3), (17.0, 5.3)) == ((2.0, 5.3), (17.0, 5.3))
             @test DT.get_point(pts, 1, 2, (17.0, -2.0), (57.0, 23.0)) ==
-                ((2.0, 3.5), (1.7, 23.3), (17.0, -2.0), (57.0, 23.0))
+                  ((2.0, 3.5), (1.7, 23.3), (17.0, -2.0), (57.0, 23.0))
             @inferred DT.get_point(pts, 1, 2, (17.0, -2.0), (57.0, 23.0)) ==
-                ((2.0, 3.5), (1.7, 23.3), (17.0, -2.0), (57.0, 23.0))
+                      ((2.0, 3.5), (1.7, 23.3), (17.0, -2.0), (57.0, 23.0))
         end
     end
 
@@ -166,9 +166,9 @@ global pts4 = ((2.0, 3.5), (1.7, 23.3), (-1.0, 0.0))
         DT.set_point!(points, 2, 3.4, 6.7)
         @test points == [[1.0, 5.0], [3.4, 6.7]]
 
-        points = [SVector{2, Float64}(1.0, 5.4), SVector{2, Float64}(6.5, 2.3)]
+        points = [SVector{2,Float64}(1.0, 5.4), SVector{2,Float64}(6.5, 2.3)]
         DT.set_point!(points, 2, 3.4, 6.7)
-        @test points == [SVector{2, Float64}(1.0, 5.4), SVector{2, Float64}(3.4, 6.7)]
+        @test points == [SVector{2,Float64}(1.0, 5.4), SVector{2,Float64}(3.4, 6.7)]
 
         points = [Float32[1.0, 5.0], Float32[2.3, -6.7]]
         DT.set_point!(points, 1, 2.3, 6.9)
@@ -177,6 +177,11 @@ global pts4 = ((2.0, 3.5), (1.7, 23.3), (-1.0, 0.0))
         points = [Point2f(2.0, 10.0), Point2f(3.0, 4.0)]
         DT.set_point!(points, 2, 3.0, 4.0)
         @test points == [Point2f(2.0, 10.0), Point2f(3.0, 4.0)]
+
+        points = rand(3, 50)
+        _points = copy(points)
+        DT.set_point!(points, 7, 3.0, 4.0)
+        @test points[:, 7] == [3.0, 4.0, _points[3, 7]]
     end
 end
 
@@ -184,7 +189,7 @@ end
     points1 = [(1.0, 2.0), (5.0, 9.0), (3.0, 4.0)]
     points2 = [1.0 5.0 3.0; 2.0 9.0 4.0]
     points3 = rand(2, 75)
-    points4 = [SVector{2, Float32}((1.0, 2.0)), SVector{2, Float32}((5.0, 9.0)), SVector{2, Float32}((3.0, 4.0))]
[…]
