### 2024-09-02 e88077b
Better default `set_point!` for matrices, and check for dimension of the user's points (#178)
