+++ test/triangulation/check_args.jl
@@ -3,6 +3,17 @@ const DT = DelaunayTriangulation
 
 _test_throws(e1, e2 = e1) = @static VERSION ≥ v"1.9" ? e1 : e2
 
+@testset "check_dimension" begin
+    points = rand(2, 50)
+    boundary_nodes = nothing
+    hierarchy = DT.construct_polygon_hierarchy(points)
+    @test DT.check_args(points, boundary_nodes, hierarchy)
+
+    points = rand(3, 50) 
+    boundary_nodes = nothing
+    hierarchy = DT.construct_polygon_hierarchy(points)
+    @test_logs (:warn, "The provided points are not in the plane. All but the first two coordinates of each point will be ignored.") DT.check_args(points, boundary_nodes, hierarchy)
+end
 
 @testset "A simple case" begin
     points = rand(2, 50)
