### 2024-09-03 #6937 @vissarion
Hi @mtsch  @atzberg @yossibokor @evetion @HaoZeke would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
