### 2024-09-03 #6937 @yossibokor
Hiiiii, depending on the timeframe I can review this.

> On 3 Sep 2024, at 12.52, Vissarion Fisikopoulos ***@***.***> wrote:
>
>
> Hi @mtsch <https://github.com/mtsch> @atzberg <https://github.com/atzberg> @yossibokor <https://github.com/yossibokor> @evetion <https://github.com/evetion> @HaoZeke <https://github.com/HaoZeke> would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
