### 2024-09-04 #6937 @evetion
Happy to review, but note that I'm close to the domain/language and have used the package incidentally.
