### 2024-09-04 #7174 @editorialbot
[REVIEW]: DelaunayTriangulation.jl: A Julia package for Delaunay triangulations and Voronoi tessellations in the plane
**Submitting author:** @DanielVandH (<a href="http://orcid.org/0000-0001-6462-0135">Daniel VandenHeuvel</a>)
**Repository:** https://github.com/JuliaGeometry/DelaunayTriangulation.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.4.2
