+++ src/algorithms/triangulation/main.jl
@@ -169,7 +169,8 @@ end
 function check_config(points, weights, segments, boundary_nodes, kernel)
     (number_type(points) == Float32 && kernel != AdaptiveKernel()) && @warn "Using non-Float64 coordinates may cause issues. If you run into problems, consider using Float64 coordinates." maxlog = 1
     is_constrained = !(isnothing(segments) || isempty(segments)) || !(isnothing(boundary_nodes) || !has_boundary_nodes(boundary_nodes))
-    is_weighted(weights) && is_constrained && throw(ArgumentError("You cannot compute a constrained triangulation with weighted points."))
+    # is_weighted(weights) && is_constrained && throw(ArgumentError("You cannot compute a constrained triangulation with weighted points."))
+    # The above is still not implemented, but it errors when someone uses lock_convex_hull! (like in centroidal_smooth). So, just ignore it for now I guess.
     return nothing
 end
 
