+++ src/data_structures/triangulation/methods/weights.jl
@@ -23,16 +23,6 @@ struct ZeroWeight{T} end
 ZeroWeight() = ZeroWeight{Float64}()
 get_weight(weights::ZeroWeight{T}, i) where {T} = zero(T)
 
-function Base.:(==)(w₁::ZeroWeight, w₂::ZeroWeight)
-    return true
-end
-function Base.:(==)(w₁::ZeroWeight, w₂)
-    return all(iszero, w₂)
-end
-function Base.:(==)(w₁, w₂::ZeroWeight)
-    return all(iszero, w₁)
-end
-
 """
     add_weight!(tri::Triangulation, w)
 
