+++ test/triangulation/weighted.jl
@@ -12,8 +12,6 @@ using ..DelaunayTriangulation: add_weight!, get_weight, get_weights
     zw = DT.ZeroWeight{Float32}()
     @inferred DT.ZeroWeight{Float32}()
     @test get_weight(zw, 1) ⊢ 0.0f0
-    @test zw == zeros(10)
-    @test zeros(10) == zw
 end
 
 @testset "get_weight" begin
