### 2024-09-07 ae1f7e2
Fix tests, fix doc example
