### 2024-09-07 e5b3e07
Fix clipping, add nearest_neighbour with power metric
