+++ src/algorithms/voronoi/clipped_coordinates.jl
@@ -137,8 +137,6 @@ function _get_ray(vorn::VoronoiTessellation, i, ghost_vertex, predicates::Abstra
     u, v, _ = triangle_vertices(ghost_tri) # w is the ghost vertex
     p, q = get_generator(vorn, u, v)
     qx, qy = getxy(q)
-    mx, my = midpoint(p, q)
-    m = (mx, my)
     is_first = is_first_ghost_vertex(C, ghost_vertex)
     if is_first
         prev_index = previndex_circular(C, ghost_vertex)
@@ -146,6 +144,11 @@ function _get_ray(vorn::VoronoiTessellation, i, ghost_vertex, predicates::Abstra
     else
         next_index = nextindex_circular(C, ghost_vertex)
         r = get_polygon_point(vorn, C[next_index])
+    end
+    if is_weighted(vorn)
+        m = project_onto_line(p, q, r)
+    else 
+        m = midpoint(p, q)
     end
     if r == m # It's possible for the circumcenter to lie on the edge and exactly at the midpoint (e.g. [(0.0,1.0),(-1.0,2.0),(-2.0,-1.0)]). In this case, just rotate 
         dx, dy = qx - mx, qy - my
