### 2024-09-08 06ca3df
Fix clipping and smoothing
