+++ src/utils/utils.jl
@@ -35,7 +35,7 @@ Float64
 number_type
 number_type(x) = number_type(typeof(x))
 number_type(::Type{T}) where {T <: AbstractArray} = number_type(eltype(T))
-number_type(x::Type{<:NTuple{N}}) where {N} = eltype(x)
+number_type(::Type{<:NTuple{N, T}}) where {N, T} = number_type(T)
 number_type(::Type{Tuple{}}) = Any
 number_type(::Type{T}) where {T} = T
 
@@ -133,7 +133,7 @@ is_circular(A) = isempty(A) || (A[begin] == A[end])
 Compares the two circular vectors `A` and `B` for equality up to circular rotation, 
 using `by` to compare individual elements.
 """
-function circular_equality(A, B, by = isequal)
+function circular_equality(A, B, by::F = isequal) where {F}
     @assert is_circular(A) && is_circular(B) "A and B must be circular"
     length(A) ≠ length(B) && return false
     isempty(A) && return true # isempty(B) is true as well because of the previous assertion 
