+++ src/algorithms/voronoi/clipped_coordinates.jl
@@ -150,6 +150,7 @@ function _get_ray(vorn::VoronoiTessellation, i, ghost_vertex, predicates::Abstra
     else 
         m = midpoint(p, q)
     end
+    mx, my = getxy(m)
     if r == m # It's possible for the circumcenter to lie on the edge and exactly at the midpoint (e.g. [(0.0,1.0),(-1.0,2.0),(-2.0,-1.0)]). In this case, just rotate 
         dx, dy = qx - mx, qy - my
         if is_right(point_position_relative_to_line(predicates, p, q, r))
@@ -308,8 +309,9 @@ Use the keyword arguments `predicates` to determine how predicates are computed.
 """
 function get_bounded_polygon_coordinates(vorn::VoronoiTessellation, i, bounding_box; predicates::AbstractPredicateKernel = AdaptiveKernel())
     if isnothing(bounding_box)
-        C = get_polygon(vorn, i)
         F = number_type(vorn)
+        !has_polygon(vorn, i) && return Vector{NTuple{2,F}}()
+        C = get_polygon(vorn, i)
         coords = Vector{NTuple{2, F}}(undef, length(C))
         for j in eachindex(C)
             coords[j] = get_polygon_point(vorn, C[j])
