+++ test/voronoi/power.jl
@@ -335,6 +335,7 @@ end
     pts = rand(rng, 2, n)
     weights = 10randn(rng, n)
     tri = triangulate(pts; weights, rng)
+    PT = AdaptiveKernel
     vorn = voronoi(tri, predicates=PT())
     flag1 = validate_tessellation(vorn, predicates=PT())
     vorn = voronoi(tri, clip=true, predicates=PT())
@@ -430,7 +431,7 @@ end
     circ = CircularArc((0.75, 0.5), (0.75, 0.5), (0.5, 0.5))
     t = LinRange(0, 1, 20)
     clip_points = circ.(t)[begin:end-1]
-    clip_vertices = [1:(length(circ_pts)); 1]
+    clip_vertices = [1:(length(clip_pts)); 1]
     clip_polygon = (clip_points, clip_vertices)
     weights = [6.12365, -1.50081, 0.853172, 6.57105, 0.863264, 10.1487, -5.45746, 5.78284, 4.70237, 2.69588]
     tri = triangulate(pts; weights)
