### 2024-09-08 22b2807
note about nearest neighbours, remove old warnings about weighted
