+++ src/data_structures/triangulation/methods/weights.jl
@@ -6,20 +6,26 @@ Pushes the weight `w` into `weights`. The default definition for this is `push!(
 add_weight!(weights, w) = push!(weights, w)
 
 """
-    get_weight(weights, i) -> Float64
+    get_weight(weights, i) -> Number
 
 Gets the `i`th weight from `weights`. The default definition for this is `weights[i]`,
 but this can be extended - e.g., [`ZeroWeight`](@ref) uses `get_weight(weights, i) = 0.0`.
+
+If `i` is not an integer, then the default definition is `is_point3(i) ? i[3] : zero(number_type(weights))`.
 """
-get_weight(weights, i) = Float64(weights[i])
+get_weight(weights, i::Integer) = weights[i]
+get_weight(weights, i) = is_point3(i) ? i[3] : zero(number_type(weights))
 
 """
-    ZeroWeight
+    ZeroWeight{T}()
 
-Struct used for indicating that a triangulation has zero weights. The weights are `Float64`.
+Struct used for indicating that a triangulation has zero weights. The weights have type `T`,
+which is `Float64` by default.
 """
-struct ZeroWeight end
-get_weight(::ZeroWeight, i) = zero(Float64)
+struct ZeroWeight{T} end
+ZeroWeight() = ZeroWeight{Float64}()
+get_weight(weights::ZeroWeight{T}, i) where {T} = is_point3(i) ? i[3] : zero(T)
+get_weight(weights::ZeroWeight{T}, i::Integer) where {T} = zero(T)
 
 """
     add_weight!(tri::Triangulation, w)
@@ -41,6 +47,10 @@ function get_weight(tri::Triangulation, i)
     weights = get_weights(tri)
     return get_weight(weights, i)
 end
+function get_weight(tri::Triangulation, i::Integer) # ambiguity
+    weights = get_weights(tri)
+    return get_weight(weights, i)
+end
 
 """
     is_weighted(weights) -> Bool
@@ -53,7 +63,7 @@ is_weighted(weights::ZeroWeight) = false
 is_weighted(weights) = true
 
 """
-    get_lifted_point(p, w) -> Tuple{Float64, Float64, Float64}
+    get_lifted_point(p, w) -> NTuple{3, Number}
 
 Returns the lifted companion of the point `p`, in particular `(x, y, x^2 + y^2 - w)`, where `(x, y)` is `p`.
 """
@@ -63,7 +73,7 @@ function get_lifted_point(p, w)
 end
 
 """
-    get_lifted_point(tri::Triangulation, i) -> Tuple{Float64, Float64, Float64}
+    get_lifted_point(tri::Triangulation, i) -> NTuple{3, Number}
 
 Returns the lifted companion of the `i`th vertex of `tri`, in particular `(x, y, x^2 + y^2 - w)`, where `w` is the
 `i`th weight of `tri` and `(x, y)` is the `i`th point of `tri`.
@@ -75,7 +85,7 @@ function get_lifted_point(tri::Triangulation, i)
 end
 
 """
-    get_power_distance(tri::Triangulation, i, j) -> Float64
+    get_power_distance(tri::Triangulation, i, j) -> Number
 
 Returns the power distance between vertices `i` and `j`, defined by 
 `||pᵢ - pⱼ||^2 - wᵢ - wⱼ`, where `wᵢ` and `wⱼ` are the respective weights.
@@ -88,7 +98,7 @@ function get_power_distance(tri::Triangulation, i, j)
 end
 
 """"
-    get_distance_to_witness_plane(tri::Triangulation, i, V) -> Float64 
+    get_distance_to_witness_plane(tri::Triangulation, i, V) -> Number
 
[…]
