+++ docs/make.jl
@@ -126,12 +126,15 @@ const _PAGES = [
         "Triangulating Rectangular Regions" => "tutorials/lattice.md",
         "Triangulating Convex Polygons" => "tutorials/convex.md",
         "Triangulating Curve-Bounded Domains" => "tutorials/curve_bounded.md",
+        "Weighted Triangulations" => "tutorials/weighted.md",
         "Voronoi Tessellations" => "tutorials/voronoi.md",
         "Clipped Voronoi Tessellations" => [
             "Clipping to the Convex Hull" => "tutorials/clipped.md",
             "Clipping to a Rectangle" => "tutorials/clipped_rectangle.md",
+            "Clipping to a Generic Convex Polygon" => "tutorials/clipped_polygon.md",
         ],
         "Centroidal Voronoi Tessellations" => "tutorials/centroidal.md",
+        "Power Diagrams" => "tutorials/power.md",
         "Point Location" => "tutorials/point_location.md",
         "Nearest Neighbour Queries" => "tutorials/nearest.md",
         "Convex Hulls" => "tutorials/convex_hull.md",
@@ -179,9 +182,11 @@ const _PAGES = [
         "Mesh Refinement" => "math/refinement.md",
         "Curves" => "math/curves.md",
         "Triangulating Curve-Bounded Domains" => "math/curve_bounded.md",
+        "Weighted Delaunay Triangulations" => "math/weighted.md",
         "Voronoi Tessellations" => "math/voronoi.md",
         "Clipped Voronoi Tessellations" => "math/clipped.md",
         "Centroidal Voronoi Tessellations" => "math/centroidal.md",
+        "Power Diagrams" => "math/power.md",
     ],
     "Example Applications" => [
         "Overview" => "applications/overview.md",
