### 2024-09-09 JuliaGeometry/DelaunayTriangulation.jl#181 @PieterjanRobbe
[JOSS] Review comments
Hi @DanielVandH, thank you so much for such an exciting Julia package!

I just finished my JOSS review at [https://github.com/openjournals/joss-reviews/issues/7174](https://github.com/openjournals/joss-reviews/issues/7174).

I think you did an amazing job with this package. The software is well-scoped, with a clear purpose to generate triangulations and tessellations in Julia in an efficient manner. I think it fills a clear gap described in the statement of need, and I appreciate the time and effort you took comparing your code to other existing (Julia or non-Julia) codes with similar purposes. The plotting capabilities (in `Makie.jl`) look great too!

I think the documentation and extensive tutorials sets this package apart from the competition. The topic of this package is somewhat far from my field of expertise, but I felt like I learned a lot by just browsing the tutorial section. Thank you!

I just have some (very minor) comments / suggestions / questions that, in my opinion, would make the package even better:

**Manuscript**

- I would suggest swapping paragraphs one and two, or rephrasing the opening sentence. I think the definitions of Delaunay triangulation and Voronoi tessellation are on point, but I feel like they should come later in the introduction, perhaps after introducing the package and its core features.
- `DelaunayTriangulation.jl is the most feature-rich ...` I think this is somewhat of a bold claim, especially since there's no reference confirming this. I feel like you do address this appropriately in the section on [similar packages](https://github.com/JuliaGeometry/DelaunayTriangulation.jl?tab=readme-ov-file#similar-packages) in the README, so perhaps a reference to that section is sufficient here?

**README**

- When running the example code in the `README.md` file, I get some seemingly strange results for the `Power Diagram` plot. There appears to be only one point. I'm not sure if this is intentional?
- Also, on my laptop, the lines with the plotting commands to not appear fully on screen. Perhaps these `triplot!` and `voronoiplot!` commands could go on a new line?
- Finally, I noticed there is a `CONTRIBUTING.md` file, but perhaps you could also refer to that file somewhere in the `README`.

Thanks!
