+++ spherical.jl
@@ -1,0 +1,126 @@
+using DelaunayTriangulation, Random, GeometryBasics, GLMakie
+GLMakie.activate!()
+const DT = DelaunayTriangulation
+struct UnitSphere{T} end
+UnitSphere() = UnitSphere{Float64}()
+struct SphericalPoint{T} # assumed to be on the unit sphere
+    x::T # could also parametrise using latitude and longitude
+    y::T
+    z::T
+end
+Random.eltype(::Type{UnitSphere{T}}) where {T} = SphericalPoint{T}
+function Random.rand(rng::AbstractRNG, ::Random.SamplerTrivial{UnitSphere{T}}) where {T}
+    x, y, z = ntuple(_ -> randn(rng, T), Val(3))
+    r⁻¹ = 1 / sqrt(x^2 + y^2 + z^2)
+    return SphericalPoint(x * r⁻¹, y * r⁻¹, z * r⁻¹)
+end
+GeometryBasics.Point3(p::SphericalPoint) = Point3(p.x, p.y, p.z)
+project(p::SphericalPoint) = (projectx(p), projecty(p))
+function invproject(p)
+    X, Y = getxy(p)
+    den = 1 + X^2 + Y^2
+    return SphericalPoint(2X / den, 2Y / den, (X^2 + Y^2 - 1) / den)
+end
+projectx(p::SphericalPoint) = p.x / (1 - p.z)
+projecty(p::SphericalPoint) = p.y / (1 - p.z)
+DT.getx(p::SphericalPoint) = projectx(p)
+DT.gety(p::SphericalPoint) = projecty(p)
+DT.number_type(::Type{SphericalPoint{T}}) where {T} = T
+
+function DT.triangle_circumcenter(tri::Triangulation{Vector{SphericalPoint{T}}}, V) where {T} # If we used a SphericalTriangle type instead, we'd dispatch on that
+    points = get_points(tri)
+    u, v, w = triangle_vertices(V)
+    p, q, r = points[u], points[v], points[w] # can't use get_point because it will return the projection instead
+    return spherical_circumcenter(p, q, r)
+end
+function spherical_circumcenter(p::SphericalPoint, q::SphericalPoint, r::SphericalPoint)
+    _p, _q, _r = Point3(p), Point3(q), Point3(r)
+    cross_pq = Point3(_p[2] * _q[3] - _q[2] * _p[3], _q[1] * _p[3] - _p[1] * _q[3], _p[1] * _q[2] - _q[1] * _p[2])
+    cross_qr = Point3(_q[2] * _r[3] - _r[2] * _q[3], _r[1] * _q[3] - _q[1] * _r[3], _q[1] * _r[2] - _r[1] * _q[2])
+    cross_rp = Point3(_r[2] * _p[3] - _p[2] * _r[3], _p[1] * _r[3] - _r[1] * _p[3], _r[1] * _p[2] - _p[1] * _r[2])
+    num = cross_pq + cross_qr + cross_rp
+    den = sqrt(num[1]^2 + num[2]^2 + num[3]^2)
+    return project(SphericalPoint(num[1] / den, num[2] / den, num[3] / den)) # project because DelaunayTriangulation uses NTuple{2}
+end
+
+function spherical_triangulate(points; kwargs...)
+    idx = findfirst(==(SphericalPoint(0.0, 0.0, 1.0)), DT.each_point(points))
+    skip_points = isnothing(idx) ? 0 : idx
+    tri = triangulate(points; skip_points = skip_points, kwargs...) # assuming skip_points not in kwargs
+    # We could just note that the missing part of the triangulation near the north pole 
+    # can be represented by each_ghost_triangle(tri), mapping -1 to the north pole. To use Mesh, 
[…]
