### 2024-09-10 2a4acbf
Change order of paragraphs in summary
