### 2024-09-10 7bec998
Some links, and looser messaging about the package's comparison to other software
