+++ README.md
@@ -37,8 +37,8 @@ If you would like to make an issue or contribute, please see the [CONTRIBUTING.m
 Here is a quick example of some ways the package can be used. As mentioned, see the docs for many examples.
 
 ```julia
-# using Pkg; Pkg.add(["DelaunayTriangulation", "CairoMakie"])
-using DelaunayTriangulation, CairoMakie
+# using Pkg; Pkg.add(["DelaunayTriangulation", "CairoMakie", "StableRNGs"])
+using DelaunayTriangulation, CairoMakie, StableRNGs
 
 # Unconstrained 
 points = rand(2, 50)
@@ -91,35 +91,40 @@ tri8 = triangulate(points; boundary_nodes)
 refine!(tri8; max_area=1e-2get_area(tri8)) # could also use find_polygon to help define a custom refinement function for each shape
 
 # Weighted triangulation example
-points = tuple.(rand(20), rand(20))
-weights = 3randn(20)
-tri9 = triangulate(points; weights)
+rng = StableRNG(1) # You can pass rngs for reproducibility. Alternatively, just use Random.seed!(n)
+points = tuple.(rand(rng, 20), rand(rng, 20))
+weights = 3randn(rng, 20)
+tri9 = triangulate(points; weights, rng)
 
 # Power diagram example 
-vorn10 = voronoi(tri9) # can also use clip/smooth here 
+points = tuple.(rand(rng, 100), rand(rng, 100))
+weights = rand(rng, 100)
+tri10 = triangulate(points; weights, rng)
+vorn10 = voronoi(tri10) # can also use clip/smooth here 
 
 # Clipped Voronoi example with a generic convex polygon
-points = 10randn(2, 100)
-weights = rand(100)
-circ = CircularArc((0.0, 2.0), (0.0, 2.0), (0.0, 0.0)) # clip to a circle 
+points = 10randn(rng, 2, 100)
+weights = rand(rng, 100)
+circ = CircularArc((0.0, 5.0), (0.0, 5.0), (0.0, 0.0)) # clip to a circle 
 clip_points = [circ(t) for t in LinRange(0, 1, 100)]
 clip_vertices = [1:(length(clip_points)-1); 1]
-tri11 = triangulate(points; weights)
-vorn11 = voronoi(tri11, clip=true, clip_polygon=(clip_points, clip_vertices))
+tri11 = triangulate(points; weights, rng)
+vorn11 = voronoi(tri11, clip=true, clip_polygon=(clip_points, clip_vertices), rng=rng)
 
 # Plotting 
 fig = Figure(fontsize = 42, size = (2800, 2200))
-ax = Axis(fig[1, 1], title = "Unconstrained", width = 600, height = 600);            triplot!(ax, tri1)
-ax = Axis(fig[1, 2], title = "Voronoi", width = 600, height = 600);                  voronoiplot!(ax, vorn2)
-ax = Axis(fig[1, 3], title = "Clipped Voronoi", width = 600, height = 600);          voronoiplot!(ax, vorn3)
-ax = Axis(fig[1, 4], title = "Centroidal Voronoi", width = 600, height = 600);       voronoiplot!(ax, vorn4)
-ax = Axis(fig[2, 1], title = "Constrained", width = 600, height = 600);              triplot!(ax, tri5)
-ax = Axis(fig[2, 2], title = "Disjoint Constrained", width = 600, height = 600);     triplot!(ax, tri6)
-ax = Axis(fig[2, 3], title = "Curve-Bounded", width = 600, height = 600);            triplot!(ax, tri7)
[…]
