+++ paper/paper.md
@@ -29,7 +29,7 @@ Given a set of points $\mathcal P$, edges $\mathcal E$, and piecewise linear bou
 
 Delaunay triangulations and Voronoi tessellations have applications in a myriad of fields. Delaunay triangulations have been used for point location [@mucke1999fast], solving differential equations [@golias1997delaunay; @ju2006adaptive], path planning [@yan2008path], etc. Voronoi tessellations are typically useful when there is some notion of _influence_ associated with a point, and have been applied to problems such as geospatial interpolation [@bobach2009natural], image processing [@du1999centroidal], and cell biology [@hermann2008delaunay; @wang2024calibration].
 
-Several software packages with support for computing Delaunay triangulations and Voronoi tessellations in two dimensions already exist, such as [_Triangle_](https://www.cs.cmu.edu/~quake/triangle.html) [@shewchuk1996triangle], [_MATLAB_](https://uk.mathworks.com/help/matlab/computational-geometry.html?s_tid=CRUX_lftnav) [@MATLAB], [_SciPy_](https://docs.scipy.org/doc/scipy/tutorial/spatial.html) [@SciPy], [_CGAL_](https://www.cgal.org/) [@CGAL], and [_Gmsh_](https://gmsh.info/) [@GMSH]. There are also other Julia packages supporting some of these features, although none are as developed as DelaunayTriangulation.jl; a comparison with these other packages is given in DelaunayTriangulation.jl's [README](https://github.com/JuliaGeometry/DelaunayTriangulation.jl?tab=readme-ov-file#similar-packages). DelaunayTriangulation.jl supports a larger set of features than most of these other software packages, such as triangulation of curve-bounded domains and power diagrams, and benefits from the high-performance of Julia to efficiently support many operations. Julia's multiple dispatch [@bezanson2017julia] 
+Several software packages with support for computing Delaunay triangulations and Voronoi tessellations in two dimensions already exist, such as [_Triangle_](https://www.cs.cmu.edu/~quake/triangle.html) [@shewchuk1996triangle], [_MATLAB_](https://uk.mathworks.com/help/matlab/computational-geometry.html?s_tid=CRUX_lftnav) [@MATLAB], [_SciPy_](https://docs.scipy.org/doc/scipy/tutorial/spatial.html) [@SciPy], [_CGAL_](https://www.cgal.org/) [@CGAL], and [_Gmsh_](https://gmsh.info/) [@GMSH]. There are also other Julia packages supporting some of these features, although none are as developed as DelaunayTriangulation.jl; a comparison with these other packages is given in DelaunayTriangulation.jl's [README](https://github.com/JuliaGeometry/DelaunayTriangulation.jl?tab=readme-ov-file#similar-packages). DelaunayTriangulation.jl supports a larger set of features than most of these other software packages, such as power diagrams and the triangulation of curve-bounded domains, and benefits from the high-performance of Julia to efficiently support many operations. Julia's multiple dispatch [@bezanson2017julia] 
 is leveraged to allow for complete customisation in how a user wishes to represent geometric primitives such as points and domain boundaries, a useful feature for allowing users to represent primitives in a way that suits their application without needing to sacrfice performance. The [documentation](https://juliageometry.github.io/DelaunayTriangulation.jl/stable/) lists many more features, including its ability to represent a wide range of domains, even those that are disjoint and with holes.
 
[…]
