### 2024-09-10 #7174 @DanielVandH
I've made some changes in response to @PieterjanRobbe's comments over at https://github.com/JuliaGeometry/DelaunayTriangulation.jl/issues/181, and made some other minor changes. To summarise:
1. Since I just finished implementing weighted triangulations and power diagrams this weekend, I've mentioned them in the summary as a feature that the package supports.
2. I changed the language of "The most feature-rich" in comparing my package to other software. I now just claim to support features that most of the other packages don't support, which I think should be appropriate. I also now mention that there are other Julia packages that support triangulations, and link to where I discuss that in my package's README since it would be a bit cluttered to also do a full comparison in the paper.
3. The summary paragraphs have been reordered. I now mention the package first and then, in the second paragraph, define triangulations and tessellations.
4. Since random power diagrams often look bad, I changed the README example to use a `rng` object for reproducibility. I also made the example use a bit less text width.

Thanks again to @PieterjanRobbe for the comments, I appreciate it.
