### 2024-09-11 JuliaGeometry/DelaunayTriangulation.jl#181 @PieterjanRobbe
Hi @DanielVandH, thanks for the quick follow-up! All of these changes look good to me. I must have missed the reference to the `CONTRIBUTING.md` file, thanks for pointing this out! I'll close the issue.
