+++ test/triangulation/weighted.jl
@@ -33,7 +33,7 @@ end
     w3 = DT.ZeroWeight()
     @test DT.get_weight(w1, 1) == w1[1]
     @test DT.get_weight(w1, ()) == 0.0
-    @test DT.get_weight(w1, (1.0, 3.0, 4.0)) == 4.0 
+    @test DT.get_weight(w1, (1.0, 3.0, 4.0)) == 4.0
     @test DT.get_weight(w2, 1) == w2[1]
     @test DT.get_weight(w2, ()) === 0.0f0
     @test DT.get_weight(w2, (1.0, 3.0, 4.0)) == 4.0f0
@@ -54,11 +54,11 @@ end
 @testset "is_weighted" begin
     tri = Triangulation(rand(2, 10))
     @test !DT.is_weighted(tri)
-    tri = Triangulation(rand(2, 10); weights = rand(10))
+    tri = Triangulation(rand(2, 10); weights=rand(10))
     @test DT.is_weighted(tri)
-    tri = Triangulation(rand(2, 10); weights = DT.ZeroWeight())
+    tri = Triangulation(rand(2, 10); weights=DT.ZeroWeight())
     @test !DT.is_weighted(tri)
-    tri = Triangulation(rand(2, 10); weights = zeros(Float32, 10))
+    tri = Triangulation(rand(2, 10); weights=zeros(Float32, 10))
     @test DT.is_weighted(tri)
 end
 
@@ -135,15 +135,15 @@ end
     _weights[58] = weights[2]
     _weights[498] = weights[3]
     _weights[5] = weights[4]
-    tri = Triangulation(_points; weights = _weights)
+    tri = Triangulation(_points; weights=_weights)
     d = DT.get_distance_to_witness_plane(tri, 5, (137, 58, 498))
     @test d ≈ -2.129523129725314
     @test DT.get_distance_to_witness_plane(tri, 5, (137, 58, 498)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (58, 137, 498)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (498, 58, 137)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (137, 498, 58)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (58, 498, 137)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (498, 137, 58))
+          DT.get_distance_to_witness_plane(tri, 5, (58, 137, 498)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (498, 58, 137)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (137, 498, 58)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (58, 498, 137)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (498, 137, 58))
     _points[5] = (-3.9, 2.01)
     _weights[5] = -15.0
     d = DT.get_distance_to_witness_plane(tri, 5, (137, 58, 498))
@@ -228,8 +228,8 @@ end
         for i in 3:10
             for j in 3:10
                 tri = triangulate_rectangle(0, 10, 0, 10, i, j)
-                tri = triangulate(get_points(tri); weights = zeros(i * j))
-                @test validate_triangulation(tri) 
+                tri = triangulate(get_points(tri); weights=zeros(i * j))
+                @test validate_triangulation(tri)
                 validate_statistics(tri)
             end
         end
@@ -249,7 +249,7 @@ end
         for i in 3:10
             for j in 3:10
                 tri = triangulate_rectangle(0, 10, 0, 10, i, j)
[…]
