+++ src/algorithms/triangulation/check_args.jl
@@ -108,9 +108,9 @@ function Base.showerror(io::IO, err::InconsistentOrientationError)
 end
 
 function check_dimension(points)
-    valid = is_planar(points) 
-    if !valid 
-        @warn "The provided points are not in the plane. All but the first two coordinates of each point will be ignored." maxlog=1
+    valid = is_planar(points)
+    if !valid
+        @warn "The provided points are not in the plane. All but the first two coordinates of each point will be ignored." maxlog = 1
     end
     return valid
 end
