+++ src/algorithms/triangulation/unconstrained_triangulation.jl
@@ -153,10 +153,6 @@ function add_point_bowyer_watson!(tri::Triangulation, new_point, initial_search_
     q = get_point(tri, _new_point)
     V = find_triangle(tri, q; predicates, m = nothing, point_indices = nothing, try_points = nothing, k = initial_search_point, rng)
     if is_weighted(tri)
-        #=
-        This part here is why the Bowyer-Watson algorithm is not implemented for weighted triangulations yet. I don't understand why it sometimes 
-        fails here for submerged points. Need to look into it some more.
-        =#
         cert = point_position_relative_to_circumcircle(predicates, tri, V, _new_point) # redirects to point_position_relative_to_witness_plane
         is_outside(cert) && return V # If the point is submerged, then we don't add it
     end
@@ -300,13 +296,13 @@ Determines whether to enter the cavity in `tri` through the edge `(i, j)` when i
 - `true` if the cavity should be entered, and `false` otherwise. See also [`dig_cavity!`](@ref) and [`point_position_relative_to_circumcircle`](@ref).
 """
 function enter_cavity(tri::Triangulation, r, i, j, ℓ, predicates::AbstractPredicateKernel = AdaptiveKernel())
-    contains_segment(tri, i, j)  && return false 
+    contains_segment(tri, i, j)  && return false
     if is_ghost_vertex(ℓ)
         cert = point_position_relative_to_circumcircle(tri, j, i, ℓ, r)
-    else 
+    else
         cert = point_position_relative_to_circumcircle(tri, r, i, j, ℓ)
     end
-    return is_inside(cert)
+    return is_weighted(tri) ? !is_outside(cert) : is_inside(cert)
 end
 
 """
