### 2024-09-13 6f24f16
Fix degenerate weighted triangulation (#184)
