+++ test/data_structures/triangulation.jl
@@ -6,80 +6,81 @@ using Setfield
 using Random
 using StatsBase
 using StableRNGs
+using AdaptivePredicates
 using ..DelaunayTriangulation: add_weight!, get_weight, get_weights
-
-@struct_equal DT.TriangulationCache
-
 
 using ..DelaunayTriangulation: Triangulation
 
 global pts = rand(2, 500)
-global tri = Triangulation(pts; IntegerType = Int32)
+global tri = Triangulation(pts; IntegerType=Int32)
 
 @testset "Initialising a triangulation" begin
     @test tri ⊢ Triangulation(
         pts, # points
-        Set{NTuple{3, Int32}}(), # triangles
+        Set{NTuple{3,Int32}}(), # triangles
         Int32[], # boundary_nodes
-        Set{NTuple{2, Int32}}(), # interior_segments
-        Set{NTuple{2, Int32}}(), # all_segments
+        Set{NTuple{2,Int32}}(), # interior_segments
+        Set{NTuple{2,Int32}}(), # all_segments
         DT.ZeroWeight(), # weights
-        DT.Adjacent{Int32, NTuple{2, Int32}}(), # adjacent
-        DT.Adjacent2Vertex{Int32, Set{NTuple{2, Int32}}}(), # adjacent2vertex
+        DT.Adjacent{Int32,NTuple{2,Int32}}(), # adjacent
+        DT.Adjacent2Vertex{Int32,Set{NTuple{2,Int32}}}(), # adjacent2vertex
         DT.Graph{Int32}(), # graph
         (), # boundary_curves
-        Dict{NTuple{2, Int32}, Tuple{Vector{Int32}, Int32}}(), # boundary_edge_map
-        Dict{Int32, Vector{Int32}}(DT.𝒢 => Int32[]), # ghost_vertex_map
-        Dict{Int32, UnitRange{Int32}}(-1 => -1:-1), # ghost_vertex_ranges
+        Dict{NTuple{2,Int32},Tuple{Vector{Int32},Int32}}(), # boundary_edge_map
+        Dict{Int32,Vector{Int32}}(DT.𝒢 => Int32[]), # ghost_vertex_map
+        Dict{Int32,UnitRange{Int32}}(-1 => -1:-1), # ghost_vertex_ranges
         DT.ConvexHull(pts, Int32[]), # convex_hull
-        Dict{Int32, DT.RepresentativeCoordinates{Int32, Float64}}(), # representative_point_coordinate
-        DT.construct_polygon_hierarchy(pts; IntegerType = Int32),
+        Dict{Int32,DT.RepresentativeCoordinates{Int32,Float64}}(), # representative_point_coordinate
+        DT.construct_polygon_hierarchy(pts; IntegerType=Int32),
         nothing,
         DT.TriangulationCache(
             DT.Triangulation(
                 pts,
-                Set{NTuple{3, Int32}}(), # triangles
+                Set{NTuple{3,Int32}}(), # triangles
                 Int32[], # boundary_nodes
-                Set{NTuple{2, Int32}}(), # interior_segments
-                Set{NTuple{2, Int32}}(), # all_segments
+                Set{NTuple{2,Int32}}(), # interior_segments
+                Set{NTuple{2,Int32}}(), # all_segments
                 DT.ZeroWeight(), # weights
-                DT.Adjacent{Int32, NTuple{2, Int32}}(), # adjacent
-                DT.Adjacent2Vertex{Int32, Set{NTuple{2, Int32}}}(), # adjacent2vertex
+                DT.Adjacent{Int32,NTuple{2,Int32}}(), # adjacent
+                DT.Adjacent2Vertex{Int32,Set{NTuple{2,Int32}}}(), # adjacent2vertex
                 DT.Graph{Int32}(), # graph
                 (), # boundary_curves
[…]
