### 2024-09-15 7890824
Upgrade to AdaptivePredicates 1.2 (#185)
