+++ test/triangulation/weighted.jl
@@ -54,11 +54,11 @@ end
 @testset "is_weighted" begin
     tri = Triangulation(rand(2, 10))
     @test !DT.is_weighted(tri)
-    tri = Triangulation(rand(2, 10); weights = rand(10))
+    tri = Triangulation(rand(2, 10); weights=rand(10))
     @test DT.is_weighted(tri)
-    tri = Triangulation(rand(2, 10); weights = DT.ZeroWeight())
+    tri = Triangulation(rand(2, 10); weights=DT.ZeroWeight())
     @test !DT.is_weighted(tri)
-    tri = Triangulation(rand(2, 10); weights = zeros(Float32, 10))
+    tri = Triangulation(rand(2, 10); weights=zeros(Float32, 10))
     @test DT.is_weighted(tri)
 end
 
@@ -135,18 +135,24 @@ end
     _weights[58] = weights[2]
     _weights[498] = weights[3]
     _weights[5] = weights[4]
-    tri = Triangulation(_points; weights = _weights)
-    d = DT.get_distance_to_witness_plane(tri, 5, (137, 58, 498))
+    tri = Triangulation(_points; weights=_weights)
+    d = DT.get_distance_to_witness_plane(tri, 5, (137, 58, 498); cache=nothing)
     @test d ≈ -2.129523129725314
-    @test DT.get_distance_to_witness_plane(tri, 5, (137, 58, 498)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (58, 137, 498)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (498, 58, 137)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (137, 498, 58)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (58, 498, 137)) ≈
-        DT.get_distance_to_witness_plane(tri, 5, (498, 137, 58))
+    @test DT.get_distance_to_witness_plane(tri, 5, (137, 58, 498); cache=DT.get_incircle_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (58, 137, 498); cache=DT.get_incircle_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (498, 58, 137); cache=DT.get_incircle_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (137, 498, 58); cache=DT.get_incircle_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (58, 498, 137); cache=DT.get_incircle_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(tri, 5, (498, 137, 58); cache=DT.get_incircle_cache(tri))
+    @test DT.get_distance_to_witness_plane(AdaptiveKernel(), tri, 5, (137, 58, 498), cache=DT.get_incircle_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(FastKernel(), tri, 5, (58, 137, 498), cache=DT.get_orient3_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(AdaptiveKernel(), tri, 5, (498, 58, 137), cache=DT.get_incircle_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(AdaptiveKernel(), tri, 5, (137, 498, 58), cache=DT.get_incircle_cache(tri)) ≈
+          DT.get_distance_to_witness_plane(ExactKernel(), tri, 5, (58, 498, 137), cache=DT.get_incircle_cache(tri)) ≈
[…]
