### 2024-09-22 0e4df03
Spherical triangulation
