+++ test/spherical.jl
@@ -1,0 +1,207 @@
+using DelaunayTriangulation
+using DelaunayTriangulation.SphericalDelaunay
+using Test
+using Makie, CairoMakie, GLMakie
+CairoMakie.activate!()
+using Random
+using LinearAlgebra
+
+const DT = DelaunayTriangulation
+const SD = SphericalDelaunay
+
+Makie.to_vertices(p::Vector{<:SphericalPoint}) = Makie.Point3.(SD.__getxyz.(p))
+Makie.convert_arguments(p::PointBased, points::Vector{<:SphericalPoint}) = Makie.convert_arguments(p, Makie.to_vertices(points))
+Makie.convert_arguments(p::PointBased, L::SphericalLine) = Makie.convert_arguments(p, SD._to_coords(L))
+Makie.convert_arguments(p::PointBased, T::SphericalTriangle) = Makie.convert_arguments(p, SD._to_coords(T))
+Makie.convert_arguments(p::PointBased, P::SphericalPolygon) = Makie.convert_arguments(p, SD._to_coords(P))
+function Makie.convert_arguments(p::Type{<:Lines}, tri::SphericalTriangulation)
+    points = SD._to_lines(tri)
+    return Makie.convert_arguments(p, points)
+end
+function Makie.convert_arguments(p::Type{<:Lines}, vorn::SphericalTessellation)
+    points = SD._to_lines(vorn)
+    return Makie.convert_arguments(p, points)
+end
+# These mesh recipes render SO poorly...
+function Makie.convert_arguments(p::Type{<:Mesh}, T::SphericalTriangle)
+    points = SD._to_coords(T, false)
+    faces = SD._to_mesh(points)
+    return Makie.convert_arguments(p, points, faces)
+end
+function Makie.convert_arguments(p::Type{<:Mesh}, tri::SphericalTriangulation)
+    points, vertices = SD._to_triangles(tri)
+    return Makie.convert_arguments(p, points, vertices)
+end
+function Makie.convert_arguments(p::Type{<:Poly}, tri::SphericalTriangulation)
+    res = Makie.convert_arguments(Mesh, tri)
+    return Makie.convert_arguments(p, res...)
+end
+
+function uniform_rand_sphere(N)
+    # http://web.archive.org/web/20120421191837/http://www.cgafaq.info/wiki/Evenly_distributed_points_on_sphere
+    coords = Vector{SphericalPoint{NTuple{3,Float64}}}(undef, N)
+    s = 3.6 / sqrt(N)
+    dz = 2 / N
+    z = 1 - dz / 2
+    long = 0.0
+    for k in 1:N
+        r = sqrt(1 - z^2)
+        latitude = asind(z)
+        longitude = rad2deg(long)
+        coords[k] = SphericalPoint(latitude, longitude % 360)
+        long += s / r
+        z -= dz
+    end
+    return coords
+end
+
+@testset "Basic" begin
+    @test SD.UnitSphere() == SD.UnitSphere{Float64}()
+    p = (1.0, 0.0, 0.0)
+    @test SphericalPoint(p) == SphericalPoint{typeof(p)}(p)
+    @test SphericalPoint(p) == SphericalPoint(p)
+    @test SphericalPoint(0.3, 0.2, sqrt(1 - 0.3^2 - 0.2^2)) == SphericalPoint((0.3, 0.2, sqrt(1 - 0.3^2 - 0.2^2)))
+    q = (0.0, 1.0, 0.0)
+    @test SphericalPoint(p) != SphericalPoint(q)
+    p = rand(UnitSphere())
+    @test p isa SphericalPoint
[…]
