+++ src/algorithms/triangulation/basic_operations/delete_holes.jl
@@ -171,8 +171,7 @@ function find_all_triangles_to_delete(tri::Triangulation, points_to_process)
             u, v = edge_vertices(e)
             V = construct_triangle(T, u, v, node)
             if !contains_triangle(V, triangles_to_delete)[2]
-                p, q, r = get_point(tri, u, v, node)
-                c = triangle_centroid(p, q, r)
+                c = triangle_centroid(tri, V)
                 δ = dist(tri, c)
                 if δ < zero(δ)
                     add_triangle!(triangles_to_delete, V)
