+++ quarto/Project.toml
@@ -1,0 +1,6 @@
+[deps]
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+DelaunayTriangulation = "927a84f5-c5f4-47a5-9785-b46e178433df"
+FiniteVolumeMethod = "d4f04ab7-4f65-4d72-8a28-7087bc7f46f4"
+NaturalNeighbours = "f16ad982-4edb-46b1-8125-78e5a8b5a9e6"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
