### 2024-09-23 b26276e
address point location failures partially
