+++ src/predicates/predicates.jl
@@ -798,7 +798,7 @@ the two does not intersect any segments.
 - `cert`: A [`Certificate`](@ref). This will be `Visible` if `i` is visible from `q`, and `Invisible` otherwise.
 """
 function test_visibility(kernel::AbstractPredicateKernel, tri::Triangulation, q, i)
-    V, invisible_flag = find_triangle(tri, q; use_barriers = Val(true), k = i, concavity_protection = true, predicates = kernel)
+    V, invisible_flag = find_triangle(tri, q; use_barriers = Val(true), k = i, concavity_protection = true, predicates = kernel, check_sphere = false)
     if invisible_flag
         return Certificate.Invisible
     else
