+++ quarto/presentation.qmd
@@ -13,8 +13,10 @@ author:
 date: "2024-10-09" 
 engine: "julia"
 execute:
-    daemon: true
-format: revealjs
+    daemon: 600
+format: 
+  revealjs:
+      slide-number: true
 include-in-header:
   - text: |
       <style>
@@ -24,36 +26,2047 @@ include-in-header:
       </style>
 css: "style.css"
 ---
+ 
+## Overview
 
-<!--
-The title image was generated using 
+- Delaunay triangulations
+- Constrained Delaunay triangulations
+- Mesh refinement
+- Voronoi tessellations
+- Interpolation and solving PDEs
 
-using DelaunayTriangulation, CairoMakie
-outer = CircularArc((1.5, 0.0), (1.5, 0.0), (0.0, 0.0), positive=false)
-inner = CircularArc((0.8, 0.0), (0.8, 0.0), (0.0, 0.0), positive=true)
-square = [(-2.0, -2.0), (2.0, -2.0), (2.0, 2.0), (-2.0, 2.0)]
-boundary = [[[outer]], [[inner]], [[1, 2, 3, 4, 1]]]
-tri = triangulate(square; boundary_nodes=boundary)
-refine!(tri; max_area = 1e-2)
-fig, ax, sc = triplot(tri, figure = (; size = (1600, 800)), axis = (; width = 800, height = 800))
-xlims!(ax, -2, 2.5)
-ylims!(ax, -2, 2)
+## Delaunay Triangulations 
+
+```{julia}
+#| echo: false
+#| output: false
+using DelaunayTriangulation
+EllipticalArc((0.0, 0.0), (1.0, 0.0), (0.0, 0.0), 0.5, 0.5, 0.0)
+```
+
+- A triangulation of a set of points $\mathcal P$ in the plane such that no triangle's circumcircle contains any other point in the set. 
+```{julia}
+#| echo: false
+#| output: false
+using DelaunayTriangulation, CairoMakie, StableRNGs
+rng = StableRNG(123)
+points = [(0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.0, 1.0),
+    (0.25, 0.25), (0.5, 0.7), (0.2, 0.5), (0.7, 0.7),
+    (0.5, 0.9), (0.1, 0.1), (0.8, 0.2)]
+tri = triangulate(points; rng)
+T = (6, 5, 11)
+u, v, w = triangle_vertices(T)
+p, q, r = get_point(tri, u, v, w)
+cent = DelaunayTriangulation.triangle_circumcenter(p, q, r)
+rad = DelaunayTriangulation.triangle_circumradius(p, q, r)
+circles = [Circle(Point2(cent), rad)]
+
+fig, ax, sc = triplot(tri, figure=(; size=(400, 400)))
+translate!(sc, 0, 0, 1)
+poly!(ax, [p, q, r], color=(:blue, 0.3))
+poly!(ax, circles, strokecolor=:red, strokewidth=2.5, color=(:white, 0.0), linestyle=:dashdot)
+xlims!(ax, -0.01, 1.01)
+ylims!(ax, -0.01, 1.01)
+hidedecorations!(ax)
+hidespines!(ax)
+save("delaunay.png", fig)
+```
+
+![&nbsp;](delaunay.png)
+
+## Delaunay Triangulations {visibility="uncounted"}
+
+- A triangulation of a set of points $\mathcal P$ in the plane such that no triangle's circumcircle contains any other point in the set. 
+- Maximizes the minimum angle over all other triangulations of $\mathcal P$.
+
+## Delaunay Triangulations {visibility="uncounted"}
+
[…]
