+++ quarto/Project.toml
@@ -1,6 +1,10 @@
 [deps]
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 DelaunayTriangulation = "927a84f5-c5f4-47a5-9785-b46e178433df"
 FiniteVolumeMethod = "d4f04ab7-4f65-4d72-8a28-7087bc7f46f4"
+GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
 NaturalNeighbours = "f16ad982-4edb-46b1-8125-78e5a8b5a9e6"
+OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
+SteadyStateDiffEq = "9672c7b4-1e72-59bd-8a11-6ac3964bc41f"
