### 2024-09-25 JuliaGeometry/DelaunayTriangulation.jl#189 @mtsch
[JOSS] Review comments
Hi,

First of all, sorry for taking such a long time to review your JOSS submission! I must say this is the perfect example of a high-quality package. I'm especially impressed by the docs as I've found them easy to read and I feel like they cover everything.

I think overall the paper is very good, but I have two minor comments:

- Since the docs contains such a nice mathematical details section, I'd suggest you link to them in the paper at the end of the summary section. This way, an interested reader will be immediately know where look to learn more details about the methods.
- In the example, I suggest adding a few more words about mean exit times, in particular where they come up or even a short definition of them. I'm not familiar with the field, so I got a bit confused there until you started discussing the equations in more detail later.

Cheers!
