+++ quarto/presentation_files/libs/revealjs/plugin/quarto-support/support.js
@@ -1,338 +1,0 @@
-// catch all plugin for various quarto features
-window.QuartoSupport = function () {
-  function isPrintView() {
-    return /print-pdf/gi.test(window.location.search);
-  }
-
-  // helper for theme toggling
-  function toggleBackgroundTheme(el, onDarkBackground, onLightBackground) {
-    if (onDarkBackground) {
-      el.classList.add('has-dark-background')
-    } else {
-      el.classList.remove('has-dark-background')
-    }
-    if (onLightBackground) {
-      el.classList.add('has-light-background')
-    } else {
-      el.classList.remove('has-light-background')
-    }
-  }
-
-  // implement controlsAudo
-  function controlsAuto(deck) {
-    const config = deck.getConfig();
-    if (config.controlsAuto === true) {
-      const iframe = window.location !== window.parent.location;
-      const localhost =
-        window.location.hostname === "localhost" ||
-        window.location.hostname === "127.0.0.1";
-      deck.configure({
-        controls:
-          (iframe && !localhost) ||
-          (deck.hasVerticalSlides() && config.navigationMode !== "linear"),
-      });
-    }
-  }
-
-  // helper to provide event handlers for all links in a container
-  function handleLinkClickEvents(deck, container) {
-    Array.from(container.querySelectorAll("a")).forEach((el) => {
-      const url = el.getAttribute("href");
-      if (/^(http|www)/gi.test(url)) {
-        el.addEventListener(
-          "click",
-          (ev) => {
-            const fullscreen = !!window.document.fullscreen;
-            const dataPreviewLink = el.getAttribute("data-preview-link");
-
-            // if there is a local specifcation then use that
-            if (dataPreviewLink) {
-              if (
-                dataPreviewLink === "true" ||
-                (dataPreviewLink === "auto" && fullscreen)
-              ) {
-                ev.preventDefault();
-                deck.showPreview(url);
-                return false;
-              }
-            } else {
-              const previewLinks = !!deck.getConfig().previewLinks;
-              const previewLinksAuto =
-                deck.getConfig().previewLinksAuto === true;
-              if (previewLinks == true || (previewLinksAuto && fullscreen)) {
-                ev.preventDefault();
-                deck.showPreview(url);
-                return false;
-              }
-            }
-
-            // if the deck is in an iframe we want to open it externally
-            // (don't do this when in vscode though as it has its own
-            // handler for opening links externally that will be play)
-            const iframe = window.location !== window.parent.location;
-            if (
-              iframe &&
-              !window.location.search.includes("quartoPreviewReqId=")
-            ) {
-              ev.preventDefault();
-              ev.stopImmediatePropagation();
-              window.open(url, "_blank");
-              return false;
-            }
-
-            // if the user has set data-preview-link to "auto" we need to handle the event
-            // (because reveal will interpret "auto" as true)
-            if (dataPreviewLink === "auto") {
-              ev.preventDefault();
-              ev.stopImmediatePropagation();
-              const target =
-                el.getAttribute("target") ||
-                (ev.ctrlKey || ev.metaKey ? "_blank" : "");
-              if (target) {
-                window.open(url, target);
-              } else {
-                window.location.href = url;
-              }
-              return false;
-            }
-          },
-          false
-        );
-      }
-    });
-  }
-
[…]
