+++ paper/paper.bib
@@ -35,6 +35,13 @@
   school={Technische Universit{\"a}t Kaiserslautern},
 }
 
+@phdthesis{li2019first,
+  title={First passage times of diffusion processes and their applications to finance},
+  author={Luting Li},
+  year={2019},
+  school={London School of Economics and Political Science},
+}
+
 @phdthesis{gosselin2009delaunay,
 title = {Delaunay refinement mesh generation of curve-bounded domains},
 author = {Serge Gosselin},
