### 2024-09-25 JuliaGeometry/DelaunayTriangulation.jl#189 @DanielVandH
Hi @mtsch, thanks for your review and kind words! No worries at all for the delay.

> Since the docs contains such a nice mathematical details section, I'd suggest you link to them in the paper at the end of the summary section. This way, an interested reader will be immediately know where look to learn more details about the methods.

This is a good suggestion. I have now added "A more detailed description of these mathematical details can
be found in the package’s [documentation](https://juliageometry.github.io/DelaunayTriangulation.jl/stable/math/overview/)." at the end of the summary.

> In the example, I suggest adding a few more words about mean exit times, in particular where they come up or even a short definition of them. I'm not familiar with the field, so I got a bit confused there until you started discussing the equations in more detail later.

Thanks for pointing this out. You're right that it's a bit confusing currently. To address this, I have mentioned its applications when introducing the concept, and I introduce the notation $T(x, y)$ before the list of equations describing the mathematical model. Hopefully that makes it clearer!

I have regenerated the PDF over at https://github.com/openjournals/joss-reviews/issues/7174 so you can see these changes.
