### 2024-09-25 #7174 @DanielVandH
I've made some changes in response to @mtsch's comments over at https://github.com/JuliaGeometry/DelaunayTriangulation.jl/issues/189. The pdf has been regenerated, see above. To summarise:

1. I've added a sentence at the end of the summary pointing the reader to the mathematical details section of DelaunayTriangulation.jl's document, allowing them to easily know where to find more information about these methods.
2. To help make the mean exit time example clearer, I have mentioned some applications where it arises, mentioning diffusive transport, economics, and a specific example in finance. I also introduce the notation $T(x, y)$ prior to showing the equations describing the model, to avoid the need to have to read ahead and then look back up to the equations.

Thanks again @mtsch, I appreciate your feedback.
