+++ paper/paper.bib
@@ -39,7 +39,7 @@
   title={First passage times of diffusion processes and their applications to finance},
   author={Luting Li},
   year={2019},
-  school={London School of Economics and Political Science},
+  school={{London School of Economics and Political Science}},
 }
 
 @phdthesis{gosselin2009delaunay,
