### 2024-09-25 JuliaGeometry/DelaunayTriangulation.jl#189 @mtsch
Looks great! I think there's a typo in line 66. It reads:

> one model describing the mean exit time of a particle existing Ω

I believe you meant

> one model describing the mean exit time of a particle exiting Ω

Might even read better as

> one model describing the mean time of a particle exiting Ω

but that's up to you.

After this change I'll be happy to approve your submission.
