### 2024-09-26 8697a95
Update paper.md
