+++ paper/paper.md
@@ -38,7 +38,7 @@ DelaunayTriangulation.jl has already seen use in several areas. DelaunayTriangul
 
 We give one example of how the package can be used, focusing on Delaunay triangulations rather than Voronoi tessellations. Many more examples are given in the [documentation](https://juliageometry.github.io/DelaunayTriangulation.jl/stable/), including [tutorials](https://juliageometry.github.io/DelaunayTriangulation.jl/stable/tutorials/overview/) and [fully detailed applications](https://juliageometry.github.io/DelaunayTriangulation.jl/stable/applications/overview/) such as cell simulations. To fully demonstrate the utility of the package, we consider a realistic application. We omit code used for plotting with Makie.jl [@danisch2021makie] in the example below for space reasons. The complete code is available [here](https://github.com/JuliaGeometry/DelaunayTriangulation.jl/blob/paper/paper/paper.jl).
 
-We consider a domain motivated by mean exit time, relating to the time taken for a particle to reach a certain target, with applications to areas such as diffusive transport [@carr2022mean] and economics [@li2019first]. For example, mean exit time can be used to describe the expected time for a stock to reach a certain threshold [@li2019first; @redner2001guide]. Denoting the mean exit time of a particle at a point $(x, y)$ by $T(x, y)$, one model describing the mean exit time of a particle existing $\Omega$ with diffusivity $D$ starting at $(x, y)$ is given by [@redner2001guide; @carr2022mean]
+We consider a domain motivated by mean exit time, relating to the time taken for a particle to reach a certain target, with applications to areas such as diffusive transport [@carr2022mean] and economics [@li2019first]. For example, mean exit time can be used to describe the expected time for a stock to reach a certain threshold [@li2019first; @redner2001guide]. Denoting the mean exit time of a particle at a point $(x, y)$ by $T(x, y)$, one model describing the mean exit time of a particle exiting $\Omega$ with diffusivity $D$ starting at $(x, y)$ is given by [@redner2001guide; @carr2022mean]
 $$
 \begin{array}{rcll}
 D\nabla^2 T(x, y) & = & -1 & (x, y) \in \Omega, \\
