### 2024-09-26 JuliaGeometry/DelaunayTriangulation.jl#189 @DanielVandH
Thanks for spotting that! I have now fixed the typo.
