### 2024-09-26 918f87a
Define constrained/unconstrained separatey; Fix bib capitalisation; Mention programming languages
