+++ paper/paper.bib
@@ -9,7 +9,7 @@
 
 @book{aurenhammer2013voronoi,
     author = {Franz Aurenhammer and Rolf Klein and Der-Tsai Lee},
-    title = {Voronoi Diagrams and Delaunay Triangulations},
+    title = {Voronoi Diagrams and {D}elaunay Triangulations},
     publisher = {World Scientific},
     year = {2013},
     url = {https://doi.org/10.1142/8685},
