+++ README.md
@@ -162,578 +162,568 @@ In addition to these Julia packages, software packages in other programming lang
 
 Compared to all these other libraries, and only in the context of planar triangulations, DelaunayTriangulation.jl is one of the most developed in terms of the features provided, except possibly with the exception of CGAL and Gmsh who provide many features although neither are in the public domain (CGAL being GPL v3+ and Gmsh being GPL v2+), unlike DelaunayTriangulation.jl which is MIT licensed. A tabular comparison of all these packages is given below (focusing only on two dimensional meshing). If there are any errors in this comparison, please let me know. Also please note that the features listed are not intended to be exhaustive, but rather to give a general idea of the capabilities of each package, and certainly not all software packages are listed here.
 
-<style type="text/css">
-.tg  {border-collapse:collapse;border-spacing:0;}
-.tg td{border-color:black;border-style:solid;border-width:1px;font-family:Arial, sans-serif;font-size:14px;
-  overflow:hidden;padding:10px 5px;word-break:normal;}
-.tg th{border-color:black;border-style:solid;border-width:1px;font-family:Arial, sans-serif;font-size:14px;
-  font-weight:normal;overflow:hidden;padding:10px 5px;word-break:normal;}
-.tg .tg-c3ow{border-color:inherit;text-align:center;vertical-align:top}
-.tg .tg-0pky{border-color:inherit;text-align:left;vertical-align:top}
-.tg .tg-0lax{text-align:left;vertical-align:top}
-</style>
-<table class="tg"><thead>
+<table><thead>
   <tr>
-    <th class="tg-0pky"></th>
-    <th class="tg-0pky"></th>
-    <th class="tg-0pky"></th>
-    <th class="tg-c3ow" colspan="10">Delaunay Triangulations</th>
-    <th class="tg-c3ow" colspan="5">Voronoi</th>
-    <th class="tg-0pky"></th>
-    <th class="tg-0pky"></th>
-    <th class="tg-0pky"></th>
+    <th></th>
+    <th></th>
+    <th></th>
+    <th colspan="10">Delaunay Triangulations</th>
+    <th colspan="5">Voronoi</th>
+    <th></th>
+    <th></th>
+    <th></th>
   </tr></thead>
 <tbody>
   <tr>
-    <td class="tg-0pky">Package</td>
-    <td class="tg-0pky">Language</td>
-    <td class="tg-0pky">Open Source License</td>
-    <td class="tg-0pky">Unconstrained</td>
-    <td class="tg-0pky">Constrained</td>
-    <td class="tg-0pky">Weighted</td>
-    <td class="tg-0pky">Refinement<br></td>
-    <td class="tg-0pky">Curve-Bounded</td>
[…]
