+++ README.md
@@ -162,568 +162,11 @@ In addition to these Julia packages, software packages in other programming lang
 
 Compared to all these other libraries, and only in the context of planar triangulations, DelaunayTriangulation.jl is one of the most developed in terms of the features provided, except possibly with the exception of CGAL and Gmsh who provide many features although neither are in the public domain (CGAL being GPL v3+ and Gmsh being GPL v2+), unlike DelaunayTriangulation.jl which is MIT licensed. A tabular comparison of all these packages is given below (focusing only on two dimensional meshing). If there are any errors in this comparison, please let me know. Also please note that the features listed are not intended to be exhaustive, but rather to give a general idea of the capabilities of each package, and certainly not all software packages are listed here.
 
-<table><thead>
-  <tr>
-    <th></th>
-    <th></th>
-    <th></th>
-    <th colspan="10">Delaunay Triangulations</th>
-    <th colspan="5">Voronoi</th>
-    <th></th>
-    <th></th>
-    <th></th>
-  </tr></thead>
-<tbody>
-  <tr>
-    <td>Package</td>
-    <td>Language</td>
-    <td>Open Source License</td>
-    <td>Unconstrained</td>
-    <td>Constrained</td>
-    <td>Weighted</td>
-    <td>Refinement<br></td>
-    <td>Curve-Bounded</td>
-    <td>Dynamic Updates</td>
-    <td>Convex Polygons</td>
-    <td>Periodic</td>
-    <td>Spherical</td>
-    <td>Furthest-site</td>
-    <td>Standard</td>
-    <td>Clipped (Rect. or Hull)</td>
-    <td>Generic Clipped</td>
-    <td>Centroidal</td>
-    <td>Power</td>
-    <td>Point Location</td>
-    <td>Robust Predicates</td>
-    <td>Customisable Interface</td>
-  </tr>
-  <tr>
-    <td>DelaunayTriangulation.jl</td>
-    <td>Julia</td>
-    <td>✅ (MIT)</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>❌</td>
-    <td>❌</td>
-    <td>❌</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-    <td>✅</td>
-  </tr>
-  <tr>
-    <td>VoronoiDelaunay.jl</td>
-    <td>Julia</td>
-    <td>✅ (MIT)</td>
-    <td>✅</td>
-    <td>❌</td>
-    <td>❌</td>
[…]
