+++ README.md
@@ -163,12 +163,11 @@ In addition to these Julia packages, software packages in other programming lang
 Compared to all these other libraries, and only in the context of planar triangulations, DelaunayTriangulation.jl is one of the most developed in terms of the features provided, except possibly with the exception of CGAL and Gmsh who provide many features although neither are in the public domain (CGAL being GPL v3+ and Gmsh being GPL v2+), unlike DelaunayTriangulation.jl which is MIT licensed. A tabular comparison of all these packages is given below (focusing only on two dimensional meshing). If there are any errors in this comparison, please let me know. Also please note that the features listed are not intended to be exhaustive, but rather to give a general idea of the capabilities of each package, and certainly not all software packages are listed here.
 
 <div style="font-size: 10px;">
-<table><thead><tr><th></th><th>Package</th><th>DelaunayTriangulation.jl</th><th>VoronoiDelaunay.jl</th><th>VoronoiCells.jl</th><th>Delaunay.jl</th><th>MiniQhull.jl</th><th>DirectQhull.jl</th><th>Delaunator.jl</th><th>TriangleMesh.jl</th><th>Triangulate.jl</th><th>Triangle.jl</th><th>GMT.jl</th><th>Quickhull.jl</th><th>Triangle</th><th>CGAL</th><th>Gmsh</th><th>SciPy</th><th>MATLAB</th><th>Qhull</th><th>Delaunator</th><th>Spade</th><th>Acute</th><th>Tinfour</th><th>JTS</th></tr></thead><tbody><tr><td></td><td>Language</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>Julia</td><td>C</td><td>C++</td><td>Own/C++</td><td>Python</td><td>MATLAB</td><td>C</td><td>JavaScript</td><td>Rust</td><td>C</td><td>Java</td><td>Java</td></tr>
[…]
