### 2024-09-26 444085e
Software comparison (#191)
