+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 DelaunayTriangulation.jl is a feature-rich Julia [@bezanson2017julia] package for computing Delaunay triangulations and Voronoi tessellations. The package, amongst many other features, supports unconstrained and constrained triangulations, mesh refinement, clipped and centroidal Voronoi tessellations, power diagrams, and dynamic updates. Thanks to the speed and genericity of Julia, the package is both performant and robust---making use of [AdaptivePredicates.jl](https://github.com/JuliaGeometry/AdaptivePredicates.jl) [@churavy2024adaptive; @shewchuk1997adaptive] and [ExactPredicates.jl](https://github.com/lairez/ExactPredicates.jl) [@lairez2024exact] for computing predicates with robust arithmetic---while still allowing for generic representations of geometric primitives.
 
-Given a set of points $\mathcal P$, a _Delaunay triangulation_ is a subdivision of the convex hull of $\mathcal P$ into triangles, with the vertices of the triangles coming from $\mathcal P$, constructed such that no triangle's circumcircle contains any point from $\mathcal P$ in its interior [@cheng2013delaunay; @aurenhammer2013voronoi]. A _constrained Delaunay triangulation_ extends this definition to additionally allow for edges $\mathcal E$ and piecewise linear boundaries $\mathcal B$ to be included, ensuring that each segment from $\mathcal E$ and $\mathcal B$ is an edge of some triangle and the boundaries of the domain come from $\mathcal B$ [@cheng2013delaunay]. For constrained Delaunay triangulations, the triangles must still obey the empty circumcircle property above, except for allowing a point $p$ to be in a triangle $T$'s circumcircle if all line segments from $T$'s interior to $p$ intersects a segment from $\mathcal E$ or $\mathcal B$ [@cheng2013delaunay]. The boundaries $\mathcal B$ may also be given as parametric curves, in which case they are discretised until they accurately approximate the curved boundary [@gosselin2009delaunay]. A related geometric structure is the _Voronoi tessellation_ that partitions the plane into convex polygons for each $p \in \mathcal P$ such that, for a given polygon, each point in that polygon is closer to the associated polygon's point than to any other $q \in \mathcal P$ [@cheng2013delaunay; @aurenhammer2013voronoi]. Weighted triangulations and power diagrams are generalisations of these structures that allow for the inclusion of weights associated with the points [@cheng2013delaunay]. A more detailed description of these mathematical details can be found in the package's [documentation](https://juliageometry.github.io/DelaunayTriangulation.jl/stable/math/overview/).
[…]
