### 2024-09-26 #7174 @mtsch
@vissarion I'm finished with my review and I think this paper will make a great addition to JOSS! The package, its documentation, and paper are all top notch in my opinion.
