+++ quarto/presentation.qmd
@@ -18,6 +18,7 @@ format:
   revealjs:
       slide-number: true
       self-contained: true
+      background-color: "white"
 include-in-header:  
   - text: |
       <style>
@@ -1397,114 +1398,45 @@ end
 - Typically converges for $\theta_{\min} < 33.9^\circ$.
 - Extra care needed for small angles.
 
-## Ruppert's Algorithm
-
-```{julia}
-#| echo: false 
-#| output: false
-Random.seed!(0)
-outer = [(0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0)]
-inner = [(0.25, 0.25), (0.25, 0.75), (0.75, 0.75), (0.75, 0.25), (0.25, 0.25)]
-boundary = [[outer], [inner]]
-boundary_nodes, points = convert_boundary_points_to_indices(boundary)
-append!(points, [(0.2, 0.1), (0.2, 0.7), (0.5, 0.8), (0.9, 0.45)])
-segments = Set([(1, 5), (7, 3)])
-tri = triangulate(points; boundary_nodes, segments)
-args = DelaunayTriangulation.RefinementArguments(tri; max_area=0.01)
-
-triobs = Observable(tri)
-encroached_segments = lift(triobs) do tri
-    segments = NTuple{2,Float64}[]
-    for e in each_segment(tri)
-        if DelaunayTriangulation.is_encroached(tri, args, e)
-            u, v = edge_vertices(e)
-            p, q = get_point(tri, u, v)
-            push!(segments, p, q, (NaN, NaN))
-        end
-    end
-    return segments
-end
-bad_triangles = lift(triobs) do tri
-    triangles = typeof(Polygon([Point2f(0.0, 0.0), Point2f(0.0, 0.0), Point2f(0.0, 0.0)]))[]
-    for T in each_solid_triangle(tri)
-        _, flag = DelaunayTriangulation.assess_triangle_quality(tri, args, T)
-        if flag
-            u, v, w = triangle_vertices(T)
-            p, q, r = get_point(tri, u, v, w)
-            push!(triangles, Polygon([Point2f(p), Point2f(q), Point2f(r)]))
-        end
-    end
-    return triangles
-end
-
-fig, ax, sc = triplot(triobs; axis=(; width=600, height=600), strokewidth=3)
-poly!(ax, bad_triangles, color=(:blue, 0.3))
-linesegments!(ax, encroached_segments, color=:magenta, linewidth=6)
-hidedecorations!(ax)
-hidespines!(ax)
-resize_to_layout!(fig)
-record(fig, "rupperts-algorithm.mp4"; framerate=10) do io
-    [recordframe!(io) for _ in 1:(5*4)]
-    DelaunayTriangulation.enqueue_all_encroached_segments!(args, tri)
-    while DelaunayTriangulation.keep_splitting(tri, args)
-        e, _ = DelaunayTriangulation.popfirst_segment!(args.queue)
-        !DelaunayTriangulation.unoriented_edge_exists(tri, e) && continue
-        DelaunayTriangulation.split_subsegment!(tri, args, e)
-        notify(triobs)
-        recordframe!(io)
-    end
-    DelaunayTriangulation.enqueue_all_bad_triangles!(args, tri)
-    while DelaunayTriangulation.keep_iterating(tri, args)
-        T, ρ = DelaunayTriangulation.popfirst_triangle!(args.queue)
-        u, v, w = triangle_vertices(T)
[…]
