### 2024-09-27 #7174 @vissarion
@DanielVandH thanks for addressing my comments.

In particular, the table of related software and the comparisons looks great! However I think it would be good to double-check the table data once more especially the licenses part (e.g. I think that CGAL has multiple open source licenses for several parts and the option for commercial, but in your case only the triangulations package should be considered). Although, I am not an expert in licensing so this is just an advice and can be done independently from the submission.
