### 2024-09-27 #7174 @DanielVandH
Thanks for pointing out that issue with CGAL's licenses. Indeed there is also a triangulation package there with a GNU GPL license. I mention now that it has GPL/GNU GPL/Commerical licenses.

I've tagged a new version (v1.4.2) of the package. I already had a Zenodo setup at https://zenodo.org/records/13847646 which has been updated accordingly; the DOI is https://doi.org/10.5281/zenodo.13847646. Is that all correct?
