### 2024-09-27 06613b8
added support for calling Triangulation(points, triangles)
