### 2024-09-27 #7174 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
