### 2024-09-27 #7174 @danielskatz
👋 @DanielVandH - As the track editor, I'm now reviewing this for final publication.  It looks very good, though I have one question. There's no acknowledgment section, so I want to check to make sure this is intentional and that you don't want to mention any funding or other acknowledgements. Is this correct?
