### 2024-09-27 #7174 @danielskatz
Congratulations to @DanielVandH (Daniel VandenHeuvel) on your publication!!

JOSS depends on volunteers and couldn't be successful without you
